% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{add_box_above}
\alias{add_box_above}
\title{Add a box above a plot}
\usage{
add_box_above(
  nboxes = 0,
  labels = NULL,
  hfr = 0.07,
  hsp = 0.01,
  hfac = 1,
  base = 0,
  col = "black",
  bg = grey(0.35),
  cex = 1,
  offset = 0,
  logscale = ""
)
}
\description{
Add a box above a plot
}
\seealso{
Other plotting-related functions provided by utilbox: 
\code{\link{Colors}()},
\code{\link{add_tight_axes}()},
\code{\link{all_devs}},
\code{\link{as_color}()},
\code{\link{hist2}()},
\code{\link{palette_ub}()},
\code{\link{show_Colors}()}
}
\concept{plotting-related functions provided by utilbox}
