% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\name{add_col}
\alias{add_col}
\title{Safe column addition}
\usage{
add_col(.data, ...)
}
\description{
Adds a column(s) into the data frame \code{x} if the column(s) 
is/are not present.
}
\examples{
add_col(tibble(x=1:2), y=2)

}
\seealso{
Other matrix/data-frame functions provided by utilbox: 
\code{\link{identical_cols}()},
\code{\link{join_recurse}()},
\code{\link{rep_df}()},
\code{\link{rev_rows}()},
\code{\link{split_rows}()}
}
\concept{matrix/data-frame functions provided by utilbox}
