% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.R
\name{anti_join_sym}
\alias{anti_join_sym}
\title{Symmetric anti-join}
\usage{
anti_join_sym(x, y, ..., join_fun = "anti_join")
}
\description{
\code{anti_join_sym()} performs two anti-joins on the two data frames (or
tibbles) supplied to it as \code{x} and \code{y}. It returns a list of length
2 with \code{anti_join(x, y, ...)} and \code{anti_join(y, x, ...)} as it elements.
By default, it depends on the existence of a function \code{anti_join()}, but
an alternative function name can be supplied via \code{join_fun}.
}
\examples{
x = data.frame(a=1:3, b=letters[1:3], stringsAsFactors=FALSE)
y = data.frame(a=0:2, b=c('z',letters[1:2]), stringsAsFactors=FALSE)
anti_join_sym(x, y)

}
\seealso{
Other matrix/data-frame functions provided by utilbox: 
\code{\link{add_col}()},
\code{\link{identical_cols}()},
\code{\link{join_recurse}()},
\code{\link{rep_df}()},
\code{\link{rev_rows}()},
\code{\link{split_rows}()}
}
\concept{matrix/data-frame functions provided by utilbox}
