% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numchecks.R
\name{anyNonNA}
\alias{anyNonNA}
\title{Non-NA value check}
\usage{
anyNonNA(x, rowwise = TRUE)
}
\description{
Check whether a row of \code{x} contains at least one non-NA value. Works for vectors
and 2-arrays.
}
\examples{
x = rbind(c(1,NA),c(NA,NA)); anyNonNA(x)

}
\seealso{
Other check-performing functions provided by utilbox: 
\code{\link{drop_positive_negative}},
\code{\link{is_all_same}()},
\code{\link{is_decreasing}()},
\code{\link{is_even}()},
\code{\link{is_increasing}()},
\code{\link{is_integer}()},
\code{\link{is_na}()},
\code{\link{is_number}()},
\code{\link{is_odd}()},
\code{\link{is_permutation}()},
\code{\link{is_round}()},
\code{\link{is_scientific}()},
\code{\link{is_unique}()},
\code{\link{last_nonzero}()}
}
\concept{check-performing functions provided by utilbox}
