% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{append_recurse}
\alias{append_recurse}
\title{Recursive append}
\usage{
append_recurse(lists, ...)
}
\description{
Appends multiple objects together by recursively calling [base::append].
}
\examples{
append_recurse(as.list(1:3), as.list(4:7), as.list(10:15))

}
