% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{append_system_PATH}
\alias{append_system_PATH}
\title{Add to the system search paths}
\usage{
append_system_PATH(path)
}
\description{
Adds the supplied path among the system search paths.
}
\seealso{
Other system-related functions provided by utilbox: 
\code{\link{.q}()},
\code{\link{R_gui_adjust}()},
\code{\link{R_session_info}()},
\code{\link{R_system_call_args}()},
\code{\link{code_startup}},
\code{\link{environment_variables}},
\code{\link{get_system_PATH}()},
\code{\link{is_win}()},
\code{\link{timest}()}
}
\concept{system-related functions provided by utilbox}
