% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{apply_pckg}
\alias{apply_pckg}
\title{Apply function to an object in a package}
\usage{
apply_pckg(objs, pckg, f, ..., workhorse = sapply)
}
\description{
Applies function \code{f} to the object(s) whose names are in 
\code{objs} and that are found in the package \code{pckg}. Useful
for applying a function (e.g. \code{base::class}) to non-exported 
objects inside a package.
}
\examples{
apply_pckg('rev_cols', 'utilbox', class)    # returns 'function'

}
\seealso{
Other package-related functions provided by utilbox: 
\code{\link{R_add_lib_startup}()},
\code{\link{R_default_Rprofile_file}()},
\code{\link{R_user_home}()},
\code{\link{list_package}},
\code{\link{llibrary}()},
\code{\link{set_pkglib}()}
}
\concept{package-related functions provided by utilbox}
