% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_character.R
\name{as.character.closure}
\alias{as.character.closure}
\title{Method for conversion to class \code{character} for functions (closures)}
\usage{
\method{as.character}{closure}(fun)
}
\description{
This method extends the S3 class \code{as.character} to allow the
conversion of function source code to class \code{character}.
}
\examples{
as.character(base::mean)

}
