% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{as_color}
\alias{as_color}
\title{Generate a vector of colors}
\usage{
as_color(x, palette = palette_ub)
}
\description{
Generate a vector of colors based on the values in the supplied
vector \code{x}.
}
\examples{
x = c('a','b','a','c'); as_color(x)

}
\seealso{
Other plotting-related functions provided by utilbox: 
\code{\link{Colors}()},
\code{\link{add_box_above}()},
\code{\link{add_tight_axes}()},
\code{\link{all_devs}},
\code{\link{hist2}()},
\code{\link{palette_ub}()},
\code{\link{show_Colors}()}
}
\concept{plotting-related functions provided by utilbox}
