% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_factor.R
\name{as_factor}
\alias{as_factor}
\alias{unname_factor}
\title{Convert to factor and back}
\usage{
as_factor(x, ordered = FALSE)

unname_factor(x)
}
\description{
\code{as_factor()} converts a vector to class \code{factor} with levels
determined by order of appearance and \strong{without} sorting of
levels (unlike \code{base::as.factor}).

\code{unname_factor()} removes the type factor produced by
\code{as_factor()}.
}
\examples{
as_factor(c('Germany','Italy','Czechia'))

unfactor(as_factor(c('Germany','Italy','Czechia')))

}
\seealso{
Other sequence-related functions provided by utilbox: 
\code{\link{every_nth}()},
\code{\link{filter_by_position}},
\code{\link{filter_by}},
\code{\link{groups_of_unique}()},
\code{\link{h1}()},
\code{\link{insert}()},
\code{\link{is_between}()},
\code{\link{is_run}},
\code{\link{last_element}()},
\code{\link{rep2}()},
\code{\link{replace2}()},
\code{\link{replace3}()},
\code{\link{seq2}()},
\code{\link{seq_around}()},
\code{\link{sizechecks}},
\code{\link{split_into_groups}()},
\code{\link{what_bag}()}
}
\concept{sequence-related functions provided by utilbox}
