% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_number.R
\name{as_numeric}
\alias{as_numeric}
\title{Type conversion}
\usage{
as_numeric(x)
}
\description{
Convert to type \code{numeric} without any warnings about non-numeric 
elements. Non-numeric elements are turned into \code{NA}. Basically,
a less verbose version of \code{base::as.numeric}.
}
\examples{
as_numeric(c("1","2"))
as_numeric(c("1","a"))

}
\seealso{
Other numeric functions provided by utilbox: 
\code{\link{Invert1}()},
\code{\link{Invert}()},
\code{\link{all_permutations}()},
\code{\link{downweight}()},
\code{\link{force_as_}},
\code{\link{frac_part}()},
\code{\link{frac}()},
\code{\link{int_part}()},
\code{\link{lower_tri}()},
\code{\link{make_numeric}()},
\code{\link{max0}()},
\code{\link{min_dif}()},
\code{\link{mintersect}()},
\code{\link{round2}()},
\code{\link{round_nearest_power}()},
\code{\link{round_nearest}()},
\code{\link{rsignif}()},
\code{\link{signif2}()},
\code{\link{to_base_k}()},
\code{\link{unspace}()},
\code{\link{unzero}()},
\code{\link{upper_tri}()},
\code{\link{value_locale}},
\code{\link{which_max}()}
}
\concept{numeric functions provided by utilbox}
