% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{ask_to_confirm}
\alias{ask_to_confirm}
\alias{ask_to_confirm_with_timeout}
\title{Ask for confirmation}
\usage{
ask_to_confirm(msg = "Press ENTER to continue or ESC to quit ...", ...)

ask_to_confirm_with_timeout(msg = "", timeout = 10, no = NULL)
}
\description{
\code{ask_to_confirm()} and \code{ask_to_confirm_with_timeout()} ask the user for
confirmation with the given message. For the latter, a lack of input within
\code{timeout} seconds is considered non-confirmation, and so are the answers
listed in \code{no}.
}
