% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{assign2}
\alias{assign2}
\title{Assign into multiple variables}
\usage{
assign2(where, what, envir = .GlobalEnv, what_as_list = FALSE)
}
\description{
Assign a value or multiple values into multiple variables with a
single call. \code{assign2} takes the name(s) of variables in \code{what} and
assigns the value in \code{what} to all of them when \code{what_as_list=FALSE}.
Otherwise, it takes a list in \code{what} of the same length as \code{where}
and assigns into variables named in \code{where} the corresponding
elements in \code{what}. The assignment takes place in environment \code{envir}.
}
\examples{
assign2(c('x','y'), 1:2)
assign2(c('x','y'), list(1:2, 'a'), what_as_list=TRUE)

}
