% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{auto_install_packages}
\alias{auto_install_packages}
\alias{.eai}
\alias{.dai}
\alias{enable_auto_install}
\alias{disable_auto_install}
\title{Enable/disable auto-installation of packages}
\usage{
.eai(call_type = c("library", "require"))

.dai(call_type = c("library", "require"))

enable_auto_install(call_type = c("library", "require"))

disable_auto_install(call_type = c("library", "require"))
}
\description{
\code{enable_auto_install()} enables auto-installation of packages by placing
functions \code{library()} and/or \code{require()} into the global environment
(i.e., \code{.GlobalEnv}), which then supperseed the calls to \code{base::library()}
and/or \code{base::require()}.
}
\details{
\code{disable_auto_install()} deletes any functions by those names in the global
environment, thereby disabling the auto-installation.

\code{.eai()} and \code{.dai()} are aliases for \code{enable_auto_install()} and
\code{disable_auto_install()}, respectively.
}
\seealso{
Other package-related functions provided by utilbox: 
\code{\link{apply_pckg}()},
\code{\link{lib_dependencies}()},
\code{\link{list_installed_packages}()},
\code{\link{llibrary}()},
\code{\link{set_pkglib}()},
\code{\link{which_package}()}
}
\concept{package-related functions provided by utilbox}
