% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap}
\usage{
bootstrap(x, T, B = 100, ..., portion = 1e+08)
}
\description{
Returns a vector of B bootstrap values of real-valued statistic T, where T 
should be an R-function which returns a scalar. Arguments of T can be supplied
via the ellipsis '...'. If length of x and B are small enough, do it all at once
Otherwise, do it for each iteration separately.
}
