% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revalue.R
\name{bound_between}
\alias{bound_between}
\title{Bound values}
\usage{
bound_between(x, lower, upper, na.rm = FALSE)
}
\description{
Lower- and upper-bounds the elements in `x` with `lower` and `upper`
}
\examples{
bound_between(-10:10, 0, 5)
bound_between(-4:5, rep(c(-1,-2),5), rep(

}
