% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{browse}
\alias{browse}
\title{Insert a browser into a pipe}
\usage{
browse(x, ...)
}
\description{
A function that is intended to be inserted as a step
in a pipe sequence during debugging to explore the state
of the input object. It only calls the function \code{base::browser()}
and returns the input object unmodified. Exit the browser
with the call \code{cont} (or \code{c} for short). See \code{?browser} for more
information.
}
\examples{
mtcars \%>\% browse() \%>\% print()   # use `.`, `cyl`, tidyverse-style
1:10 \%>\% browse() \%>\% print()     # use `x` to get the pipe input

}
