% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{browser_with_timeout}
\alias{browser_with_timeout}
\title{Browser with timeout}
\usage{
browser_with_timeout(
  text = "",
  condition = NULL,
  expr = TRUE,
  skipCalls = 0L,
  timeout = 10
)
}
\description{
This is a wrapper around \code{base::browser()} that asks the user for confirmation
before calling \code{base::browser()}. The confirmation must happen within \code{timeout}
seconds, otherwise no browser is launched and the execution continues as normal.
}
