% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halting.R
\name{call_info}
\alias{call_info}
\title{Get the details about where a call came from}
\usage{
call_info(print_msg = FALSE, level = 1)
}
\value{
A list with the following elements:
\code{fun} ... calling function
\code{dir} ... path to the script file
\code{file} ... name of the script file
\code{line} ... line in the script where called
\code{msg} ... message with the sequence
}
\description{
\code{call_info()} returns a list with information about the calling sequence.
Optionally, a message with the information is directly printed.
}
\concept{halting utilities provided by utilbox}
