% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halting.R
\name{caller_info}
\alias{caller_info}
\title{Print caller information}
\usage{
caller_info(fmtstring = NULL, level = 1)
}
\description{
\code{caller_info()} is a function that prints the caller information,
i.e. the sequence of calling fuctions and the invocation locations.
}
\details{
Taken from: https://stackoverflow.com/q/59537482/684229

Serves as inspiration for the function \code{get_call_info()}.
}
