% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caller.R
\name{caller_info2}
\alias{caller_info2}
\title{Print caller information}
\usage{
caller_info2(fmtstring = NULL, level = 1)
}
\description{
\code{caller_info2()} is a function that prints the caller information,
i.e. the sequence of calling fuctions and the invocation locations.
}
\details{
Taken from: https://stackoverflow.com/q/59537482/684229

Serves as inspiration for the function \code{caller_info()}.
}
