% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing.R
\name{cat0}
\alias{cat0}
\alias{catn}
\alias{catnn}
\title{Print}
\usage{
cat0(
  ...,
  file = "",
  sep = "",
  fill = FALSE,
  labels = NULL,
  append = FALSE,
  flush = TRUE,
  flush_cycle = 1,
  envir = utilbox_namespace()
)

catn(
  ...,
  file = "",
  sep = "",
  fill = FALSE,
  labels = NULL,
  append = FALSE,
  eol = "\\n"
)

catnn(
  ...,
  file = "",
  sep = "\\n",
  fill = TRUE,
  labels = NULL,
  append = FALSE,
  eol = "\\n"
)
}
\description{
`cat0` is basically an alias for [base::cat] that automatically 
flushes the console buffer (when `flush=TRUE`). This only affects 
buffered output enabled R sessions (e.g. Rgui on Windows)
}
\details{
`catn` is an alias for `cat0` which adds a single end-of-line symbol 
by default.
}
