% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.R
\name{cc}
\alias{cc}
\title{Evaluate an expression given as a string}
\usage{
cc(..., envir = parent.frame())
}
\description{
Takes an expression in a character vector, pastes it
into a single string and evaluates it in the \code{envir}
environment (the parent frame by default).
}
\examples{
x = NULL
print(x)
cc('x','=1')
print(x)

}
