% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_management.R
\name{check_same_code}
\alias{check_same_code}
\title{Compare R scripts}
\usage{
check_same_code(file1, file2, verbose = TRUE)
}
\value{
A logical indicating whether the two codes were found to
         to be identical together with the number of the first failed
         check (attribute \code{check_failed}, where value 0 indicates
         'no failure') and the names of the objects that were found 
         to be different (attribute \code{objects})..
}
\description{
Takes two R scripts and performs a comparison. First, the
actual codes are compared as character strings. If they fail
this sameness check, the two scripts are sourced into separate
environments, which are then compared. In this comparison, the
names of all objects are compared (check 1), then the classes 
of all objects (check 2), then functions are checked for identical 
arguments and bodies (check 3), and finally non-function objects 
are compared in terms of values (check 4).
}
