% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{class_huge}
\alias{class_huge}
\alias{as_huge}
\alias{add_class_huge}
\alias{print.huge}
\alias{as_huge.list}
\alias{as_huge.default}
\title{Convert objects to class `huge`}
\usage{
as_huge(x)

add_class_huge(x)

\method{print}{huge}(x, n = 10, show_note = TRUE)

\method{as_huge}{list}(x)

\method{as_huge}{default}(x)
}
\description{
Add `huge` as the first class of an object. Class `huge`
is intended to be used for objects with many elements which, 
if all printed, would clutter the screen. With class `huge` 
set, only a limited number of elements are actually printed.
(extracted by [base::head]).
}
\examples{
as_huge(list(1:100))

}
\seealso{
Other class-related functions provided by utilbox: 
\code{\link{classes}}
}
\concept{class-related functions provided by utilbox}
