% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{class_truncated}
\alias{class_truncated}
\alias{is_truncated}
\alias{[.truncated}
\alias{[[.truncated}
\alias{as.truncated.default}
\alias{print.truncated}
\title{rdname class_huge
export
Methods for class truncated}
\usage{
is_truncated(x)

\method{[}{truncated}(x, i)

\method{[[}{truncated}(x, i)

as.truncated.default(x)

\method{print}{truncated}(x)
}
\description{
rdname class_huge
export
Methods for class truncated
}
