% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock.R
\name{clock}
\alias{clock}
\alias{start_clock}
\alias{read_clock}
\alias{stop_clock}
\title{Clock control functions}
\usage{
clock(announce = TRUE, with_date = TRUE, lead = "", digs = 3, eol = "\\n")

start_clock(print = TRUE, envir = parent.frame(), lead = "", what = "")

read_clock(envir = utilbox_environment())

stop_clock(
  d1 = NULL,
  d12 = NULL,
  start_print = FALSE,
  end_print = TRUE,
  runtime_print = TRUE,
  envir = utilbox_environment(),
  lead = "",
  what = ""
)
}
\description{
\code{clock()} gets the current system time and optionally prints it on screen
\code{start_clock()} starts the runtime clock
\code{stop_clock()} stops the runtime clock
\code{read_clock()} retrieves the runtime clock
}
