% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{closest_day}
\alias{closest_day}
\title{Find the closest day among dates}
\usage{
closest_day(dates, day, position = FALSE)
}
\description{
Takes a vector of dates in \code{dates} and a single date in \code{day},
which are assumed to be in the 'YYYY-MM-DD' format, and find the
closest value in \code{dates} to the value in \code{day}. If \code{position} is
\code{TRUE}, the index of the closest date is returned, otherwise the
closest date itself is returned (default).
}
\examples{
closest_day(c('2010-01-01', '2010-03-15', '2012-06-20'), '2011-01-08')

}
\seealso{
Other calendar functions provided by utilbox: 
\code{\link{t_day}()},
\code{\link{y_day}()}
}
\concept{calendar functions provided by utilbox}
