% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{coalesce2}
\alias{coalesce2}
\title{Coalesce function}
\usage{
coalesce2(...)
}
\description{
A custom version of the coalesce function (to avoid dependence
on other packages).
}
\details{
Taken from https://stackoverflow.com/questions/19253820/how-to-implement-coalesce-efficiently-in-r
}
\examples{
coalesce2(c(NA,'a','b',NA), letters[1:4])

}
