% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{code_startup}
\alias{code_startup}
\alias{R_add_code_startup}
\alias{R_del_code_startup}
\title{Add/remove code to/from the .Rprofile file}
\usage{
R_add_code_startup(code, Rprof_file)

R_del_code_startup(
  code,
  Rprof_file,
  exact_match = FALSE,
  remove_all_matches = TRUE
)
}
\description{
`R_add_code_startup` adds code (supplied as character 
string vector `code`) and adds lines to the .Rprofile 
file for the code to evaluated at R's start up.
}
\details{
`R_del_code_startup` removes all lines that match the
supplied code.


Do not forget to double-backslash the supplied code.
}
\examples{
R_add_code_startup("print('Hi!')")   # check .Rprofile
R_del_code_startup("print('Hi!')")   # delete what was added

}
\seealso{
Other system-related functions provided by utilbox: 
\code{\link{.q}()},
\code{\link{R_gui_adjust}()},
\code{\link{R_session_info}()},
\code{\link{R_system_call_args}()},
\code{\link{append_system_PATH}()},
\code{\link{environment_variables}},
\code{\link{get_system_PATH}()},
\code{\link{is_win}()},
\code{\link{timest}()}
}
\concept{system-related functions provided by utilbox}
