% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{collapse}
\alias{collapse}
\alias{collapse0}
\alias{str_collapse}
\alias{collapse1}
\alias{collapse0n}
\alias{collapsen}
\alias{collapse0nq}
\title{Collapse a string vector}
\usage{
collapse0(x, ..., sep = "")

str_collapse(x, ..., sep = "")

collapse1(x, ..., sep = " ")

collapse0n(x, ..., sep = "\\n")

collapsen(x, ..., sep = "\\n")

collapse0nq(x, ..., sep = "'\\n'")
}
\description{
\code{collapse0} takes a vector and collapses it into a single string.
Equivalent to \code{\link[base:paste]{base::paste}} with \code{collapse=''}. The functions
\code{collapse1}, \code{collapse0n}, \code{collapse0nq} are aliases for \code{collapse0}
with different default values for \code{sep}.
}
\examples{
collapse0(c(0,1,1,1,0,1,0,0))

}
\seealso{
Other string-manipulation functions provided by utilbox: 
\code{\link{ascii}()},
\code{\link{cumpaste0}()},
\code{\link{multi_sub}()},
\code{\link{padding}},
\code{\link{patternize}()},
\code{\link{str_cut}()},
\code{\link{str_diff}()},
\code{\link{str_grab}()},
\code{\link{str_insert}()},
\code{\link{str_is_empty}()},
\code{\link{str_paste_grid}()},
\code{\link{str_pos}()},
\code{\link{str_rev.character}()},
\code{\link{str_trim_space}()},
\code{\link{sub_at}()},
\code{\link{toupperfirst}()}
}
\concept{string-manipulation functions provided by utilbox}
