% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{Colors}
\alias{Colors}
\title{A color palette}
\usage{
Colors(n = 20, skip = 0, randomize = FALSE, reverse = FALSE)
}
\description{
Returns a list of \code{n} interesting and contrasting colors (at least
as far as neighbours go). Up to 100 unique colors. For \code{n} above 100
the colors are recycled in a loop.
}
\examples{
Colors(1)                         # gives 1 (the first) color
Colors(5)                         # gives 5 colors from the beginning
Colors(21:25)                     # gives 5 colors from a different place
Colors(c('a','b','b'))            # gives 3 colors, the last two are the same
Colors(seq(0,1,l=10))             # gives colors from the entire palette sampled equidistantly

}
\seealso{
Other plotting-related functions provided by utilbox: 
\code{\link{add_box_above}()},
\code{\link{add_tight_axes}()},
\code{\link{all_devs}},
\code{\link{as_color}()},
\code{\link{get_legend}()},
\code{\link{hist2}()},
\code{\link{p16}()},
\code{\link{palette_ub}()},
\code{\link{show_Colors}()}
}
\concept{plotting-related functions provided by utilbox}
