% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{combine_names}
\alias{combine_names}
\title{Names for a list of combinations of vectors}
\usage{
combine_names(..., vars, sep1 = "=", sep2 = "_")
}
\description{
Produces names for a list of combinations of vectors. Takes vectors with 
parameter values and a vector of names and pastes them together in a 
cartesian product way. Can be used to names the elements of a list
which contains the results of a run of analysis for each combination 
on a grid (cartesian product) of parameter combinations.
}
\examples{
a = 1:3; b = 10:11; combine_names(a, b, names=c('a','b'))

}
\seealso{
Other string-manipulation functions provided by utilbox: 
\code{\link{ascii}()},
\code{\link{collapse}()},
\code{\link{cumpaste0}()},
\code{\link{padding}},
\code{\link{patternize}()},
\code{\link{str_add_punct}()},
\code{\link{str_diff}()},
\code{\link{str_insert}()},
\code{\link{str_is_empty}()},
\code{\link{str_last}()},
\code{\link{str_replace}()},
\code{\link{str_scrub_space}()},
\code{\link{str_split}()},
\code{\link{str_trim_space}()},
\code{\link{strpos}()},
\code{\link{strrev}()},
\code{\link{sub_at}()},
\code{\link{subm}()},
\code{\link{toupperfirst}()}
}
\concept{string-manipulation functions provided by utilbox}
