% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.R
\name{combine_pvalues}
\alias{combine_pvalues}
\title{P-value combination}
\usage{
combine_pvalues(
  P,
  lam,
  method = c("jelle", "jakub", "fisher", "stouffer"),
  summarize = FALSE,
  ...
)
}
\description{
Combines p-values using one of the available methods
}
