% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_compare.R
\name{compare_functions}
\alias{compare_functions}
\title{Compare functions}
\usage{
compare_functions(fun1, fun2, by_name = FALSE, envir1, envir2)
}
\description{
Compares two functions by looking at their arguments (via
\code{\link[base:args]{base::args()}}) and their bodies (via \code{\link[base:body]{base::body()}}).
}
\examples{
f1 = function (x) .Internal(which.max(x))
identical(f1, base::which.max)                              # returns FALSE, since they are defined in different environments
compare_functions(f1, base::which.max)                      # returns TRUE
compare_functions('f1', 'base::which.max', by_name=TRUE)    # returns TRUE

}
\seealso{
Other coding-related functions provided by utilbox: 
\code{\link{append_body}()},
\code{\link{call_to_list}},
\code{\link{fun_code_to_text}()},
\code{\link{hide}()},
\code{\link{hijack}()},
\code{\link{is_error}()},
\code{\link{namespaces}},
\code{\link{this_fun_name}()}
}
\concept{coding-related functions provided by utilbox}
