% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{concatenate}
\alias{concatenate}
\alias{\%.\%}
\alias{\%..\%}
\alias{\%_\%}
\alias{\%__\%}
\title{Concatenation operators}
\usage{
... \%.\% NA

... \%..\% NA

... \%_\% NA

... \%__\% sep
}
\description{
Concatenation operators, which are aliases for `base::paste`.
}
\details{
\code{\%.\%} is an operator versions of \code{base::paste0}.

\code{\%..\%} is an operator versions of \code{base::paste}.

\code{\%_\%} is an operator versions of \code{base::paste} with
`sep='_'`.

\code{\%__\%} is an operator versions of \code{base::paste} with
`sep='_'` and which takes symbols as arguments (see the examples below).
}
\examples{
a = 'multi'; b = 'tasking'
a \%.\% b            # yields 'multitasking'
a \%..\% b           # yields 'multi tasking'
a \%_\% b            # yields 'multi_tasking'
a \%__\% b           # BEWARE: yields 'a_b' (the inputs are treated as symbolic!)

}
\seealso{
Other operators provided by utilbox: 
\code{\link{\%anyin\%}()},
\code{\link{\%append\%}()},
\code{\link{\%c\%}()},
\code{\link{\%modify\%}()},
\code{\link{not_in}},
\code{\link{operator_NULL}},
\code{\link{operator_match}},
\code{\link{renaming_operators}}
}
\concept{operators provided by utilbox}
