% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{convert_unit}
\alias{convert_unit}
\title{Unit conversion}
\usage{
convert_unit(x, unit, append_unit = TRUE, ndigit = 3)
}
\description{
Converts which is assumed to be in unitary units (i.e. bytes) to an appropriate 
other unit
}
