% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encrypt.R
\name{crypt_files}
\alias{crypt_files}
\alias{decrypt_files}
\alias{encrypt_files}
\alias{encrypt_R_files}
\alias{decrypt_R_files}
\alias{generate_keys}
\title{Encrypt and decrypt files}
\usage{
crypt_files(
  files = NULL,
  pattern = NULL,
  path = ".",
  out_path = ".",
  appendix = "",
  key_file = "id_rsa.pub",
  delete_existing = FALSE,
  path_create = FALSE,
  action = c("encrypt", "decrypt")
)

decrypt_files(
  files = NULL,
  pattern = NULL,
  path = ".",
  out_path = ".",
  appendix = "",
  private_key_file = "id_rsa",
  delete_existing = FALSE,
  path_create = FALSE
)

encrypt_files(
  files = NULL,
  pattern = NULL,
  path = ".",
  out_path = ".",
  appendix = "",
  public_key_file = "id_rsa",
  delete_existing = FALSE,
  path_create = FALSE
)

encrypt_R_files(
  files = NULL,
  pattern = NULL,
  path = ".",
  appendix = "",
  out_path = "R_encrypted",
  delete_existing = TRUE,
  path_create = TRUE,
  public_key_file = file.path(out_path, "id_rsa.pub")
)

decrypt_R_files(
  files = NULL,
  pattern = NULL,
  path = "R_encrypted",
  appendix = "",
  out_path = "R_decrypted",
  delete_existing = TRUE,
  path_create = TRUE,
  private_key_file = file.path(path, "id_rsa")
)

generate_keys(
  private_key_file = "id_rsa",
  public_key_file = paste0(private_key_file, ".pub")
)
}
\description{
Utilities for encrypting and decrypting files using the package encryptr.
}
\details{
\code{encrypt_files()} encrypts the supplied files (directly via \code{files} and/or
via \code{pattern}, which is matched against the files in found in the path in
\code{path}). \code{decrypt_files()} achieves the opposite effect, i.e., it decrypts
the supplied files. Both function require key files (public key file for
\code{encrypt_files()} and private key file for \code{decrypt_files()}), which can
be generated using \code{generate_keys()}.

\code{encrypt_R_files()} and \code{decrypt_R_files()} are special cases of the above
functions that look for R script files (i.e., *.R) in the given \code{path} and
encrypt/decrypt them from and into the specified folders (i.e., 'R_decrypted'
and 'R_decrypted' relative to the current working directory as default).
}
\examples{
cat('print("hello world")', file='foo.R')
generate_keys()
encrypt_R_files('foo.R', appendix = '.backup', public_key_file = 'id_rsa.pub')
decrypt_R_files('foo.R.backup.encryptr.bin', private_key_file = 'id_rsa')
file_compare('foo.R', 'R_decrypted/foo.R.backup')

}
