% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{cumpaste0}
\alias{cumpaste0}
\title{Cumulative paste0}
\usage{
cumpaste0(x, .sep = "")
}
\description{
Takes a vector and cumulatively pastes it together.
}
\examples{
cumpaste0(c(0,1,1,1,0,1,0,0))

}
\seealso{
Other string-manipulation functions provided by utilbox: 
\code{\link{ascii}()},
\code{\link{collapse}()},
\code{\link{combine_names}()},
\code{\link{padding}},
\code{\link{patternize}()},
\code{\link{str_add_punct}()},
\code{\link{str_diff}()},
\code{\link{str_insert}()},
\code{\link{str_is_empty}()},
\code{\link{str_last}()},
\code{\link{str_replace}()},
\code{\link{str_scrub_space}()},
\code{\link{str_split}()},
\code{\link{str_trim_space}()},
\code{\link{strpos}()},
\code{\link{strrev}()},
\code{\link{sub_at}()},
\code{\link{subm}()},
\code{\link{toupperfirst}()}
}
\concept{string-manipulation functions provided by utilbox}
