% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revalue.R
\name{de_na}
\alias{de_na}
\title{Remove NA values}
\usage{
de_na(x, y, use_all_y = FALSE)
}
\description{
Replaces all \code{NA} elements in its first argument (\code{x})
with the corresponding values in its second argument (\code{y}).
Same effect as \code{dplyr::coalesce}, except when \code{use_all_y}
is \code{TRUE}, when the \code{NA} elements in \code{x}, say there
are \\emph{n} of them, are replaced with the first \emph{n} elements
of \code{y} (with possible recycling of \code{y}).
}
\examples{
de_na(c(1,2,NA,4,5,NA), 0)
de_na(c(1,2,NA,4,5,NA), 1:6)
de_na(c(1,2,NA,4,5,NA), Inf)

}
\concept{sequence-modifying functions provided by utilbox}
