% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{def_Sigma}
\alias{def_Sigma}
\title{Produce a variance matrix}
\usage{
def_Sigma(
  rho,
  N,
  nb = 1,
  random = FALSE,
  type = c("block", "band"),
  bandsize = 0
)
}
\description{
Produces a variance matrix with \code{rho} on the diagonal according to the 
specified type which can be either "random" (see \code{rvariance}) or has 
the structure according to \code{type}.
}
