% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizechecks.R
\name{dim_check}
\alias{dim_check}
\alias{is_scalar}
\alias{is_dimtwo}
\alias{is_dimtwoplus}
\title{Dimension check}
\usage{
is_scalar(x)

is_dimtwo(x)

is_dimtwoplus(x)
}
\description{
`is_scalar` checks whether the supplied is a scalar (vector, list, etc.).
}
\details{
`is_dimtwo` checks for two-dimensionality (e.g. matrix, data frame).

`is_dimtwoplus` checks for dimensionality 2 or more (e.g. matrix, data frame, 2D array, 3D array).
}
\seealso{
Other sequence-related functions provided by utilbox: 
\code{\link{as_factor}()},
\code{\link{every_nth}()},
\code{\link{filter_by_position}},
\code{\link{filter_by}},
\code{\link{groups_of_unique}()},
\code{\link{h1}()},
\code{\link{insert}()},
\code{\link{is_between}()},
\code{\link{is_empty}()},
\code{\link{is_run}},
\code{\link{last_element}()},
\code{\link{midpoints}()},
\code{\link{nunique}()},
\code{\link{or_between}()},
\code{\link{rep2}()},
\code{\link{replace2}()},
\code{\link{replace3}()},
\code{\link{seq2}()},
\code{\link{seq_around}()},
\code{\link{setdiffsym}()},
\code{\link{shift}()},
\code{\link{split_into_groups}()},
\code{\link{unname_factor}()},
\code{\link{what_bag}()}

Other sequence-related functions provided by utilbox: 
\code{\link{as_factor}()},
\code{\link{every_nth}()},
\code{\link{filter_by_position}},
\code{\link{filter_by}},
\code{\link{groups_of_unique}()},
\code{\link{h1}()},
\code{\link{insert}()},
\code{\link{is_between}()},
\code{\link{is_empty}()},
\code{\link{is_run}},
\code{\link{last_element}()},
\code{\link{midpoints}()},
\code{\link{nunique}()},
\code{\link{or_between}()},
\code{\link{rep2}()},
\code{\link{replace2}()},
\code{\link{replace3}()},
\code{\link{seq2}()},
\code{\link{seq_around}()},
\code{\link{setdiffsym}()},
\code{\link{shift}()},
\code{\link{split_into_groups}()},
\code{\link{unname_factor}()},
\code{\link{what_bag}()}

Other sequence-related functions provided by utilbox: 
\code{\link{as_factor}()},
\code{\link{every_nth}()},
\code{\link{filter_by_position}},
\code{\link{filter_by}},
\code{\link{groups_of_unique}()},
\code{\link{h1}()},
\code{\link{insert}()},
\code{\link{is_between}()},
\code{\link{is_empty}()},
\code{\link{is_run}},
\code{\link{last_element}()},
\code{\link{midpoints}()},
\code{\link{nunique}()},
\code{\link{or_between}()},
\code{\link{rep2}()},
\code{\link{replace2}()},
\code{\link{replace3}()},
\code{\link{seq2}()},
\code{\link{seq_around}()},
\code{\link{setdiffsym}()},
\code{\link{shift}()},
\code{\link{split_into_groups}()},
\code{\link{unname_factor}()},
\code{\link{what_bag}()}
}
\concept{sequence-related functions provided by utilbox}
