% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{dir_create}
\alias{dir_create}
\alias{dir_remove}
\alias{dir_slash_check}
\title{Create/remove a directory}
\usage{
dir_create(dir, ask = interactive(), verbose = FALSE)

dir_remove(dir)

dir_slash_check(dir, fsep = .Platform$file.sep)
}
\description{
\code{dir_create()} is a verbose version of \code{base::dir.create()}.

\code{dir_remove()} removes (unlinks) supplied directory/ies.

\code{dir_slash_check()} checks whether the argument (\code{dir}) ends with
a forward/backward slash and if not it appends one.
}
\seealso{
Other file system functions provided by utilbox: 
\code{\link{dir_exist_check}()},
\code{\link{file_path}()},
\code{\link{file_rename}()},
\code{\link{list_files}()}
}
\concept{file system functions provided by utilbox}
