% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{dir_exist_check}
\alias{dir_exist_check}
\title{Directory existence check}
\usage{
dir_exist_check(dir, stop_if_fail = TRUE, create_on_missing = TRUE, trace = 1)
}
\description{
Checks if directory \code{dir} exists (relative to the current working directory)
and if not it tries to create it. On success the value supplied in \code{dir} is returned. 
If directory creation fails, then an error is thrown unless \code{stop_if_fail=FALSE} 
was supplied, in which case the value \code{./} is returned. Verboseness is controled
via \code{trace} being set to 0 (no message) or otherwise.
}
\seealso{
Other file system functions provided by utilbox: 
\code{\link{dir_work}},
\code{\link{file_exists}()},
\code{\link{file_work}},
\code{\link{list_files}()}
}
\concept{file system functions provided by utilbox}
