% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.R
\name{do_pval_comb}
\alias{do_pval_comb}
\title{Combine p-values}
\usage{
do_pval_comb(
  P,
  lam = 1,
  method = "jelle",
  eps_p = 0.001,
  fac_lam = 0.01,
  Debug = FALSE,
  na.rm = TRUE,
  trace = 1
)
}
\description{
Performs combination of p-values using Fisher method or Box method
(the latter via Jelle's code in .pAsymptotic or Jakub's code below
}
