% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halting.R
\name{.halt}
\alias{.halt}
\title{Stop execution}
\usage{
.halt()
}
\description{
A different way of stopping the execution. Intended to be called by halt but the behavior
wasn't as hoped/expected. Abandoned for now.
}
