% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{.q}
\alias{.q}
\title{An alias for the quit function}
\usage{
.q(save = "no")
}
\description{
This does essentially the same as \code{base::q} except it quits without saving
or asking to save the current workspace.
}
\seealso{
Other system-related functions provided by utilbox: 
\code{\link{R_gui_adjust}()},
\code{\link{R_session_info}()},
\code{\link{R_system_call_args}()},
\code{\link{append_system_PATH}()},
\code{\link{code_startup}},
\code{\link{environment_variables}},
\code{\link{get_system_PATH}()},
\code{\link{is_win}()},
\code{\link{timest}()}
}
\concept{system-related functions provided by utilbox}
