% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{.rme}
\alias{.rme}
\title{Delete objects}
\usage{
.rme(..., envir = parent.frame())
}
\description{
\code{rme()} deletes all supplied objects in the environment specified
in \code{envir}. Unlike the basic \code{base::rm} function, the current
function checks if the supplied objects exist in order to avoid
errors. Objects can be supplied as character strings or as symbols.
}
\examples{
random_object = '1'
rme(random_object)      # delete successfully
rme(random_object)      # nothing to delete, but no error/warning

}
