% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{.sow}
\alias{.sow}
\title{Toggle stopping on warning}
\usage{
.sow(turn_on, announce = TRUE)
}
\description{
Sets `options()$warn` to either 0 or 2, which makes R either stop when a warning
is issued (>=2) or not (0). If argument `turn_on` is missing, the function acts
as a toggle turning stopping on warnings on and off on alternative calls.
}
