% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{download_file}
\alias{download_file}
\title{Download file}
\usage{
download_file(url, destfile = separate_path(url)$files, ...)
}
\description{
A wrapper for \code{download.file()} that handles missing files without
stopping on error. It uses the input URL to define the destination
file.
}
\examples{
download_file('http://www.karlin.mff.cuni.cz/~kpms/index.php')
}
