% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighing.R
\name{downweight}
\alias{downweight}
\title{Down-weighing}
\usage{
downweight(
  x,
  pivot,
  speed = 0.15,
  side = c("left", "right", "both"),
  nonnegative = FALSE
)
}
\description{
Down-weighs the values in \code{x} if they are below the pivot (\code{from})
at the rate of determined by \code{speed}. More specifically, the value of
\code{speed} is used as the standard deviation of the normal density which
is used to bring the values in \code{x} to zero as they deviate from the
pivot.

The argument \code{side} determines whether the down-weighing is
applied only to the values below the pivot (\code{side='left'}),
above the pivot (\code{side='right'}), or to both sides of the pivot
(\code{side='both'}).

The argument \code{speed} expects values between '0' (no down-weighing)
and '1' (maximum down-weighing). Values outside this range are set to
one of the two boundaries.
}
\examples{
downweight(1:10, 5)

}
\seealso{
Other numeric functions provided by utilbox: 
\code{\link{Invert}()},
\code{\link{as_numeric}()},
\code{\link{extract_by_value}},
\code{\link{force_as_}},
\code{\link{locate_position_of_value}},
\code{\link{lower_tri}()},
\code{\link{make_numeric}()},
\code{\link{min_dif}()},
\code{\link{permutations}},
\code{\link{replace_value}},
\code{\link{sequences}},
\code{\link{sets}},
\code{\link{to_base_k}()},
\code{\link{which_max}()}
}
\concept{numeric functions provided by utilbox}
