% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{drop_positive_negative}
\alias{drop_positive_negative}
\alias{positive}
\alias{negative}
\alias{nonnegative}
\alias{nonpositive}
\title{Drop elements based on their positiveness/negativeness}
\usage{
positive(x)

negative(x)

nonnegative(x)

nonpositive(x)
}
\description{
`positive` drops all non-positive elements in `x`.
}
\details{
`negative` drops all non-negative elements in `x`.

`nonpositive` drops all negative elements in `x`.

`nonnegative` drops all positive elements in `x`.
}
\seealso{
Other check-performing functions provided by utilbox: 
\code{\link{anyNonNA}()},
\code{\link{is_all_same}()},
\code{\link{is_decreasing}()},
\code{\link{is_even}()},
\code{\link{is_increasing}()},
\code{\link{is_integer}()},
\code{\link{is_na}()},
\code{\link{is_number}()},
\code{\link{is_odd}()},
\code{\link{is_permutation}()},
\code{\link{is_round}()},
\code{\link{is_scientific}()},
\code{\link{is_unique}()}
}
\concept{check-performing functions provided by utilbox}
