% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{edf}
\alias{edf}
\title{Empirical probability frequency (density)}
\usage{
edf(x)
}
\description{
\code{edf()} gives the empirical probability frequency (i.e. the
density).
}
\examples{
ecf(sample(1:3, 100, repl=TRUE))

}
