% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.R
\name{empty_symbol}
\alias{empty_symbol}
\alias{is_empty_symbol}
\title{Empty symbols}
\usage{
empty_symbol(x)

is_empty_symbol(x)
}
\description{
Empty symbols
}
\examples{
print(empty_symbol())  # prints the empty symbol
is_empty_symbol(empty_symbol())              # TRUE
is_empty_symbol(as.list(function(x) {})$x)   # TRUE

}
