% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{environment_variables}
\alias{environment_variables}
\alias{get_envir_var}
\alias{set_envir_var}
\title{Get and set environment variables}
\usage{
get_envir_var(name)

set_envir_var(name)
}
\description{
Just wrappers for the built-in R functions so that it
is easier to find them.
}
\details{
`set_envir_var` sets a system environment variable,
`get_envir_var` reads a value of a system
environment variable.
}
\seealso{
Other system-related functions provided by utilbox: 
\code{\link{.q}()},
\code{\link{R_gui_adjust}()},
\code{\link{R_session_info}()},
\code{\link{R_system_call_args}()},
\code{\link{append_system_PATH}()},
\code{\link{code_startup}},
\code{\link{get_system_PATH}()},
\code{\link{is_win}()},
\code{\link{timest}()}
}
\concept{system-related functions provided by utilbox}
