% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drives.R
\name{eval_with_drive}
\alias{eval_with_drive}
\title{Evaluate an expression}
\usage{
eval_with_drive(
  call,
  path = ".",
  drive = find_unmapped_drive(),
  unmap = TRUE,
  dir_create = FALSE,
  ask_new_dir = TRUE,
  silent = FALSE
)
}
\description{
Evaluate \code{call} with the effective path (either \code{path} when supplied
or the current working directory when not) mapped as a Windows drive.
This is useful when working with long paths and/or file names.
With \code{path = NULL} eval_with_drive simply evaluates the call without
any mapping of drives (i.e., behaves simply as \code{base::eval()}).
}
\examples{
## Use eval_with_drive() to work around long file names
# start by setting a long file name
long_file_name = paste(rep('a',255), collapse='')

# change path to something non-trivially long
orig_wdir = getwd()
setwd('~')

# Try opening png device with the long file name
try({png(long_file_name); dev.off()})
print(file.exists(long_file_name))        # FALSE

# Do the same but with the curren working directory mapped as drive T
eval_with_drive({png(long_file_name); dev.off()}, silent = TRUE)
# A regular call to file.exists still fails to find the file
print(file.exists(long_file_name))        # FALSE
print(eval_with_drive(file.exists(long_file_name), silent = TRUE))

# Clean up
eval_with_drive(file.remove(long_file_name), silent = TRUE)
setwd(orig_wdir)

}
