% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{every_nth}
\alias{every_nth}
\title{Get every n-th element of an object}
\usage{
every_nth(x, n, start = 1L)
}
\description{
Extracts every n-th element of an object (vector, list, etc.)
starting from 'start'.
}
\examples{
every_nth(1:10, 2)
every_nth(1:10, 2, 2)

}
\seealso{
Other sequence-related functions provided by utilbox: 
\code{\link{as_factor}()},
\code{\link{filter_by_position}},
\code{\link{filter_by}},
\code{\link{groups_of_unique}()},
\code{\link{h1}()},
\code{\link{insert}()},
\code{\link{is_between}()},
\code{\link{is_run}},
\code{\link{last_element}()},
\code{\link{rep2}()},
\code{\link{replace2}()},
\code{\link{replace3}()},
\code{\link{seq2}()},
\code{\link{seq_around}()},
\code{\link{sizechecks}},
\code{\link{split_into_groups}()},
\code{\link{what_bag}()}
}
\concept{sequence-related functions provided by utilbox}
