% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_by_value}
\alias{extract_by_value}
\alias{first_nonzero}
\alias{kth_nonzero}
\alias{last_nonzero}
\alias{first_positive}
\alias{last_positive}
\alias{first_above}
\alias{first_above_soft}
\alias{last_above}
\alias{last_above_soft}
\alias{first_below}
\alias{first_below_soft}
\alias{last_below}
\alias{last_below_soft}
\alias{first_nonempty}
\title{Locate value}
\usage{
first_nonzero(x)

kth_nonzero(x)

last_nonzero(x)

first_positive(x)

last_positive(x)

first_above(x, y)

first_above_soft(x, y)

last_above(x, y, ...)

last_above_soft(x, y)

first_below(x, y)

first_below_soft(x, y)

last_below(x, y)

last_below_soft(x, y)

first_nonempty(x)
}
\description{
\code{first_nonzero} finds the value of the last non-zero element in
\code{x}.

\code{kth_nonzero} finds the value of the n-th non-zero element in \code{x}.

\code{last_nonzero} finds the value of the last non-zero element in \code{x}.

\code{first_positive} finds the value of the first positive element in
\code{x}.

\code{last_positive} finds the value of the last positive element in
\code{x}.

\code{first_above} (\code{first_above_soft}) finds the value of the first
element in \code{x} with value larger than (larger than or equal to) \code{y}.

\code{last_above} (\code{last_above_soft}) finds the value of the last
element in \code{x} with value larger than (larger than or equal to) \code{y}.

\code{first_below} (\code{first_below_soft}) finds the value of the first
element in \code{x} with value smaller than (smaller than or equal to) \code{y}.

\code{last_below} (\code{last_below_soft}) finds the value of the last
element in \code{x} with value smaller than (smaller than or equal to) \code{y}.
}
\seealso{
Other numeric functions provided by utilbox: 
\code{\link{Invert}()},
\code{\link{downweight}()},
\code{\link{locate_position_of_value}},
\code{\link{lower_tri}()},
\code{\link{min_dif}()},
\code{\link{permutations}},
\code{\link{replace_value}},
\code{\link{sequences}},
\code{\link{sets}},
\code{\link{to_base_k}()},
\code{\link{which_max}()}
}
\concept{numeric functions provided by utilbox}
