% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{file_backup}
\alias{file_backup}
\title{Backup a file}
\usage{
file_backup(file, path, path_backup, pid = FALSE, announce = TRUE)
}
\description{
\code{file_backup()} creates a backup of the given file by creating a copy.
}
\seealso{
Other file system function provided by utilbox: 
\code{\link{check_file_locked}()},
\code{\link{clean_filename}()},
\code{\link{file_can_open_check}()},
\code{\link{file_empty}()},
\code{\link{file_size}()},
\code{\link{file_sort}()},
\code{\link{random_filename}()},
\code{\link{separate_path}()}
}
\concept{file system function provided by utilbox}
