% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{file_can_open_check}
\alias{file_can_open_check}
\title{File open check}
\usage{
file_can_open_check(filename)
}
\description{
Checks whether a file can be opened.
}
\seealso{
Other file system function provided by utilbox: 
\code{\link{check_file_locked}()},
\code{\link{clean_filename}()},
\code{\link{file_empty}()},
\code{\link{file_size}()},
\code{\link{file_sort}()},
\code{\link{random_filename}()},
\code{\link{separate_path}()}
}
\concept{file system function provided by utilbox}
