% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{file_empty}
\alias{file_empty}
\title{Check file existence and emptiness}
\usage{
file_empty(file)
}
\description{
\code{file_empty} checks if file exists, tries to read it to see
whether there is any data in it.
}
\seealso{
Other file system function provided by utilbox: 
\code{\link{check_file_locked}()},
\code{\link{clean_filename}()},
\code{\link{file_can_open_check}()},
\code{\link{file_size}()},
\code{\link{file_sort}()},
\code{\link{random_filename}()},
\code{\link{separate_path}()}
}
\concept{file system function provided by utilbox}
