% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{file_exists}
\alias{file_exists}
\alias{file_exists_check}
\alias{file_exists_should}
\alias{file_exists_must}
\title{Check file existence}
\usage{
file_exists(files, path, stop_on_missing = FALSE, warn_on_missing = TRUE)

file_exists_check(
  files,
  path,
  stop_on_missing = FALSE,
  warn_on_missing = FALSE
)

file_exists_should(
  files,
  path,
  stop_on_missing = FALSE,
  warn_on_missing = TRUE
)

file_exists_must(files, path, stop_on_missing = TRUE, warn_on_missing = TRUE)
}
\description{
`file_exists` checks if file(s) exist(s) relative to the path in `path`. 
If the existence check returns any `FALSE`, the behaviour depends on
`stop_on_missing` and `warn_on_missing`. If the former is `TRUE`, an 
error is printed and the execution halts. If only the former is `TRUE`,
a warning is shown but the execution continues. When both are `FALSE`
no message is shown and a logical indicating each file's existence is
returned.
}
\details{
In good code, it is advised to use one of the three following aliases:

`file_exists_checks` shows not messages and only returns the logicals.

`file_exists_should` warns on any files missing but proceeds and returns
the logicals.

`file_exists_must` is a strict version that halts on any files missing.
}
\seealso{
Other file system functions provided by utilbox: 
\code{\link{dir_exist_check}()},
\code{\link{dir_work}},
\code{\link{file_work}},
\code{\link{list_files}()}
}
\concept{file system functions provided by utilbox}
