% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{file_path}
\alias{file_path}
\alias{is_absolute_path}
\title{Work with files and paths}
\usage{
file_path(path, file, normalize = TRUE, fsep = .Platform$file.sep)

is_absolute_path(path)
}
\description{
\code{is_absolute_path()} checks whether the supplied path (\code{path}),
which may or may not include a file name, is absolute or not. On
Windows, absolute means that it begins with '\<drive letter\>:/', on
Linux/Unix it checks for '/' at the beginning of the path.

\code{file_path()} binds the supplied path (\code{path}) with the supplied
file name (\code{file}). If \code{file} contains absolute path or if path is an
an empty string, it ignores path. Otherwise, it concatenates them
using \code{fsep} as separator.
}
\examples{
file_path('home','text.txt')

}
\seealso{
Other file system functions provided by utilbox: 
\code{\link{dir_create}()},
\code{\link{dir_exist_check}()},
\code{\link{file_rename}()},
\code{\link{list_files}()}
}
\concept{file system functions provided by utilbox}
