% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{file_size}
\alias{file_size}
\title{Get file sizes}
\usage{
file_size(files)
}
\description{
\code{file_size()} returns the file sizes in appropriate units.
}
\seealso{
Other file system function provided by utilbox: 
\code{\link{check_file_locked}()},
\code{\link{clean_filename}()},
\code{\link{file_backup}()},
\code{\link{file_can_open_check}()},
\code{\link{file_empty}()},
\code{\link{file_sort}()},
\code{\link{map_drives}},
\code{\link{random_filename}()},
\code{\link{separate_path}()}
}
\concept{file system function provided by utilbox}
