% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{file_sort}
\alias{file_sort}
\alias{file_sort_time}
\title{Sort files names}
\usage{
file_sort(
  files,
  by = c("time", "name", "isdir", "mode", "mtime", "ctime", "atime", "exe"),
  decreasing = switch(by, time = , mtime = , ctime = , atime = TRUE, FALSE),
  nget = NULL
)

file_sort_time(files, by = "mtime", decreasing = TRUE)
}
\description{
\code{file_sort()} sorts files (specified in \code{files}) by their
attributes such as time or size. The attribute for sorting is
specified via \code{by}. The logical \code{decreasing} switches between
descending and ascending order.

\code{file_sort_time()} is a shortcut to sorting according to attribute
"mtime".
}
\seealso{
Other file system function provided by utilbox: 
\code{\link{check_file_locked}()},
\code{\link{clean_filename}()},
\code{\link{file_backup}()},
\code{\link{file_can_open_check}()},
\code{\link{file_empty}()},
\code{\link{file_size}()},
\code{\link{map_drives}},
\code{\link{random_filename}()},
\code{\link{separate_path}()}
}
\concept{file system function provided by utilbox}
