% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{file_work}
\alias{file_work}
\alias{file_rename}
\alias{file_remove}
\alias{file_rename_timestamp}
\alias{file_move}
\alias{file_timestamp}
\title{Rename or remove multiple files}
\usage{
file_rename(from, to, nretry = 1, time_to_sleep_prior = 0)

file_remove(file, nretry = 10)

file_rename_timestamp(filename, attrib = "mtime", nretry = 10)

file_move(files, destination, strip.dir = TRUE)

file_timestamp(filename, attrib = "mtime")
}
\description{
`file_rename` renames one or multiple files.
}
\details{
`file_remove` removes one or multiple files.

In case of a failure, the functions re-attempt up to `nretry` times.

`file_rename_timestamp` renames a single file by adding a time stamp 
based on the file's modification time.

`file_move` moves the file to the supplied path (`destination`)

`file_timestamp` produces the time stamp for renaming by `file_rename_timestamp`.
}
\seealso{
Other file system functions provided by utilbox: 
\code{\link{dir_exist_check}()},
\code{\link{dir_work}},
\code{\link{file_exists}()},
\code{\link{list_files}()}
}
\concept{file system functions provided by utilbox}
