% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{find_object}
\alias{find_object}
\title{Find objects in files}
\usage{
find_object(
  object_name,
  files = NULL,
  dir = ".",
  pattern = "^.*[.]RData$",
  stop_on_found = TRUE,
  announce = TRUE
)
}
\description{
Find an object with name in \code{object_name} inside files in \code{files} 
relative to the path supplied in \code{dir}. If a pattern is given (via 
\code{pattern}) instead of a list of file names, then all files that match 
the pattern are searched through looking for the object in \code{object_name}.
}
