% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drives.R
\name{find_unmapped_drive}
\alias{find_unmapped_drive}
\title{Find an unused drive letter}
\usage{
find_unmapped_drive(candidates = letters)
}
\description{
\code{find_unmapped_drive()} searches for an unmapped drive by going
through the letters of the alphabet (starting with 'a', 'b', etc.)
and returns the first letter for which there is no mapped drive.
Otherwise, it returns \code{FALSE}.
}
