% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_number.R
\name{force_as_}
\alias{force_as_}
\alias{force_as_integer}
\alias{force_as_real}
\title{Force type conversion}
\usage{
force_as_integer(x, ignore_sign = TRUE, na_val)

force_as_real(x, ignore_sign = TRUE, dec = ".", dec_fixed = TRUE, na_val)
}
\description{
Convert to type integer (\code{force_as_integer()}) or double
(\code{force_as_real()}) by stripping all non-number substrings in each
element of \code{x}.

This is a very radically forced conversion, which can be useful
when strange artefacts polluted the data, but caution is advised.
}
\examples{
# force as integers
force_as_integer(c('1','2'))
force_as_integer(c('1','x'))
force_as_integer(c('1','2.0','a3','-17.x','not-17.2','not-17.2-'))

# force as real numbers
force_as_real(c('1','2.0','a3','-17.x','not-17.2','not-17.2-'))
force_as_real(c('1','2.0','a3','-17.x','not-17.2','not-17.2-'), ignore_signFALSE)
force_as_real(c('1','x'))

}
\seealso{
Other numeric functions provided by utilbox: 
\code{\link{Invert}()},
\code{\link{as_numeric}()},
\code{\link{downweight}()},
\code{\link{extract_by_value}},
\code{\link{locate_position_of_value}},
\code{\link{lower_tri}()},
\code{\link{make_numeric}()},
\code{\link{min_dif}()},
\code{\link{permutations}},
\code{\link{replace_value}},
\code{\link{sequences}},
\code{\link{sets}},
\code{\link{to_base_k}()},
\code{\link{which_max}()}

Other numeric functions provided by utilbox: 
\code{\link{Invert}()},
\code{\link{as_numeric}()},
\code{\link{downweight}()},
\code{\link{extract_by_value}},
\code{\link{locate_position_of_value}},
\code{\link{lower_tri}()},
\code{\link{make_numeric}()},
\code{\link{min_dif}()},
\code{\link{permutations}},
\code{\link{replace_value}},
\code{\link{sequences}},
\code{\link{sets}},
\code{\link{to_base_k}()},
\code{\link{which_max}()}
}
\concept{numeric functions provided by utilbox}
