% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{frac}
\alias{frac}
\title{Lagged ratios}
\usage{
frac(x)
}
\description{
Similar to \code{base::diff} except returns lagged ratios instead of lagged differences.
}
\seealso{
Other numeric functions provided by utilbox: 
\code{\link{Invert1}()},
\code{\link{Invert}()},
\code{\link{all_permutations}()},
\code{\link{as_numeric}()},
\code{\link{downweight}()},
\code{\link{force_as_}},
\code{\link{frac_part}()},
\code{\link{int_part}()},
\code{\link{lower_tri}()},
\code{\link{make_numeric}()},
\code{\link{max0}()},
\code{\link{min_dif}()},
\code{\link{mintersect}()},
\code{\link{round2}()},
\code{\link{round_nearest_power}()},
\code{\link{round_nearest}()},
\code{\link{rsignif}()},
\code{\link{signif2}()},
\code{\link{to_base_k}()},
\code{\link{unspace}()},
\code{\link{unzero}()},
\code{\link{upper_tri}()},
\code{\link{value_locale_position}},
\code{\link{value_locale}},
\code{\link{which_max}()}
}
\concept{numeric functions provided by utilbox}
