% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.R
\name{fun2call}
\alias{fun2call}
\title{Convert a function to a call}
\usage{
fun2call(fun)
}
\description{
Takes a function and returns an object of class \code{call} with the
code of the function.
}
\examples{
# compare with the output of print(sample)
code = fun2call(sample)

}
