% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.R
\name{fun_code_to_text}
\alias{fun_code_to_text}
\title{Get the source code of a function}
\usage{
fun_code_to_text(fun, file = NULL)
}
\arguments{
\item{fun}{function for which the body of the code is desired}
}
\value{
A character vector with each command of the input function as an element.
}
\description{
`fun_code_to_text` returns the body of the supplied function
as a character string.
}
\details{
This function takes an R function on input and returns its source code as character vector.
}
\examples{
f <- function(x) { 
  print(x)
  x = 2
  invisible(x)
}
ff <- fun_code_to_text(f)
print(ff)

}
\seealso{
Other coding-related functions provided by utilbox: 
\code{\link{append_body}()},
\code{\link{call_to_list}},
\code{\link{compare_functions}()},
\code{\link{get_package_namespace}()},
\code{\link{hijack}()},
\code{\link{this_fun_name}()}
}
\concept{coding-related functions provided by utilbox}
