% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fun_dump_code}
\alias{fun_dump_code}
\title{Dump the source code of a function to a file}
\usage{
fun_dump_code(fun, file)
}
\description{
Dumps the source code of a function \code{fun} to a file \code{file}.
}
\seealso{
Other coding-related functions provided by utilbox: 
\code{\link{append_body}()},
\code{\link{call_to_list}},
\code{\link{compare_functions}()},
\code{\link{fun_code_to_text}()},
\code{\link{hide}()},
\code{\link{hijack}()},
\code{\link{is_error}()},
\code{\link{namespaces}},
\code{\link{this_fun_name}()}
}
\concept{coding-related functions provided by utilbox}
