% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fun_separate_lines}
\alias{fun_separate_lines}
\title{Separate lines of the code of a function}
\usage{
fun_separate_lines(fun)
}
\description{
\code{fun_separate_lines()} converts the body of a function (\code{fun}) to a list where
each element in the list corresponds to one line for the purposes of altering
the code for instance via \code{append_body()} below or via \code{base::trace()}.
}
