% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{function_separate_code}
\alias{function_separate_code}
\title{Separate lines of the code of a function}
\usage{
function_separate_code(fun)
}
\description{
\code{function_separate_code()} converts the body of a function (\code{fun}) to a list
where each element in the list corresponds to a block of code (often a
line). This is used when altering the code for instance via \code{append_body()}
below or via \code{base::trace()}.
}
\seealso{
Other coding-related functions provided by utilbox: 
\code{\link{call_to_list}},
\code{\link{compare_functions}()},
\code{\link{fun_name}},
\code{\link{function_code_to_text}()},
\code{\link{function_dump_code}()},
\code{\link{function_find_dependencies}()},
\code{\link{function_modify}},
\code{\link{hide}()},
\code{\link{hijack}()},
\code{\link{is_error}()},
\code{\link{namespaces}},
\code{\link{nformals}()},
\code{\link{null}()}
}
\concept{coding-related functions provided by utilbox}
