% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{get2}
\alias{get2}
\title{Get an object}
\usage{
get2(what, envir, default_value, mode = "any", inherits = TRUE)
}
\description{
\code{get2()} gets an object from the given environment (\code{envir}). Checks if it exists
first and if the object does not exist it returns the value in
\code{default_value}. If \code{envir} is missing, it simply looks for the object
names 'what' on the search path.
}
\examples{
get2('fklasdfjskadfjlsd', envir=.GlobalEnv, NA)

}
