% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{get_all_objects}
\alias{get_all_objects}
\title{Lists objects}
\usage{
get_all_objects(files = NULL, dir = ".", pattern = "^.*[.]RData$")
}
\description{
Lists all objects inside given files or inside files that match pattern in 
\code{pattern} relative to the path in \code{dir}
}
