% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_legend}
\alias{get_legend}
\title{Extract legend from a ggplot}
\usage{
get_legend(p, direction = NULL, nrow_color = NULL)
}
\description{
For usage see for instance the following link or the example below.

See also https://github.com/hadley/ggplot2/wiki/Share-a-legend-between-two-ggplot2-graphs.
}
\examples{
# Example usage:
\dontrun{
require(ggplot)
require(dplyr)
p = tibble(x=1:100, y=rnorm(100), z=rep(letters[1:10], e=10)) \%>\% ggplot(aes(x=x, y=y, color=z)) + geom_point()
leg = get_legend(p, direction='horizontal', nrow_color=1)
p = p + theme(legend.position='none')
gridExtra::grid.arrange(arrangeGrob(p), arrangeGrob(leg), nrow=2, heights=c(4.5,0.5))
}

}
\seealso{
Other plotting-related functions provided by utilbox: 
\code{\link{Colors}()},
\code{\link{add_box_above}()},
\code{\link{add_tight_axes}()},
\code{\link{all_devs}},
\code{\link{as_color}()},
\code{\link{hist2}()},
\code{\link{p16}()},
\code{\link{palette_ub}()},
\code{\link{show_Colors}()}
}
\concept{plotting-related functions provided by utilbox}
