% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{get_my_ip}
\alias{get_my_ip}
\title{Public IP address}
\usage{
get_my_ip(server = "https://api.ipify.org", trace = 1)
}
\description{
Gets ones public IP address using the ipify.com API. The default server can be changed.
}
\details{
The function probably needs a rewrite for when the call to the server fails for some reason.
}
