% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{get_package_namespace}
\alias{get_package_namespace}
\title{Get the namespace of a package}
\usage{
get_package_namespace(pckg, character.only = FALSE, stop_on_error = FALSE)
}
\description{
Returns the namespace of the package with name in `pckg`.
}
\examples{
get_package_namespace('base')

}
\seealso{
Other coding-related functions provided by utilbox: 
\code{\link{append_body}()},
\code{\link{call_to_list}},
\code{\link{compare_functions}()},
\code{\link{fun_code_to_text}()},
\code{\link{hijack}()},
\code{\link{this_fun_name}()}
}
\concept{coding-related functions provided by utilbox}
