% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{get_system_PATH}
\alias{get_system_PATH}
\title{System search paths}
\usage{
get_system_PATH(path)
}
\description{
\code{get_system_PATH()} reads and returns the system search paths.

\code{append_system_PATH()} adds the supplied path among the system
search paths.
}
\examples{
get_system_PATH()

}
\seealso{
Other system-related functions provided by utilbox: 
\code{\link{R_defaults}},
\code{\link{R_gui_adjust}()},
\code{\link{R_session_info}()},
\code{\link{append_system_PATH}()},
\code{\link{environment_variables}},
\code{\link{is_win}()},
\code{\link{pause}},
\code{\link{stopping_control}},
\code{\link{timest}()}
}
\concept{system-related functions provided by utilbox}
