% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%anyin\%}
\alias{\%anyin\%}
\title{Any-in-set operator}
\usage{
x \%anyin\% table
}
\description{
Similar to the 'in-set' operator except the return value is
reduced to a scalar via \code{any}.
}
\examples{
0:1 \%anyin\% 1:10

}
\seealso{
Other operators provided by utilbox: 
\code{\link{\%modify\%}()},
\code{\link{concatenate}},
\code{\link{not_in}},
\code{\link{operator_NULL}},
\code{\link{operator_append}},
\code{\link{operator_c}},
\code{\link{operator_match}},
\code{\link{renaming_operators}}
}
\concept{operators provided by utilbox}
