% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%append\%}
\alias{\%append\%}
\title{Append operator}
\usage{
k \%append\% l
}
\value{
A list.
}
\description{
Append one object to another via [base::append].
}
\examples{
list(a=1, b=2) \%append\% list(c=3)    # adds an element to the list on the LHS
c(1,2) \%append\% list(c=3)            # turns the LHS into a list, then appends

}
\seealso{
Other operators provided by utilbox: 
\code{\link{\%anyin\%}()},
\code{\link{\%c\%}()},
\code{\link{\%modify\%}()},
\code{\link{concatenate}},
\code{\link{not_in}},
\code{\link{operator_NULL}},
\code{\link{operator_match}},
\code{\link{renaming_operators}}
}
\concept{operators provided by utilbox}
