% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%modify\%}
\alias{\%modify\%}
\title{List modify operator}
\usage{
left \%modify\% right
}
\description{
Merges two list together while updating the elements of the first
list with the elements of the second list for the elements found in
both lists.
}
\examples{
list(a=1, b=2) \%modify\% list(b=3, c=4)

}
\seealso{
Other operators provided by utilbox: 
\code{\link{\%anyin\%}()},
\code{\link{concatenate}},
\code{\link{not_in}},
\code{\link{operator_NULL}},
\code{\link{operator_append}},
\code{\link{operator_c}},
\code{\link{operator_match}},
\code{\link{operator_try}},
\code{\link{renaming_operators}}
}
\concept{operators provided by utilbox}
