% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping.R
\name{groups_of_unique}
\alias{groups_of_unique}
\title{Group belonging indices}
\usage{
groups_of_unique(x, format = as.numeric)
}
\description{
For each element in \code{x}, \code{groups_of_unique()} determines which
group (first group, second group, etc.) the element belongs to and
returns the group indices. The groups are determined based on the
order of appearance of unique elements in \code{x}.

This is basically an alias for \code{replace3} with \code{new_values} equal
to \code{1:nunique(x)}.
}
\examples{
groups_of_unique(c('b','b','aa', 'a','b','c','a'))

}
\seealso{
Other sequence-related functions provided by utilbox: 
\code{\link{every_kth}()},
\code{\link{filter_by_position}},
\code{\link{filter_by}},
\code{\link{h1}()},
\code{\link{insert}()},
\code{\link{is_between}()},
\code{\link{is_run}},
\code{\link{last_element}()},
\code{\link{rep2}()},
\code{\link{replace2}()},
\code{\link{replace3}()},
\code{\link{seq2}()},
\code{\link{seq_around}()},
\code{\link{sizechecks}},
\code{\link{split_into_groups}()},
\code{\link{to_factor}()},
\code{\link{what_bag}()}
}
\concept{sequence-related functions provided by utilbox}
