% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halting.R
\name{halt}
\alias{halt}
\title{Print error and stop execution}
\usage{
halt(error = "")
}
\description{
Prints an error message and stops. The actual way of stopping, however,
depends on whether the code is run in interactive or non-interactive
modes
}
