% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtp.R
\name{hartung.test}
\alias{hartung.test}
\title{Hartung test for dependent p-values}
\usage{
hartung.test(p, alpha = 0.05, kappa = 0.2, what = "ns")
}
\description{
Requires library \code{punitroots} which can be found at r-forge (see examples below for installation).
}
\examples{
install.packages("urca")
install.packages("CADFtest")
install.packages("punitroots", repos="http://r-forge.r-project.org")
}
