% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorchecks.R
\name{hide}
\alias{hide}
\title{Hide input}
\usage{
hide(x, ...)
}
\description{
Takes input and returns it invisibly. Useful for instance for
hiding errors and warning messages within \code{\link[base:tryCatch]{base::tryCatch}} (see the
examples).
}
\examples{
tryCatch(as.numeric('a'), warning=hide)
tryCatch(sin('a'), error=hide)

}
\seealso{
Other coding-related functions provided by utilbox: 
\code{\link{append_body}()},
\code{\link{call_to_list}},
\code{\link{compare_functions}()},
\code{\link{fun_code_to_text}()},
\code{\link{hijack}()},
\code{\link{is_error}()},
\code{\link{namespaces}},
\code{\link{this_fun_name}()}
}
\concept{coding-related functions provided by utilbox}
