% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hijack}
\alias{hijack}
\title{Argument hijack function}
\usage{
hijack(FUN, ...)
}
\description{
A function that allows for changing of default values for
arguments of other functions. Similar to \code{\link[base:formals]{base::formals}} but nicer
syntax.
}
\examples{
cat = hijack(cat, sep='')        # equivalent to cat0
mean = hijack(mean, na.rm=TRUE)  # equivalent to mean2

}
\seealso{
Other coding-related functions provided by utilbox: 
\code{\link{append_body}()},
\code{\link{call_to_list}},
\code{\link{compare_functions}()},
\code{\link{fun_code_to_text}()},
\code{\link{hide}()},
\code{\link{is_error}()},
\code{\link{namespaces}},
\code{\link{this_fun_name}()}
}
\concept{coding-related functions provided by utilbox}
