% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{hist2}
\alias{hist2}
\title{Histogram}
\usage{
hist2(..., nbreaks)
}
\description{
A wrapper for \code{base::hist}, which allows to specify the
number of bins by through \code{nbreaks}.
}
\seealso{
Other plotting-related functions provided by utilbox: 
\code{\link{Colors}()},
\code{\link{add_box_above}()},
\code{\link{add_tight_axes}()},
\code{\link{all_devs}},
\code{\link{as_color}()},
\code{\link{get_legend}()},
\code{\link{p16}()},
\code{\link{palette_ub}()},
\code{\link{show_Colors}()}
}
\concept{plotting-related functions provided by utilbox}
