% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtp.R
\name{hmp_test}
\alias{hmp_test}
\title{Harmonic mean p-value test}
\usage{
hmp_test(p, alpha = 0.05, what = "ns")
}
\description{
This implements the global test based on the harmonic mean 
p-value, which is a combination p-value that is insensitive
to dependence among the combined p-values.
}
\details{
Requires package \code{harmonicmeanp}.
}
