% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\name{identical_cols}
\alias{identical_cols}
\title{Check for identical columns}
\usage{
identical_cols(x, w)
}
\arguments{
\item{x}{A two-dimensional object to be checked. @description}

\item{w}{(Optional) A vector giving a subset of columns to be
considered.}
}
\value{
Logical indicating whether all of the selected columns
are identical.
}
\description{
\code{identical_cols()} takes a matrix, data frame or any
two-dimensional object (array) and checks whether all (or a subset
of, when \code{w} given) its columns are identical in terms of content,
not necessarily in terms of equality of column names.
}
\examples{
identical_cols(data.frame(x=1:10, y=letters[1:10]))
identical_cols(data.frame(x=1:10, y=1:10))

}
\seealso{
Other matrix/data-frame functions provided by utilbox: 
\code{\link{add_col}()},
\code{\link{join_recurse}()},
\code{\link{rep_df}()},
\code{\link{rev_rows}()},
\code{\link{split_rows}()}
}
\concept{matrix/data-frame functions provided by utilbox}
