% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditioning.R
\name{ifelsem}
\alias{ifelsem}
\title{Conditional element selection with (possible) non-scalar response}
\usage{
ifelsem(test, yes, no, simplify = FALSE, do_not_recycle_no = FALSE)
}
\description{
\code{ifelsem} is an extension of \code{\link[base:ifelse]{base::ifelse}} which allows
non-scalar response. It returns a list. An attempt at simplifying the
response is made when \code{simplify=TRUE}.
}
