% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert.R
\name{Invert}
\alias{Invert}
\alias{Invert1}
\title{Find (pseudo)inverse, square root, and sqrt-inverse matrices}
\usage{
Invert(M, inv = FALSE, sq = FALSE, sqinv = FALSE, eps = 1e-12, drop = FALSE)

Invert1(x)
}
\description{
\code{Invert1()} does element-wise inversion (as in \"one over\") of
all non-zero elements \code{x}, while the zero-valued elements are kept
unchanged.

\code{Invert()} finds an inverse (or pseudoinverse), square root,
and/or sqrt-inverse matrices of \code{M}.
}
\seealso{
Other numeric functions provided by utilbox: 
\code{\link{downweight}()},
\code{\link{extract_by_value}},
\code{\link{locate_position_of_value}},
\code{\link{lower_tri}()},
\code{\link{min_dif}()},
\code{\link{permutations}},
\code{\link{replace_value}},
\code{\link{sequences}},
\code{\link{sets}},
\code{\link{to_base_k}()},
\code{\link{which_max}()}
}
\concept{numeric functions provided by utilbox}
