% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.R
\name{is_assign_call}
\alias{is_assign_call}
\title{Guess whether an expression is an assignment}
\usage{
is_assign_call(expr, stop_on_error = FALSE)
}
\description{
Checks whether an expression has an assignement form, namely
that it is a call to the \verb{<-} or other such functions.
}
\examples{
is_assign_call(expression(x <- 1))    # TRUE
is_assign_call(expression(x = 1))    # TRUE
is_assign_call(expression(2*x))       # FALSE

}
