% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizechecks.R
\name{is_empty}
\alias{is_empty}
\title{Zero-length check}
\usage{
is_empty(...)
}
\description{
1is_empty` checks whether the supplied object has zero
no elements.
}
\details{
`is_nonempty` is a negation of `is_empty.

The function `is_empty` is similar to [rlang::is_empty], 
except applicable also to data frames, for which it 
derives length from the production of the dimension 
(obtained via `dim`).
}
\seealso{
Other sequence-related functions provided by utilbox: 
\code{\link{as_factor}()},
\code{\link{dim_check}},
\code{\link{every_nth}()},
\code{\link{filter_by_position}},
\code{\link{filter_by}},
\code{\link{groups_of_unique}()},
\code{\link{h1}()},
\code{\link{insert}()},
\code{\link{is_between}()},
\code{\link{is_run}},
\code{\link{last_element}()},
\code{\link{midpoints}()},
\code{\link{nunique}()},
\code{\link{or_between}()},
\code{\link{rep2}()},
\code{\link{replace2}()},
\code{\link{replace3}()},
\code{\link{seq2}()},
\code{\link{seq_around}()},
\code{\link{setdiffsym}()},
\code{\link{shift}()},
\code{\link{split_into_groups}()},
\code{\link{unname_factor}()},
\code{\link{what_bag}()}
}
\concept{sequence-related functions provided by utilbox}
