% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_script.R
\name{is_empty_line}
\alias{is_empty_line}
\title{\code{is_empty_line()} identifies "empty" lines, where "empty" means
that there is at most "#'" on the line.}
\usage{
is_empty_line(x)
}
\description{
\code{is_empty_line()} identifies "empty" lines, where "empty" means
that there is at most "#'" on the line.
}
