% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.R
\name{is_error}
\alias{is_error}
\title{Checks for an error}
\usage{
is_error(x)
}
\description{
Checks whether an object is a result of a failer (error throwing)
call by checking whether it is of class `try-error` (which is 
produced by [`base::try()`]).
}
\examples{
is_error(try(x = .this.is.a.missing.variable))

}
