% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numchecks.R
\name{is_increasing}
\alias{is_increasing}
\title{Check if a sequence in 'x' is increasing}
\usage{
is_increasing(x, strictly = FALSE)
}
\description{
Check if a sequence in 'x' is increasing
}
\seealso{
Other check-performing functions provided by utilbox: 
\code{\link{anyNonNA}()},
\code{\link{drop_positive_negative}},
\code{\link{is_all_same}()},
\code{\link{is_decreasing}()},
\code{\link{is_even}()},
\code{\link{is_integer}()},
\code{\link{is_na}()},
\code{\link{is_number}()},
\code{\link{is_odd}()},
\code{\link{is_permutation}()},
\code{\link{is_round}()},
\code{\link{is_scientific}()},
\code{\link{is_unique}()},
\code{\link{last_nonzero}()}
}
\concept{check-performing functions provided by utilbox}
