% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numchecks.R
\name{is_integer}
\alias{is_integer}
\title{Check if elements in 'x' are effectively integers (i.e. close enough to one)}
\usage{
is_integer(x, tol = .Machine$double.eps)
}
\description{
Check if elements in 'x' are effectively integers (i.e. close enough to one)
}
\seealso{
Other check-performing functions provided by utilbox: 
\code{\link{anyNonNA}()},
\code{\link{drop_positive_negative}},
\code{\link{is_all_same}()},
\code{\link{is_decreasing}()},
\code{\link{is_even}()},
\code{\link{is_increasing}()},
\code{\link{is_na}()},
\code{\link{is_number}()},
\code{\link{is_odd}()},
\code{\link{is_permutation}()},
\code{\link{is_round}()},
\code{\link{is_scientific}()},
\code{\link{is_unique}()},
\code{\link{last_nonzero}()}
}
\concept{check-performing functions provided by utilbox}
