% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{is_list_class}
\alias{is_list_class}
\title{Checks whether the object is of class "list"}
\usage{
is_list_class(x)
}
\description{
\code{is_list_class()} checks whether the object is of class "list"
more strictly than \code{base::is.list()}, which returns \code{TRUE} not
only for objects of class \code{"list"} but also for data frames.
}
\examples{
is_list_class(list(1))
is_list_class(data.frame(1))

}
\seealso{
Other list utilities provided by utilbox: 
\code{\link{is_all_character}()},
\code{\link{list_clean}()},
\code{\link{list_empty}()},
\code{\link{list_flatten}()},
\code{\link{list_mirror}()},
\code{\link{list_set_attr}()},
\code{\link{list_update}()},
\code{\link{merge_recurse}()},
\code{\link{nlist}()}
}
\concept{list utilities provided by utilbox}
