% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numchecks.R
\name{is_na}
\alias{is_na}
\title{Check for \code{NA/NaN} values}
\usage{
is_na(x, freduce = all)
}
\description{
Checks if a reduction of elements of \code{x} is \code{NA}. The default 
reduction function is \code{all}, which returns \code{TRUE} whenever all elements 
of its argument are \code{NA} or `x` is 0-length. A useful alternative reduction function is 
\code{any} supplied via the argument \code{freduce}.
}
\seealso{
Other check-performing functions provided by utilbox: 
\code{\link{anyNonNA}()},
\code{\link{drop_positive_negative}},
\code{\link{is_all_same}()},
\code{\link{is_decreasing}()},
\code{\link{is_even}()},
\code{\link{is_increasing}()},
\code{\link{is_integer}()},
\code{\link{is_number}()},
\code{\link{is_odd}()},
\code{\link{is_permutation}()},
\code{\link{is_round}()},
\code{\link{is_scientific}()},
\code{\link{is_unique}()}
}
\concept{check-performing functions provided by utilbox}
