% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutate.R
\name{is_permutation}
\alias{is_permutation}
\title{Check for permutation}
\usage{
is_permutation(x, y)
}
\description{
Checks if the elements in the first argument (\code{x}) correspond to 
a permutation of the second argument (\code{y}).
}
\seealso{
Other check-performing functions provided by utilbox: 
\code{\link{anyNonNA}()},
\code{\link{drop_positive_negative}},
\code{\link{is_all_same}()},
\code{\link{is_decreasing}()},
\code{\link{is_even}()},
\code{\link{is_increasing}()},
\code{\link{is_integer}()},
\code{\link{is_na}()},
\code{\link{is_number}()},
\code{\link{is_odd}()},
\code{\link{is_round}()},
\code{\link{is_scientific}()},
\code{\link{is_unique}()},
\code{\link{last_nonzero}()}
}
\concept{check-performing functions provided by utilbox}
