% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqchecks.R
\name{is_run}
\alias{is_run}
\alias{is_start_of_run}
\alias{is_end_of_run}
\title{Starts and ends of run}
\usage{
is_start_of_run(x, is_first_start = TRUE, tol)

is_end_of_run(x, is_last_end = TRUE, tol)
}
\description{
Indicators of the beginnings (\code{is_start_of_run}) and ends
(\code{is_end_of_run}) of "runs", where "runs" are subsequences with
unchanging values.
}
\examples{
x=1:10
is_end_of_run(x)           # all TRUE
is_start_of_run(x)         # all TRUE

x = c(1,2,2,2,3,3,4)
is_end_of_run(x)           # 1 4 6 are TRUE
is_start_of_run(x)         # 1 2 5 7 are TRUE

}
\seealso{
Other sequence-related functions provided by utilbox: 
\code{\link{every_kth}()},
\code{\link{filter_by_position}},
\code{\link{filter_by}},
\code{\link{groups_of_unique}()},
\code{\link{h1}()},
\code{\link{insert}()},
\code{\link{is_between}()},
\code{\link{last_element}()},
\code{\link{rep2}()},
\code{\link{replace2}()},
\code{\link{replace3}()},
\code{\link{seq2}()},
\code{\link{seq_around}()},
\code{\link{sizechecks}},
\code{\link{split_into_groups}()},
\code{\link{to_factor}()},
\code{\link{what_bag}()}
}
\concept{sequence-related functions provided by utilbox}
