% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{is_same_environment}
\alias{is_same_environment}
\title{Comparison of environments}
\usage{
is_same_environment(x, y, assume_package_x = TRUE, assume_package_y = TRUE)
}
\description{
Compares two environments, which can be supplied either directly
or by name (when \code{x} and/or \code{y} are of type \code{character}). If supplied
by name, the names in \code{x} and \code{y} are front-appended by 'package:'
unless \code{assume_package_x} and \code{assume_package_y} are FALSE,
respectively.
}
\examples{
is_same_environment(topenv(ls), 'base')
is_same_environment(e1 <- new.env(), e1)

assign('xyzXYZ', function() {}, envir=.GlobalEnv)
is_same_environment(.GlobalEnv, environment(xyzXYZ))

}
