% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numchecks.R
\name{is_same_value}
\alias{is_same_value}
\title{Check if two objects have equal values (simultaneous NAs are considered equal)}
\usage{
is_same_value(x, y)
}
\description{
\code{is_same_value()} checks if the elements of two equally sizes objects are either
simultaneously \code{NA} or have equal non-\code{NA} values.
}
\examples{
is_same_value(1:5, 1:5)               # TRUE
is_same_value(c(NA,2:5), c(NA,2:5)    # also TRUE
is_same_value(1:5, c(NA,2:5)          # FALSE
is_same_value(0:4, 1:5)               # FALSE

}
\seealso{
Other check-performing functions provided by utilbox: 
\code{\link{anyNonNA}()},
\code{\link{extract_by_positiveness_negativeness}},
\code{\link{is_number}()},
\code{\link{numchecks}},
\code{\link{rounding}}
}
\concept{check-performing functions provided by utilbox}
