% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numchecks.R
\name{is_scientific}
\alias{is_scientific}
\title{Check if 'x' is in scientific notation (e.g. 1.234E5)}
\usage{
is_scientific(x, freduce = all)
}
\description{
Check if 'x' is in scientific notation (e.g. 1.234E5)
}
\seealso{
Other check-performing functions provided by utilbox: 
\code{\link{anyNonNA}()},
\code{\link{drop_positive_negative}},
\code{\link{is_all_same}()},
\code{\link{is_decreasing}()},
\code{\link{is_even}()},
\code{\link{is_increasing}()},
\code{\link{is_integer}()},
\code{\link{is_na}()},
\code{\link{is_number}()},
\code{\link{is_odd}()},
\code{\link{is_permutation}()},
\code{\link{is_round}()},
\code{\link{is_unique}()}
}
\concept{check-performing functions provided by utilbox}
