% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripting.R
\name{is_sourced}
\alias{is_sourced}
\alias{source_depth}
\title{Sourcing check}
\usage{
is_sourced(level = 1, exact_level = TRUE)

source_depth()
}
\value{
A logical indicator of whether the sourcing level is equal or above \code{level}.
}
\description{
Check whether the script from which it is called is being sourced (via either
\code{source()} or \code{sys.source()}) "at" (when \code{exact_level=TRUE}) or 
"above" (when \code{exact_level=FALSE}) level \code{level}. In other words, when
this function is called from a script that is being sourced from the global environment
then \code{is_sourced(1)} returns \code{TRUE}, when it is called from a script that
is being sourced by another script from the global environment
then \code{is_sourced(2)} returns \code{TRUE}, etc.

When sourcing files, it might be of interest to know how many times the \code{source}
function has been called. This function returns that number.
}
