% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_script.R
\name{is_title_line}
\alias{is_title_line}
\title{\code{is_title_line()} identifies "title lines", which are those}
\usage{
is_title_line(x)
}
\description{
\code{is_title_line()} identifies "title lines", which are those
}
