% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numchecks.R
\name{is_unique}
\alias{is_unique}
\title{Check if a sequence in 'x' has only unique elements}
\usage{
is_unique(x)
}
\description{
Check if a sequence in 'x' has only unique elements
}
\seealso{
Other check-performing functions provided by utilbox: 
\code{\link{anyNonNA}()},
\code{\link{drop_positive_negative}},
\code{\link{is_all_same}()},
\code{\link{is_decreasing}()},
\code{\link{is_even}()},
\code{\link{is_increasing}()},
\code{\link{is_integer}()},
\code{\link{is_na}()},
\code{\link{is_number}()},
\code{\link{is_odd}()},
\code{\link{is_permutation}()},
\code{\link{is_round}()},
\code{\link{is_scientific}()}
}
\concept{check-performing functions provided by utilbox}
