% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lang_sub}
\alias{lang_sub}
\alias{gsub_lang}
\alias{sub_lang}
\title{Replacement in expressions}
\usage{
lang_sub(code, pattern, repl, fixed = TRUE, workhorse = gsub)

gsub_lang(...)

sub_lang(...)
}
\description{
Similarly to \code{gsub} and \code{sub}, the functions \code{gsub_lang} and \code{sub_lang}
perform replacement in expressions. The expressions are first converted
to strings, the replacement is executed (via \code{base::gsub} or \code{base::sub})
and the result is converted back to language expressions. The input can
be a list of expressions or a function.
}
\examples{
# let's modify the `get2` function to return an NA when the requested
# object is not found.
get3 = hijack(get2, default_value2=NA)
get3 = gsub(get3, 'default_value', 'default_value2')

}
