% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{last_element}
\alias{last_element}
\title{Get the last element}
\usage{
last_element(x)
}
\description{
Returns the last element in an vector or list.
}
\seealso{
Other sequence-related functions provided by utilbox: 
\code{\link{every_nth}()},
\code{\link{filter_by_position}},
\code{\link{filter_by}},
\code{\link{groups_of_unique}()},
\code{\link{h1}()},
\code{\link{insert}()},
\code{\link{is_between}()},
\code{\link{is_run}},
\code{\link{rep2}()},
\code{\link{replace2}()},
\code{\link{replace3}()},
\code{\link{seq2}()},
\code{\link{seq_around}()},
\code{\link{sizechecks}},
\code{\link{split_into_groups}()},
\code{\link{to_factor}()},
\code{\link{what_bag}()}
}
\concept{sequence-related functions provided by utilbox}
