% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{launchProgressBar}
\alias{launchProgressBar}
\title{Start progress bar}
\usage{
launchProgressBar(
  pb_max,
  pb_print_every_n = 1,
  pb_show = TRUE,
  pb_initial = 0,
  pb_width = min(options()$width, 100),
  pb_con = stdout(),
  envir = parent.frame()
)
}
\description{
\code{launchProgressBar()} creates a progress bar (object \code{..pb..} in the environment \code{envir})
and returns the updating function (invisibly). Besides returning it, it also puts the
updating function as \code{..pb_update..()} in \code{envir}.
}
