% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{lib_dependencies}
\alias{lib_dependencies}
\alias{install_package_dependencies}
\title{Package dependencies}
\usage{
lib_dependencies(pckg, ...)

install_package_dependencies(pckg, ..., skip_installed = TRUE, quiet = FALSE)
}
\description{
\code{lib_dependencies} lists dependencies for all packages supplied in \code{pckg}.
It is nothing but a wrapper for \code{tools::package_dependencies}.
}
\details{
\code{install_package_dependencies} (re)installs all dependencies of the
supplied packages. Useful when an error occurred during a previous
package installation and some packages were not installed.
}
\seealso{
Other package-related functions provided by utilbox: 
\code{\link{apply_pckg}()},
\code{\link{auto_install_packages}},
\code{\link{list_installed_packages}()},
\code{\link{llibrary}()},
\code{\link{set_pkglib}()},
\code{\link{which_package}()}
}
\concept{package-related functions provided by utilbox}
