% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_clean}
\alias{list_clean}
\title{Removes zero-length elements from a list}
\usage{
list_clean(L, null.rm = TRUE)
}
\description{
Removes zero-length elements from a list
}
\examples{
list_clean(list(1:10, NULL, 'a'))

}
\seealso{
Other list utilities provided by utilbox: 
\code{\link{list_flatten}()},
\code{\link{list_mirror}()},
\code{\link{list_set_attr}()},
\code{\link{list_update}()},
\code{\link{merge_recurse}()},
\code{\link{nlist}()}
}
\concept{list utilities provided by utilbox}
