% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_empty}
\alias{list_empty}
\title{Flattens a list}
\usage{
list_empty(x)
}
\description{
An alias for \code{as.list(unlist(...))}.
}
\seealso{
Other list utilities provided by utilbox: 
\code{\link{is_all_character}()},
\code{\link{list_clean}()},
\code{\link{list_flatten}()},
\code{\link{list_mirror}()},
\code{\link{list_set_attr}()},
\code{\link{list_update}()},
\code{\link{merge_recurse}()},
\code{\link{nlist}()}
}
\concept{list utilities provided by utilbox}
