% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{list_files}
\alias{list_files}
\alias{list_files_latest}
\alias{list_files_biggest}
\title{List files matching regular pattern}
\usage{
list_files(
  pattern = NULL,
  path = ".",
  full.names = FALSE,
  ...,
  sort = FALSE,
  attrib = c("mtime", "size", "isdir", "mode", "ctime", "atime", "exe"),
  decreasing = FALSE,
  append_path = FALSE
)

list_files_latest(...)

list_files_biggest(...)
}
\description{
A wrapper for [base::list.files] that allows multiple patterns at once and 
files that match at least on of the patterns are returned. Optionally,
files can be sorted according to attribute in \code{attrib}.
}
\details{
Wrappers \code{list_files_latest} and \code{list_files_biggest} implement the
sorting according to attributes "modification time" and "size", respectively.
}
\seealso{
Other file system functions provided by utilbox: 
\code{\link{dir_exist_check}()},
\code{\link{dir_work}},
\code{\link{file_exists}()},
\code{\link{file_work}}
}
\concept{file system functions provided by utilbox}
