% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{list_installed_packages}
\alias{list_installed_packages}
\alias{list_loaded_packages}
\alias{list_attached_packages}
\title{Listing of installed and/or loaded packages}
\usage{
list_installed_packages()

list_loaded_packages(base = TRUE, other = TRUE)

list_attached_packages(
  only_name = TRUE,
  list_ignore = c(".GlobalEnv", "Autoloads")
)
}
\description{
\code{list_installed_packages()} lists all installed packages.

\code{list_loaded_packages()} and \code{list_attached_packages()}
both list all loaded packages. The former using
\code{utils::sessionInfo()}, the latter based on \code{base::search()}.
}
\examples{
list_installed_packages()
list_loaded_packages()
list_attached_packages()

}
\seealso{
Other package-related functions provided by utilbox: 
\code{\link{apply_pckg}()},
\code{\link{lib_dependencies}()},
\code{\link{llibrary}()},
\code{\link{set_pkglib}()}
}
\concept{package-related functions provided by utilbox}
