% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_mirror}
\alias{list_mirror}
\title{Mirror list structure}
\usage{
list_mirror(x, value, depth = 1)
}
\description{
Mirrors the structure of the list \code{x}. In other words, it returns
a list with the same numbers of elements at each list-depth. For now
only mirroring up to depth 1 is implemented.
}
\seealso{
Other list utilities provided by utilbox: 
\code{\link{is_all_character}()},
\code{\link{list_clean}()},
\code{\link{list_empty}()},
\code{\link{list_flatten}()},
\code{\link{list_set_attr}()},
\code{\link{list_update}()},
\code{\link{merge_recurse}()},
\code{\link{nlist}()}
}
\concept{list utilities provided by utilbox}
