% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripting.R
\name{list_objects_created_by_sourcing}
\alias{list_objects_created_by_sourcing}
\title{Get the list of objects created by sourcing a file}
\usage{
list_objects_created_by_sourcing(
  file,
  ...,
  all.names = TRUE,
  format_fun = as_object_table
)
}
\description{
Find out what objects would be created if a file was sourced (via
\code{\link[base:sys.source]{base::sys.source}}).
}
\examples{
# create a toy script file
tmp_file = '.~temp.R'
cat('test_function = function() NULL\n', file=tmp_file)
# list the objects (as a table with details)
list_objects_created_by_sourcing(tmp_file)
# list the objects (just a vector of object names)
list_objects_created_by_sourcing(tmp_file, format_fun=function(x, ...) x)
# delete the toy file
unlink(tmp_file)

}
