% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{list_package}
\alias{list_package}
\alias{list_package_exported}
\alias{list_package_all}
\alias{package_table}
\title{List objects in a package}
\usage{
list_package_exported(pckg, character.only = FALSE)

list_package_all(pckg, character.only = FALSE, all.names = TRUE)

package_table(objs, pckg)
}
\description{
Lists all/exported objects in a package.
}
\details{
\code{list_package_exported} lists all exported objects by a package.

\code{list_package_all} lists all objects defined in a package (including
non-exported objects).

`package_table` puts the objects into a table together
}
\examples{
list_package_exported(utilbox)
list_package_all(utilbox)

# lsf.str("package:dplyr")   # see also
# ls("package:dplyr")        # see also (works when the package is loaded)
help(package = dplyr)
# see https://stackoverflow.com/questions/30392542/is-there-a-command-in-r-to-view-all-the-functions-present-in-a-package

}
\seealso{
Other package-related functions provided by utilbox: 
\code{\link{R_add_lib_startup}()},
\code{\link{R_default_Rprofile_file}()},
\code{\link{R_user_home}()},
\code{\link{apply_pckg}()},
\code{\link{llibrary}()},
\code{\link{set_pkglib}()}
}
\concept{package-related functions provided by utilbox}
