% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_set_attr}
\alias{list_set_attr}
\title{Set list attributes (element-wise)}
\usage{
list_set_attr(x, attrib_name, attrib)
}
\description{
\code{list_set_attr} takes a list (\code{x}), attribute name (\code{attrib_name})
and a list of attributes (\code{attrib}) and sets the attribute named
\code{attrib_name} of each element of \code{x} to the corresponding value in
code{attrib}.
}
\examples{
set_list_attr(list('a','b'), 'size', c(10,20))

}
\seealso{
Other list utilities provided by utilbox: 
\code{\link{is_all_character}()},
\code{\link{is_list_class}()},
\code{\link{list_clean}()},
\code{\link{list_empty}()},
\code{\link{list_flatten}()},
\code{\link{list_mirror}()},
\code{\link{list_update}()},
\code{\link{merge_recurse}()},
\code{\link{nlist}()}
}
\concept{list utilities provided by utilbox}
