% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_update}
\alias{list_update}
\alias{modifyList2}
\title{Modify list}
\usage{
list_update(x, val, ..., drop_null_val = TRUE)

modifyList2(x, val, ..., drop_null_val = TRUE)
}
\description{
A version of 'modifyList' from `utils` which drops zero-length elements in 'val'
before updating 'x' (optionally can behave the same as modifyList)
}
\seealso{
Other list utilities provided by utilbox: 
\code{\link{list_clean}()},
\code{\link{list_flatten}()},
\code{\link{list_mirror}()},
\code{\link{list_set_attr}()},
\code{\link{merge_recurse}()},
\code{\link{nlist}()}
}
\concept{list utilities provided by utilbox}
