% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipping.R
\name{list_zip}
\alias{list_zip}
\title{Lists files in a zip archive}
\usage{
list_zip(zipfiles, pattern = ".*", exclude = FALSE)
}
\arguments{
\item{pattern}{Regular expression pattern(s) against which the
discovered file names are matched}

\item{exclude}{If \code{TRUE}, the pattern is used to exclude the
discovered files, otherwise files that match \code{pattern} are kept.
Defaults to \code{FALSE}}

\item{zip}{Name of the zip archive(s) to read files from}
}
\value{
A character vector or list depending on whether the input
in \code{zipfiles} was of type \code{list}.
}
\description{
Lists files within a zip archive that match the supplied regular
expression pattern or patterns. Keeps only files that match at least
one of the patterns in \code{pattern} or those that do not match any of
the patterns when \code{exclude} is \code{TRUE}.
}
\seealso{
Other compression-related utilities provided by utilbox: 
\code{\link{read_zip}()},
\code{\link{unzip_files}()},
\code{\link{zip_all_in_path}()},
\code{\link{zip_files_pattern}()},
\code{\link{zip_files}()},
\code{\link{zip_file}()}
}
\concept{compression-related utilities provided by utilbox}
