% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pause.R
\name{listen_for_pause}
\alias{listen_for_pause}
\title{Listen for signals to pause execution}
\usage{
listen_for_pause(
  pid = Sys.getpid(),
  file = file.path("~", ".Rgui_pause.txt"),
  pause_fun_name = "browser",
  envir = parent.frame(),
  trace = FALSE
)
}
\description{
When called, \code{listen_for_pause()} looks for the existence of the specified
file (.Rgui_pause.txt in the user's home directory by default) and checks
the contents of the file for the presence of the current process ID. If the
process ID is found, the specified pause function (via \code{pause_fun_name},
set to 'browser' by default) is called in the specified environment (\code{envir}).
Otherwise, execution continues as normal. The return value indicates whether
the process ID was found (1), not found (0), or the file does not exist (-1).
}
