% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{llibrary}
\alias{llibrary}
\alias{llib}
\alias{unload_library}
\alias{unload_all_libraries}
\alias{reload_unloaded_libraries}
\alias{package_is_installed}
\alias{is_lib_installed}
\title{Library loading and unloading operations}
\usage{
llibrary(
  pckgs = NULL,
  quietly = TRUE,
  character.only = FALSE,
  fail = warn,
  detach_first = FALSE,
  remove_first = FALSE,
  default_src = "CRAN",
  lib_fun = base::library,
  url_CRAN = "https://cloud.r-project.org/",
  suppress_startup_msgs = FALSE,
  ...
)

llib(
  ...,
  detach_first = FALSE,
  remove_first = FALSE,
  suppress_startup_msgs = FALSE
)

unload_library(pckgs, announce = TRUE)

unload_library(pckgs, announce = TRUE)

unload_all_libraries()

reload_unloaded_libraries(pckgs)

package_is_installed(pckgs, character.only = FALSE)

is_lib_installed(pckgs, character.only = FALSE)
}
\description{
\code{llibrary()} loads specified libraries. On input, \code{pckgs} must
either be a character vector of package names or a list with each of
its elements having the structure of \code{list(name=..., src=...)}
where name is the package name and src is the package source
(repository or local) from which the package is to be loaded.

\code{llib()} is an alias for \code{llibrary()} which allows unquoted input.

\code{unload_library()} unloads/detaches a loaded library.

\code{package_is_installed()} checks if a package is installed.

\code{is_lib_installed()} is an alias for \code{package_is_installed()}.

\code{unload_all_libraries()} unloads all non-base libraries in the
search path (i.e. those returned by \code{base::search()}).

\code{reload_unloaded_libraries()} attempts to reverse the action by
\code{unload_all_libraries()}.
}
\examples{
package_is_installed('utilbox')
package_is_installed(utilbox)
lib = 'utilbox'; package_is_installed(lib)
lib = 'utilbox'; package_is_installed(lib, character.only=TRUE)

}
\seealso{
Other package-related functions provided by utilbox: 
\code{\link{apply_pckg}()},
\code{\link{auto_install_packages}},
\code{\link{lib_dependencies}()},
\code{\link{list_installed_packages}()},
\code{\link{set_pkglib}()},
\code{\link{which_package}()}
}
\concept{package-related functions provided by utilbox}
