% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{loadAs}
\alias{loadAs}
\title{Load objects from file}
\usage{
loadAs(file, as, what, envir = parent.frame())
}
\description{
Loads objects from a file and assigns the contents into variables listed in
\code{as} inside the environment in \code{envir} (the parent frame of the call
to \code{loadAs} by default).
}
