% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{locate_position_of_value}
\alias{locate_position_of_value}
\alias{w_first_nonzero}
\alias{w_kth_nonzero}
\alias{w_last_nonzero}
\alias{w_first_positive}
\alias{w_last_positive}
\alias{w_first_above}
\alias{w_first_above_soft}
\alias{w_last_above}
\alias{w_last_above_soft}
\alias{w_first_below}
\alias{w_first_below_soft}
\alias{w_last_below}
\alias{w_last_below_soft}
\title{Locate position of a value}
\usage{
w_first_nonzero(x)

w_kth_nonzero(x, n)

w_last_nonzero(x)

w_first_positive(x)

w_last_positive(x)

w_first_above(x, y)

w_first_above_soft(x, y)

w_last_above(x, y)

w_last_above_soft(x, y)

w_first_below(x, y)

w_first_below_soft(x, y)

w_last_below(x, y)

w_last_below_soft(x, y)
}
\description{
\code{w_first_nonzero} finds the position of the last non-zero element
in \code{x}.

\code{w_kth_nonzero} finds the position of the n-th non-zero element in
\code{x}.

\code{w_last_nonzero} finds the position of the last non-zero element
in \code{x}.

\code{w_first_positive} finds the position of the first positive
element in \code{x}.

\code{w_last_positive} finds the position of the last positive element
in \code{x}.

\code{w_last_above} finds the position of the last element in \code{x} with
value larger than \code{y}.

\code{w_first_above} finds the position of the first element in \code{x}
with value larger than \code{y}.

\code{w_last_below} finds the position of the last element in \code{x} with
value below \code{y}.

\code{w_first_below} finds the position of the first element in \code{x}
with value below \code{y}.

\code{w_first_nonempty} finds the position of the first non-empty
element in \code{x}. It is designed as a method, which currenlty only
operates on lists.
}
\seealso{
Other numeric functions provided by utilbox: 
\code{\link{Invert}()},
\code{\link{downweight}()},
\code{\link{extract_by_value}},
\code{\link{lower_tri}()},
\code{\link{min_dif}()},
\code{\link{permutations}},
\code{\link{replace_value}},
\code{\link{sequences}},
\code{\link{sets}},
\code{\link{to_base_k}()},
\code{\link{which_max}()}
}
\concept{numeric functions provided by utilbox}
