% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{logit}
\alias{logit}
\title{Logit, probit, inverse logit functions}
\usage{
logit(p)
}
\description{
\code{logit()} gives the logit function, which is also the quantile
function of the logistic distribution. More exactly, it is the
function \emph{f(p)=log(p/(1-p))}.

\code{probit()} is the probit function, which is also the quantile
function of the standard normal distribution.

\code{ilogit()} is the inverse logit function, i.e. the distribution
function of the logistic distribution.
}
