% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{lower_tri}
\alias{lower_tri}
\alias{upper_tri}
\title{Upper- and lower-triangular}
\usage{
lower_tri(x, diag = TRUE)

upper_tri(x, diag = TRUE)
}
\description{
\code{lower_tri()} and \code{upper_tri()} return the elements on the upper-
and lower-triangular as a vectors.
}
\seealso{
Other numeric functions provided by utilbox: 
\code{\link{Invert}()},
\code{\link{downweight}()},
\code{\link{extract_by_value}},
\code{\link{locate_position_of_value}},
\code{\link{min_dif}()},
\code{\link{permutations}},
\code{\link{replace_value}},
\code{\link{sequences}},
\code{\link{sets}},
\code{\link{to_base_k}()},
\code{\link{which_max}()}
}
\concept{numeric functions provided by utilbox}
