% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{ls_parents}
\alias{ls_parents}
\alias{get_in_parents}
\title{List and get objects in parent frames}
\usage{
ls_parents(..., stop_at_global = FALSE, envir = parent.frame())

get_in_parents(x, envir = parent.frame())
}
\description{
\code{ls_parents()} returns a list of all objects in the sequence of parent
frames.
}
\details{
\code{get_in_parents()} retrieves an object by the specified name from all
parent frams where it is found.
}
