% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drives.R
\name{map_drives}
\alias{map_drives}
\alias{map_drive}
\alias{unmap_drive}
\alias{list_all_drives}
\alias{list_all_network_drives}
\alias{list_all_network_drives2}
\alias{convert_to_network_path}
\title{Work with drives on Windows}
\usage{
map_drive(
  drive = find_unmapped_drive(),
  path = ".",
  persistent = FALSE,
  delete_existing = FALSE,
  force_delete = FALSE,
  change_wdir = FALSE,
  normalize = FALSE,
  silent = FALSE
)

unmap_drive(drive, force = FALSE, silent = FALSE)

list_all_drives(append = "")

list_all_network_drives(append = "")

list_all_network_drives2(path)

convert_to_network_path(path, normalize = FALSE)
}
\description{
\code{map_drive()} maps a path to the specified drive.

\code{unmap_drive()} unmaps the specified drive.

\code{convert_to_network_path()} converts a path from the regular
regular format (e.g., 'c:\\Windows') to the network path format
(e.g., '\\\\localhost\\c$\\Windows'), which can be used to map
any path to a drive.

\code{list_all_drives()} returns a list of all existing drives
together with their mapped location in the network path format.
If supplied, a string in \code{append} is attached to the end of the
network path(s) (e.g., '\\.' can be useful).

\code{list_all_network_drives()} lists drives that are mappings
of network-storage devices (which can include local paths
mapped as drives by relying on the network format of paths
(e.g., '\\\\localhost\\c$\\Windows').

All of these functions work on Windows only and rely on the
Windows system utilities such as \code{net} and/or \code{wmic} (which
has been deprecated in Windows 10, version 21H1, see here
https://docs.microsoft.com/en-us/windows/win32/wmisdk/wmic)
}
\seealso{
Other file system function provided by utilbox: 
\code{\link{check_file_locked}()},
\code{\link{clean_filename}()},
\code{\link{file_backup}()},
\code{\link{file_can_open_check}()},
\code{\link{file_empty}()},
\code{\link{file_size}()},
\code{\link{file_sort}()},
\code{\link{random_filename}()},
\code{\link{separate_path}()}
}
\concept{file system function provided by utilbox}
