% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{matchecks}
\alias{matchecks}
\alias{is_diag}
\alias{is_posdef}
\title{Matrix checks}
\usage{
is_diag(x)

is_posdef(x)
}
\description{
\code{is_diag()} checks if 'x' is a diagonal matrix.

\code{is_posdef()} checks if \code{M} is a positive-definite matrix.
}
