% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{mb}
\alias{mb}
\title{Benchmarking}
\usage{
mb(
  ...,
  list = NULL,
  times = 100L,
  unit = NULL,
  check = NULL,
  control = list(),
  envir = base::parent.frame()
)
}
\description{
Simply a short-named alias for \code{microbenchmark::microbenchmark}.
}
