% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/means.R
\name{means}
\alias{means}
\alias{mean_na}
\alias{amean}
\alias{hmean}
\alias{gmean}
\alias{gmeanw}
\alias{lmean}
\title{Various means}
\usage{
mean_na(..., na.rm = TRUE)

amean(x, na.rm = TRUE)

hmean(..., na.rm = TRUE)

gmean(x, na.rm = TRUE)

gmeanw(x, w, na.rm = TRUE, ...)

lmean(x, w, na.rm = TRUE, ...)
}
\description{
Various mean functions. Arithmetic, geometric, weighted geometric,
harmonic and logistic means are implemented. Compares to [base::mean]
these functions have remove `NA`s by default and return zero-length
responses on zero-length input or input with no non-`NA` values.
}
\details{
`amean` gives the arithmetic mean of `x`. 

`gmean` and `gmeanw` calculate the geometric mean of `x`, while the latter
permits weighing.

`hmean` calculates the harmonic mean (single value) of a single vector created as a
combination of all its input arguments.

`lmean` calculates the (weighted) logistic mean.

`mean_na` is an alias for [base::mean] with `na.rm=TRUE`.
}
\examples{
x = c(1,NA); mean(x); mean2(x)

}
