% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{merge_recurse}
\alias{merge_recurse}
\title{Recursive merge}
\usage{
merge_recurse(dfs, ...)
}
\description{
This does a recursive merge of lists. The code is taken from the package 
reshape and modified to fix the bug in it. The bug was the missing ellipsis 
in the call to \code{Recall()}.
}
\seealso{
Other list utilities provided by utilbox: 
\code{\link{list_clean}()},
\code{\link{list_flatten}()},
\code{\link{list_mirror}()},
\code{\link{list_set_attr}()},
\code{\link{list_update}()},
\code{\link{nlist}()}
}
\concept{list utilities provided by utilbox}
