% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{Mode}
\alias{Mode}
\title{Mode of a sample}
\usage{
Mode(x, all_modes = FALSE, exclude.na = TRUE)
}
\description{
Finds the mode, i.e. the value that occurs with highest frequency in \code{x}.
The function works for both numeric and character/factor data.
}
\examples{
Mode(c(1:7,7))    # the highest frequency value is '7'

}
