% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing.R
\name{msg}
\alias{msg}
\title{Prints a warning and either stops execution or waits depending on arguments}
\usage{
msg(
  t,
  ...,
  lead = "",
  sep = "",
  quit = FALSE,
  wait = FALSE,
  skip1 = 0,
  skip2 = 0,
  flush = TRUE
)
}
\description{
Prints a warning and either stops execution or waits depending on arguments
}
