% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtp.R
\name{mtc_test}
\alias{mtc_test}
\title{Multiple testing}
\usage{
mtc_test(p, alpha = 0.05, method = "holm")
}
\value{
Logical vector (or matrix) indicating non-significance
statuses of each p-value.
}
\description{
\code{mtc_test()} takes a vector of p-values and performs a multiple
testing corrected test on it. Which test it is is determined via
\code{method}.

\code{mtc_test_mul()} is a wrapper for it which can take multiple sets
of p-values (as columns of a matrix) and performs the multiple
testing corrected tests on each of the sets separately.
}
\examples{
mtc_test(runif(10))
mtc_test_mul(cbind(runif(10),runif(10)))

}
