% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{multi_sub}
\alias{multi_sub}
\alias{subm}
\alias{gsubm}
\title{Recursive string substitution}
\usage{
multi_sub(x, what, with, workhorse = base::sub, ...)

subm(x, what, with)

gsubm(x, what, with)
}
\description{
Allows for multiple arguments in \code{what} and \code{with} which are
replaced inside \code{where}. \code{what} and \code{with} are processed in parallel
which means that they must have the same length or \code{with} must be of
length 1.

\code{subm} leverages \code{base::sub}, while \code{gsubm} leverages
\code{base::gsub}.
}
\examples{
subm(c('a','b'), paste0('|',c('a','b'),'|'), 'abracadabra')
gsubm(c('a','b'), paste0('|',c('a','b'),'|'), 'abracadabra')

}
\seealso{
Other string-manipulation functions provided by utilbox: 
\code{\link{ascii}()},
\code{\link{collapse}},
\code{\link{cumpaste0}()},
\code{\link{padding}},
\code{\link{patternize}()},
\code{\link{str_cut}()},
\code{\link{str_diff}()},
\code{\link{str_grab}()},
\code{\link{str_insert}()},
\code{\link{str_is_empty}()},
\code{\link{str_paste_grid}()},
\code{\link{str_pos}()},
\code{\link{str_rev}()},
\code{\link{str_trim_space}()},
\code{\link{sub_at}()},
\code{\link{toupperfirst}()}
}
\concept{string-manipulation functions provided by utilbox}
