% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.R
\name{my_ast_tree}
\alias{my_ast_tree}
\title{This is a slightly modified version of the function \code{lobstr:::ast_tree()}.
It is used by the function \code{fun2funcalls()}, which required the modifications.}
\usage{
my_ast_tree(x, layout = box_chars())
}
\description{
This is a slightly modified version of the function \code{lobstr:::ast_tree()}.
It is used by the function \code{fun2funcalls()}, which required the modifications.
}
