% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{n_kth_rev}
\alias{n_kth_rev}
\title{title
Get elements for filtering in reverse order.}
\usage{
n_kth_rev(k, x)
}
\description{
For logical input in the first argument (\code{k}), \code{n_kth_rev} simply reverses
the elements of the vector while ignoring its second argument (\code{x}).
For numeric input in \code{k} that is positive, it returns the indices of the k-th
elements from the end in \code{x}. For negative numeric input in \code{k}, it returns
the same as for \code{k=abs(k)} except with a negative sign. For values 0 in \code{k}
it returns \code{NA}.
}
\examples{
n_kth_rev(c(TRUE,FALSE,TRUE))
n_kth_rev(c(1:3,10),letters)
n_kth_rev(-5,letters)

}
