% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting.R
\name{name_sort}
\alias{name_sort}
\title{Name sort}
\usage{
name_sort(x, get_order = FALSE)
}
\arguments{
\item{x}{named object to be sorted. If names are missing, no sorting is done.}

\item{get_order}{Determines whether the ordering of names or the ordered object
is returned.}
}
\description{
Sorts the input according to its names attribute.
}
