% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{namespace_exists}
\alias{namespace_exists}
\title{Check for namespace}
\usage{
namespace_exists(ns)
}
\description{
Checks if the given string corresponds to a namespace by
attempting to call \code{\link[base:asNamespace]{base::asNamespace}} on it and checking whether
an error occurred.
}
