% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{nformals}
\alias{nformals}
\title{Number of arguments a function accepts}
\usage{
nformals(fun, envir = parent.frame())
}
\description{
Returns the number of arguments that a given function \code{fun}
accepts. If the arguments of \code{fun} contain an ellipse (\code{...}), it
returns \code{Inf}.
}
\examples{
nformals(mode)
nformals(mean)


}
