% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{nlist}
\alias{nlist}
\alias{nlist2}
\title{Named list}
\usage{
nlist(...)

nlist2(...)
}
\description{
`nlist` creates a named list. The code `list(a = a, b = b)` 
becomes `nlist(a,b)` and `list(a = a, b = 2)` becomes 
`nlist(a, b = 2)`, etc.
}
\details{
`nlist2` is a shorter original code which relies on existing
code (`dots_to_nlist`). It should result in the exact
same behavior as `nlist`.

Credit: The code for `nlist` was lifted from the package 
`loo` to avoid having to install that package just for this 
one function. The code has been slightly modified.
}
\examples{
# uses the variable names to assign names
a = 1:10
nlist(a, b=2)
nlist2(a, b=2)

# uses the call to assign names
nlist(1:10, sin(30))
nlist2(1:10, sin(30))

}
\seealso{
Other list utilities provided by utilbox: 
\code{\link{list_clean}()},
\code{\link{list_flatten}()},
\code{\link{list_set_attr}()},
\code{\link{merge_recurse}()},
\code{\link{modifyList2}()}
}
\concept{list utilities provided by utilbox}
