% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{normpval}
\alias{normpval}
\title{Normal distribution p-value}
\usage{
normpval(x, two.sided = FALSE, m = 0, s = 1)
}
\description{
Calculates the p-value relative to the normal distribution with parameters 
\code{m} (mean) and \code{s} (standard deviation).
}
\examples{
normpval(1.96)        # approx. 2.5\%
normpval(1.96, TRUE)  # approx. 5.0\%
}
