% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{not_in}
\alias{not_in}
\alias{\%notin\%}
\alias{\%nin\%}
\alias{\%has\%}
\title{Not-in-set operator}
\usage{
... \%notin\% NA

... \%nin\% NA

table \%has\% x
}
\description{
Logical negation of the value matching operator \code{\%in\%}. The
operator \code{\%notin\%} (just like its alias \code{\%notin\%})
returns \code{TRUE} for all elements of the first argument that are not
contained among the elements of the second argument, and \code{FALSE}
otherwise.

\code{\%has\%} is equivalent to \code{\%in\%} except with the roles
of the two sides switched.
}
\examples{
1 \%notin\% 1:3
0 \%notin\% 1:3

}
\seealso{
Other operators provided by utilbox: 
\code{\link{\%anyin\%}()},
\code{\link{\%modify\%}()},
\code{\link{operator_NULL}},
\code{\link{operator_append}},
\code{\link{operator_c}},
\code{\link{operator_match}},
\code{\link{operator_try}},
\code{\link{renaming_operators}}
}
\concept{operators provided by utilbox}
