% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notepad.R
\name{notepad_highlighting}
\alias{notepad_highlighting}
\title{Add all utilbox objects into Notepad++ highlighting}
\usage{
notepad_highlighting(
  keywords = NULL,
  field = "type1",
  npp_path,
  npp_langs,
  backup = TRUE
)
}
\description{
Modifies the notepad language keyword groups in order
for the utilbox objects to be highlighted by Notepad++.
}
