% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notepad.R
\name{notepad_keywords}
\alias{notepad_keywords}
\title{Extract the keywords for packages and package groups}
\usage{
notepad_keywords(field, do_methods = FALSE)
}
\description{
Only Keywords1 (base words), Keywords2 (methods), Keywords3 (R base),
Keywords4 (extended base R words), Keywords6 (tidyverse) and
Keywords 8 (utilbox) are currently returned.
}
