% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numchecks.R
\name{numchecks}
\alias{numchecks}
\alias{is_na}
\alias{n_na}
\alias{is_unique}
\alias{is_all_same}
\alias{is_decreasing}
\alias{is_increasing}
\alias{equals_integer}
\alias{is_odd}
\alias{is_even}
\title{Check}
\usage{
is_na(x, freduce = all)

n_na(x)

is_unique(x)

is_all_same(x)

is_decreasing(x, strictly = FALSE)

is_increasing(x, strictly = FALSE)

equals_integer(x, tol = .Machine$double.eps)

is_odd(x, tol = .Machine$double.eps)

is_even(x, tol = .Machine$double.eps)
}
\description{
\code{is_na(x)} checks if a reduction of elements of \code{x} is \code{NA}. The
default reduction function is \code{all}, which returns \code{TRUE} whenever
all elements of its argument are \code{NA} or \code{x} is 0-length. A useful
alternative reduction function is \code{any} supplied via the argument
\code{freduce}.

\code{n_na(x)} counts the number of NA elements in \code{x}.

\code{is_unique(x)} checks whether a sequence in 'x' has only unique
elements.

\code{is_all_same(x)} checks if all elements in 'x' have the same value.

\code{is_decreasing(x)} checks if a sequence in 'x' is decreasing.

\code{is_increasing(x)} checks if a sequence in 'x' is increasing.

\code{equals_integer(x)} checks if elements in 'x' are effectively integers
(i.e. close enough to one).

\code{is_odd(x)} checks if elements of 'x' are odd

\code{is_even(x)} checks if elements of 'x' are even
}
\seealso{
Other check-performing functions provided by utilbox: 
\code{\link{anyNonNA}()},
\code{\link{extract_by_positiveness_negativeness}},
\code{\link{is_number}()},
\code{\link{is_same_value}()},
\code{\link{rounding}}
}
\concept{check-performing functions provided by utilbox}
