% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{operator_append}
\alias{operator_append}
\alias{\%append\%}
\alias{\%appendR\%}
\alias{\%ap\%}
\alias{\%apR\%}
\title{Append operator}
\usage{
left \%append\% right

right \%appendR\% left

left \%ap\% right

right \%apR\% left
}
\value{
A list.
}
\description{
Append one object to another via \code{\link[base:append]{base::append}}. \verb{\\\%append\\\%}
appends the right-hand argument to the left-hand side, while
\verb{\\\%appendR\\\%} appends in the reverse order.
}
\examples{
list(a=1, b=2) \%append\% list(c=3)    # adds an element to the list on the lhs
c(1,2) \%append\% list(c=3)            # turns the lhs into a list, then appends

}
\seealso{
Other operators provided by utilbox: 
\code{\link{\%anyin\%}()},
\code{\link{\%modify\%}()},
\code{\link{not_in}},
\code{\link{operator_NULL}},
\code{\link{operator_c}},
\code{\link{operator_match}},
\code{\link{operator_try}},
\code{\link{renaming_operators}}
}
\concept{operators provided by utilbox}
