% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{operator_c}
\alias{operator_c}
\alias{\%c\%}
\alias{\%cR\%}
\title{Append operator}
\usage{
x \%c\% y

y \%cR\% x
}
\description{
Append one object to another via \code{\link[base:c]{base::c}}.
}
\examples{
1:10 \%c\% 11:15

}
\seealso{
Other operators provided by utilbox: 
\code{\link{\%anyin\%}()},
\code{\link{\%modify\%}()},
\code{\link{concatenate}},
\code{\link{not_in}},
\code{\link{operator_NULL}},
\code{\link{operator_append}},
\code{\link{operator_match}},
\code{\link{renaming_operators}}
}
\concept{operators provided by utilbox}
