% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{operator_NULL}
\alias{operator_NULL}
\alias{\%|||\%}
\alias{\%||||\%}
\alias{\%|||||\%}
\alias{\%NA\%}
\title{Default value for NULL and zero-length objects}
\usage{
x \%|||\% y

x \%||||\% y

x \%|||||\% y

x \%NA\% y
}
\description{
`\%|||\%` is intented to be equivalent "\code{\%||\%}" operator in the 
package \code{rlang}, but named differently as not to clash with it.
}
\details{
`\%||||\%` extends this behavior to all zero-length objects.

`\%|||||\%` extends this behavior further to include empty strings.

`\%NA\%` is an operator version of `ifelse(is.na(x), y, x), which
allows input of various types (vectors, lists, closures, etc.).
}
\examples{
1 \%|||\% 2
NULL \%|||\% 2
NULL \%||||\% 2
numeric(0) \%||||\% 2

}
\seealso{
Other operators provided by utilbox: 
\code{\link{\%anyin\%}()},
\code{\link{\%append\%}()},
\code{\link{\%c\%}()},
\code{\link{\%modify\%}()},
\code{\link{concatenate}},
\code{\link{not_in}},
\code{\link{operator_match}},
\code{\link{renaming_operators}}
}
\concept{operators provided by utilbox}
