% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{operator_try}
\alias{operator_try}
\alias{\%try\%}
\alias{\%otherwise\%}
\alias{\%finally\%}
\alias{\%recover\%}
\title{Operators for trying calls and recovering from failures}
\usage{
. \%try\% ...

e \%otherwise\% ...

x \%finally\% ...

lhs \%recover\% ...
}
\description{
\verb{\%try\%} attempts to execute a block of code on the right-hand
side (RHS) using \code{base::tryCatch}. The left-hand side (LHS)
is ignored.

\verb{\%otherwise\%} looks for an error result on the LHS and if one
is found, it executes the RHS code.

\verb{\%finally\%} executes the RHS code and returns the outcome whenever
the LHS was an error, otherwise it returns the LHS.

After the execution of \%otherwise\% or \%finally\% the latest error
is found in \code{.e} in the parent frame.

\verb{\%recover\%} attempts to execute the LHS and on failure it executes
the RHS. It is similar to \verb{\%ERR\%}, which works a little differently.
}
\examples{
. \%try\% print(fsdafsdaf) \%otherwise\% message('the printing did not work')
print(fsdafsdaf) \%recover\% message('the printing did not work')

}
\seealso{
Other operators provided by utilbox: 
\code{\link{\%anyin\%}()},
\code{\link{\%modify\%}()},
\code{\link{not_in}},
\code{\link{operator_NULL}},
\code{\link{operator_append}},
\code{\link{operator_c}},
\code{\link{operator_match}},
\code{\link{renaming_operators}}
}
\concept{operators provided by utilbox}
