% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqchecks.R
\name{or_between}
\alias{or_between}
\title{Between with many intervals}
\usage{
or_between(x, intervals, sharp_left = FALSE, sharp_right = FALSE)
}
\description{
Checks whether \code{x} is in at least one of the supplied 
intervals.
}
\examples{
or_between(runif(3), list(c(0,0.1), c(0.9,1)))

}
\seealso{
Other sequence-related functions provided by utilbox: 
\code{\link{as_factor}()},
\code{\link{dim_check}},
\code{\link{every_nth}()},
\code{\link{filter_by_position}},
\code{\link{filter_by}},
\code{\link{groups_of_unique}()},
\code{\link{h1}()},
\code{\link{insert}()},
\code{\link{is_between}()},
\code{\link{is_empty}()},
\code{\link{is_run}},
\code{\link{last_element}()},
\code{\link{midpoints}()},
\code{\link{nunique}()},
\code{\link{rep2}()},
\code{\link{replace2}()},
\code{\link{replace3}()},
\code{\link{seq2}()},
\code{\link{seq_around}()},
\code{\link{setdiffsym}()},
\code{\link{shift}()},
\code{\link{split_into_groups}()},
\code{\link{unname_factor}()},
\code{\link{what_bag}()}
}
\concept{sequence-related functions provided by utilbox}
