% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{orig_env}
\alias{orig_env}
\title{Get the original environment of a function or another object}
\usage{
orig_env(fun)
}
\description{
When an object defined in a namespace is printed, its namespace is
shown as the last line. If a function in a package has been defined
as an alias of a function from another package, this original
information is still present when the function is printed For
instance, \code{utilbox::head} is just an alias for \code{utils::head}, as
shown when \code{print(utilbox::head)} is called. The function
\code{orig_env()} captures that information and returns it as a character
string. For objects with no such information printed it simply
returns the last element of a print call of that object.
}
\examples{
orig_env(mean)

}
