% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{p16}
\alias{p16}
\alias{pt16}
\title{Plot and scatterplot

\code{p16} and \code{pt16} are synonyms for \code{graphics::plot} and \code{graphics::points}
with different defaults for size (\code{cex}), character (\code{pch}) and color (\code{col}).}
\usage{
p16(..., col = "royalblue3", cex = 0.6, pch = 16)

pt16(..., col = "royalblue3", cex = 0.6, pch = 16)
}
\description{
Plot and scatterplot

\code{p16} and \code{pt16} are synonyms for \code{graphics::plot} and \code{graphics::points}
with different defaults for size (\code{cex}), character (\code{pch}) and color (\code{col}).
}
\examples{
p16(rnorm(100))
pt16(rnorm(100), col='deeppink1')

}
\seealso{
Other plotting-related functions provided by utilbox: 
\code{\link{Colors}()},
\code{\link{add_box_above}()},
\code{\link{add_tight_axes}()},
\code{\link{all_devs}},
\code{\link{as_color}()},
\code{\link{get_legend}()},
\code{\link{hist2}()},
\code{\link{palette_ub}()},
\code{\link{show_Colors}()}
}
\concept{plotting-related functions provided by utilbox}
