% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting.R
\name{pattern_sort}
\alias{pattern_sort}
\title{Pattern sort}
\usage{
pattern_sort(
  x,
  pattern,
  decreasing = FALSE,
  sort_on_names = TRUE,
  exclude = FALSE,
  get_order = FALSE
)
}
\description{
Sorts input according to the order given in pattern(s) (given in 
\code{pattern}) by matching the names of the input (i.e. \code{x}) 
against the pattern(s).
}
\seealso{
Other sorting utilities provided by utilbox: 
\code{\link{sort_pval_matrix}()}
}
\concept{sorting utilities provided by utilbox}
