% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pause.R
\name{pause}
\alias{pause}
\alias{sleep}
\alias{wait}
\alias{ask}
\title{Pause execution / wait}
\usage{
sleep(time, announce = TRUE)

wait(...)

ask(
  question,
  answers = c("Y", "N"),
  affirmative = "Y",
  nmax = 1,
  case_sensitive = FALSE
)
}
\description{
\code{wait()} pauses the execution until either the ENTER key (for
continue) or ESC (for quit) are pressed.

\code{sleep()} pauses execution for \code{time} seconds with (optional)
announcement.

\code{ask()} pauses the execution until and waits for the user to
respond with one of the questions. It then evaluates whether the
response was in the affirmative (as specified via \code{affirmative}).
Only the first nmax characters of the response are compared with
}
\examples{
sleep(3)

wait("It's nice weather outside.")

wait("It's nice weather outside.")

}
\seealso{
Other system-related functions provided by utilbox: 
\code{\link{R_defaults}},
\code{\link{R_gui_adjust}()},
\code{\link{R_session_info}()},
\code{\link{environment_variables}},
\code{\link{is_win}()},
\code{\link{stopping_control}},
\code{\link{system_path}},
\code{\link{timest}()}
}
\concept{system-related functions provided by utilbox}
