% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{pchisqw}
\alias{pchisqw}
\title{Get the probability of a weighted chi-square sum}
\usage{
pchisqw(x, weights, lower.tail = TRUE, epsw = 1e-05, df0 = 2)
}
\description{
Calculates the probability of a weighted sum of chisquare-df0 variables
being below (lower.tail==TRUE) or above (lower.tail=FALSE) x, where
weights specifies the weights. Fails if some (but not all) weights are equal
}
