% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutate.R
\name{permutations}
\alias{permutations}
\alias{all_permutations}
\alias{is_permutation}
\title{Generates all permutations}
\usage{
all_permutations(x)

is_permutation(x, y)
}
\description{
\code{all_permutations()} generates all possible permutations of the
elements of its argument.

\code{is_permutation()} checks if the elements in the first argument
(\code{x}) correspond to a permutation of the second argument (\code{y}).
}
\seealso{
Other numeric functions provided by utilbox: 
\code{\link{Invert}()},
\code{\link{as_numeric}()},
\code{\link{downweight}()},
\code{\link{extract_by_value}},
\code{\link{force_as_}},
\code{\link{locate_position_of_value}},
\code{\link{lower_tri}()},
\code{\link{make_numeric}()},
\code{\link{min_dif}()},
\code{\link{replace_value}},
\code{\link{sequences}},
\code{\link{sets}},
\code{\link{to_base_k}()},
\code{\link{which_max}()}
}
\concept{numeric functions provided by utilbox}
