% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.R
\name{perturb}
\alias{perturb}
\title{Fuzzy up a numeric object}
\usage{
perturb(x, mindif = 0, preserve_sum = TRUE, maxrep = 100)
}
\description{
\code{perturb()} adds a small perturbation to those elements that very close
together (as defined by \code{mindif}). By default, it only perturbs elements
that are exactly equal (i.e., \code{mindif=0}). When \code{preserve_sum=TRUE}, the
perturbation is done in a way that preserves the sum of \code{x}.
}
\examples{
perturb(rep(1:10, e=3))

}
