% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{pmnormrect}
\alias{pmnormrect}
\title{Normal probability of a rectangle}
\usage{
pmnormrect(
  upper = c(Inf, Inf),
  lower = c(-Inf, -Inf),
  mean = rep(0, length(upper)),
  Sigma = diag(length(upper)),
  varcov
)
}
\description{
Calculates the probability of an arbitrary box with respect to an equi-correlated 
multivariate normal random vector of dimension given by the size of upper/lower
}
