% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing.R
\name{print2var}
\alias{print2var}
\title{Return the output of a print call as string}
\usage{
print2var(fun, file = NULL)
}
\description{
Returns the output of a print call as string. Useful for saving
the output of a print call into a variable. Simply an alias for
\code{\link[utils:capture.output]{utils::capture.output}}.
}
