% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic.R
\name{process_symbolic_call}
\alias{process_symbolic_call}
\title{Process symbolic call}
\usage{
process_symbolic_call(call, narg)
}
\description{
call (starting with \code{~}), which it turns into a function. Inspired
by \code{tidyverse}-style function supply.
}
