% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{qchisqw}
\alias{qchisqw}
\title{Quantile function of a weighted chi-square distribution}
\usage{
qchisqw(p, weights, lower.tail = TRUE, epsw = 1e-05, epsv = 1e-12, df0 = 2)
}
\description{
Returns the quantile of a distribution that arises when independent
chisquare-distributed variables with \code{df0} degrees of freedom
are summed on the logarithmic scale with weights in \code{w}
}
