% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{R_add_lib_startup}
\alias{R_add_lib_startup}
\title{Set a library to load automatically}
\usage{
R_add_lib_startup(pckg, Rprof_file, check_duplicate_load = TRUE)
}
\description{
Add a library to the list of libraries loaded automatically at startup of R.
}
\examples{
#R_add_lib_startup('utilbox')   # run this example only if you want to set 'utilbox' to load on startup

}
\seealso{
Other package-related functions provided by utilbox: 
\code{\link{R_default_Rprofile_file}()},
\code{\link{R_user_home}()},
\code{\link{apply_pckg}()},
\code{\link{list_package}},
\code{\link{llibrary}()},
\code{\link{set_pkglib}()}
}
\concept{package-related functions provided by utilbox}
