% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{R_append_default_pckgs}
\alias{R_append_default_pckgs}
\title{Append a package to R's defaultly loaded packages}
\usage{
R_append_default_pckgs(..., system_var_name = "R_DEFAULT_PACKAGES")
}
\description{
This appends the specified package names to the system
variable which is consulted by R for what to load on
startup.
}
\details{
Note: On Windows, this does not seem to do anything.
Best use the '.Rprofile' file in the user's home directory
for instance via [R_add_lib_startup].
}
