% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_defaults.R
\name{R_defaults}
\alias{R_defaults}
\alias{R_get_default_pckgs}
\alias{R_append_default_pckgs}
\alias{R_add_lib_startup}
\alias{R_add_code_startup}
\alias{R_del_code_startup}
\title{Default setting for R}
\usage{
R_get_default_pckgs()

R_append_default_pckgs(..., system_var_name = "R_DEFAULT_PACKAGES")

R_add_lib_startup(pckg, Rprof_file, check_duplicate_load = TRUE)

R_add_code_startup(code, Rprof_file)

R_del_code_startup(
  code,
  Rprof_file,
  exact_match = FALSE,
  remove_all_matches = TRUE
)
}
\description{
\code{R_get_default_pckgs()} lists the packages that R loads on startup.

\code{R_add_lib_startup()} adds a library to the list of libraries
loaded automatically at startup of R.

\code{R_append_default_pckgs()} appends the specified package names to
the system variable 'R_DEFAULT_PACKAGES', which is supposed to be
consulted by R for what to load on startup. However, on Windows, this
does not seem to do anything. It is better to use the '.Rprofile'
file in the user's home directory for instance.

\code{R_add_code_startup()} adds code (supplied as character string
vector \code{code}) and adds lines to the .Rprofile file for the code to
evaluated at R's start up.

\code{R_del_code_startup()} removes all lines that match the supplied
code.

Do not forget to double-backslash the supplied code.
}
\examples{
#R_add_lib_startup('utilbox')   # run this example only if you want to set 'utilbox' to load on startup

R_add_code_startup("print('Hi!')")   # check .Rprofile
R_del_code_startup("print('Hi!')")   # delete what was added

}
\seealso{
Other system-related functions provided by utilbox: 
\code{\link{R_gui_adjust}()},
\code{\link{R_session_info}()},
\code{\link{append_system_PATH}()},
\code{\link{environment_variables}},
\code{\link{get_system_PATH}()},
\code{\link{is_win}()},
\code{\link{pause}},
\code{\link{stopping_control}},
\code{\link{timest}()}
}
\concept{system-related functions provided by utilbox}
