% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{R_gui_adjust}
\alias{R_gui_adjust}
\alias{R_gui_reset}
\title{Adjust R GUI}
\usage{
R_gui_adjust(cmd = c("points = 7", "rows = 43", "columns = 177"))

R_gui_reset()
}
\description{
Change the way the GUI looks (e.g. font size, font style). Without 
changing the default arguments, a call to \code{Rgui_adjust} makes 
the font smaller so that more info fits the screen, while \code{Rgui_reset}
resets the settings to their defaults.
}
\seealso{
Other system-related functions provided by utilbox: 
\code{\link{.q}()},
\code{\link{R_session_info}()},
\code{\link{R_system_call_args}()},
\code{\link{append_system_PATH}()},
\code{\link{code_startup}},
\code{\link{environment_variables}},
\code{\link{get_system_PATH}()},
\code{\link{is_win}()},
\code{\link{timest}()}

Other system-related functions provided by utilbox: 
\code{\link{.q}()},
\code{\link{R_session_info}()},
\code{\link{R_system_call_args}()},
\code{\link{append_system_PATH}()},
\code{\link{code_startup}},
\code{\link{environment_variables}},
\code{\link{get_system_PATH}()},
\code{\link{is_win}()},
\code{\link{timest}()}
}
\concept{system-related functions provided by utilbox}
