% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_info.R
\name{R_session_info}
\alias{R_session_info}
\alias{R_system_call_args}
\alias{R_user_home}
\alias{R_default_Rprofile_file}
\title{R session info}
\usage{
R_system_call_args(trailingOnly = TRUE)

R_session_info()

R_user_home(fsep = .Platform$file.sep)

R_default_Rprofile_file()
}
\description{
\code{R_system_call_args()} returns the trailing arguments of the call
that invoked the current R session. Useful when calling Rscript on a
script and modifying its behavior via flags

\code{R_session_info()} prints information about the current R session
including the version of R and the trailing arguments of the system
call that started this session.

\code{R_user_home()} returns the default path to the .Rprofile file in
the user's home directory.

\code{R_default_Rprofile_file()} returns the default path to the
.Rprofile file in the user's home directory.
}
\examples{
R_system_call_args(trailingOnly=FALSE)

}
\seealso{
Other system-related functions provided by utilbox: 
\code{\link{R_defaults}},
\code{\link{R_gui_adjust}()},
\code{\link{environment_variables}},
\code{\link{is_win}()},
\code{\link{pause}},
\code{\link{stopping_control}},
\code{\link{system_path}},
\code{\link{timest}()}
}
\concept{system-related functions provided by utilbox}
