% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{R_system_call_args}
\alias{R_system_call_args}
\title{Get trailing arguments}
\usage{
R_system_call_args(trailingOnly = TRUE)
}
\description{
Returns the trailing arguments of the call that invoked the current R session.
Useful when calling Rscript on a script and modifying its behavior via flags
}
\examples{
R_system_call_args(trailingOnly=FALSE)

}
\seealso{
Other system-related functions provided by utilbox: 
\code{\link{.q}()},
\code{\link{R_gui_adjust}()},
\code{\link{R_session_info}()},
\code{\link{append_system_PATH}()},
\code{\link{code_startup}},
\code{\link{environment_variables}},
\code{\link{get_system_PATH}()},
\code{\link{is_win}()},
\code{\link{timest}()}
}
\concept{system-related functions provided by utilbox}
