% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{R_user_home}
\alias{R_user_home}
\title{Default location for the .Rprofile file}
\usage{
R_user_home(fsep = .Platform$file.sep)
}
\description{
Returns the default path to the .Rprofile file in the
user's home directory.
}
\seealso{
Other package-related functions provided by utilbox: 
\code{\link{R_add_lib_startup}()},
\code{\link{R_default_Rprofile_file}()},
\code{\link{apply_pckg}()},
\code{\link{list_package}},
\code{\link{llibrary}()},
\code{\link{set_pkglib}()}
}
\concept{package-related functions provided by utilbox}
