% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{random_filename}
\alias{random_filename}
\title{Generate a random file name}
\usage{
random_filename(
  path = ".",
  nchar = 3,
  chars = c(letters, LETTERS, 0:9, "_-\%#@")
)
}
\description{
\code{random_filename()} generates a random file name of length \code{nchar}
by drawing from the set defined via \code{chars}.
}
\seealso{
Other file system function provided by utilbox: 
\code{\link{check_file_locked}()},
\code{\link{clean_filename}()},
\code{\link{file_backup}()},
\code{\link{file_can_open_check}()},
\code{\link{file_empty}()},
\code{\link{file_size}()},
\code{\link{file_sort}()},
\code{\link{separate_path}()}
}
\concept{file system function provided by utilbox}
