% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_char}
\alias{read_char}
\title{Reads file}
\usage{
read_char(file, encoding)
}
\description{
Reads a text from a file as a single string. By default, reads the
entire file.
}
\concept{file-reading functions provided by utilbox}
