% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipping.R
\name{read_table_zip}
\alias{read_table_zip}
\title{Read file inside a zip archive}
\usage{
read_table_zip(
  zipfiles,
  files = NULL,
  pattern = NULL,
  nonames = FALSE,
  maxnfiles = Inf,
  skipnfiles = 0,
  trace = 0,
  maxnchar = 128,
  solve_long_name = TRUE,
  sort_by_name = TRUE,
  long_action = c("copy", "rename"),
  attempt_read_plain = FALSE,
  fun_read = read.table,
  ...
)
}
\description{
Reads a table from within a zip file (equivalent to read.table except the
input file is expected to be a zip archive)
}
\seealso{
Other archiving utilities provided by utilbox: 
\code{\link{list_zip}()},
\code{\link{unzip_files}()},
\code{\link{zip_all_in_path}()},
\code{\link{zip_files_pattern}()},
\code{\link{zip_files}()},
\code{\link{zip_file}()}
}
\concept{archiving utilities provided by utilbox}
