% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipping.R
\name{read_zip}
\alias{read_zip}
\title{Read file inside a zip archive}
\usage{
read_zip(
  zipfiles,
  files = NULL,
  pattern = NULL,
  maxnfiles = Inf,
  skipnfiles = 0,
  nonames = FALSE,
  sort_by_name = TRUE,
  solve_long_name = TRUE,
  maxnchar = 128,
  long_action = c("copy", "rename"),
  attempt_read_plain = FALSE,
  trace = 0,
  fun_read = read.table,
  ...
)
}
\description{
\code{read_zip()} reads file from within a zip archive. The contents
are read using the function in \code{fun_read} (\code{utils::read.table}
by default). Specific files to read from the zip archive can be
supplied via \code{files}. Use other arguments to control which files
get read such as \code{pattern} (only file matching the pattern are
read from the archive), \code{maxnfiles} (to limit the number of files
to read), \code{skipnfiles} (to skip reading some files). The contents
are returned as a named list (unless \code{nonames=TRUE}, when the list
is unnamed) sorted by name (unless \code{sort_by_name=FALSE}). Attempts
are made to deal with excessively long file names (unless
\code{solve_long_name=FALSE}), while \code{maxnchar} sets the cut-off for
what is considered a long file name and \code{long_action} determines
whether this happens via renaming (when \code{long_action='rename'})
or by copying (when \code{long_action='copy'}).
}
\seealso{
Other compression-related utilities provided by utilbox: 
\code{\link{list_zip}()},
\code{\link{unzip_files}()},
\code{\link{zip_all_in_path}()},
\code{\link{zip_files_pattern}()},
\code{\link{zip_files}()},
\code{\link{zip_file}()}
}
\concept{compression-related utilities provided by utilbox}
