% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{remova_objects}
\alias{remova_objects}
\alias{.rma}
\alias{.rme}
\title{Remove objects}
\usage{
.rma(all.names = FALSE, keep = ".utilbox", envir = .GlobalEnv)

.rme(..., envir = parent.frame())
}
\description{
\code{.rma()} is a shorthand for removal of all objects (including
hidden ones whose names start with a dot) from the environment in
\code{envir} (\code{.GlobalEnv} by default). This function is basically an
alias for \code{rm(list=ls(envir=envir))}. Optionally, names of
objects that should be excluded from removal can be supplied via \code{keep}.

\code{rme()} is a selective delete, which removes all supplied objects in the
environment specified
in \code{envir}. Unlike the basic \code{base::rm} function, the current
function checks if the supplied objects exist in order to avoid
errors. Objects can be supplied as character strings or as symbols.
}
\examples{
random_object = '1'
rme(random_object)      # delete successfully
rme(random_object)      # nothing to delete, but no error/warning

}
\concept{object listing, object loading and removing functions provided by utilbox}
