% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{renaming_operators}
\alias{renaming_operators}
\alias{\%->\%}
\alias{\%<-\%}
\title{Renaming operators}
\usage{
from \%->\% to

to \%<-\% from
}
\value{
The value originally found in \code{from}.
}
\description{
Operators that change the name of an object by reassigning the
value from its left-hand side (for \code{\%->\%}) or right-hand side
(for \code{\%->\%}) and assigning the value to the variable on the
other side, thus effectively performing renaming of an object.
}
\examples{
a = 100; a \%->\% b; b; exists('a')  # after the call b is 100 and a does not exist
a = 100; b \%<-\% a; b; exists('a')  # same effect, opposite direction of syntax

}
\seealso{
Other operators provided by utilbox: 
\code{\link{\%anyin\%}()},
\code{\link{\%modify\%}()},
\code{\link{concatenate}},
\code{\link{not_in}},
\code{\link{operator_NULL}},
\code{\link{operator_append}},
\code{\link{operator_c}},
\code{\link{operator_match}},
\code{\link{operator_try}}
}
\concept{operators provided by utilbox}
