% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting.R
\name{reord}
\alias{reord}
\title{Reorder an object}
\usage{
reord(x, order)
}
\arguments{
\item{x}{An object to be sorted.}

\item{order}{The new order of the elements of \code{x}.}
}
\value{
A reordering of \code{x} according to \code{order}.
}
\description{
Reorders the object in \code{x} according to the order given in \code{order}.
}
