% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeat.R
\name{rep2}
\alias{rep2}
\title{Repeat object}
\usage{
rep2(x, each, ...)
}
\description{
An upgraded version of \code{rep}. Allows for different number of
occurrences for each value supplied in the first argument.
}
\examples{
rep2(c('a','b'), e=3:4)

}
\seealso{
Other sequence-related functions provided by utilbox: 
\code{\link{every_nth}()},
\code{\link{filter_by_position}},
\code{\link{filter_by}},
\code{\link{groups_of_unique}()},
\code{\link{h1}()},
\code{\link{insert}()},
\code{\link{is_between}()},
\code{\link{is_run}},
\code{\link{last_element}()},
\code{\link{replace2}()},
\code{\link{replace3}()},
\code{\link{seq2}()},
\code{\link{seq_around}()},
\code{\link{sizechecks}},
\code{\link{split_into_groups}()},
\code{\link{to_factor}()},
\code{\link{what_bag}()}
}
\concept{sequence-related functions provided by utilbox}
