% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\name{rep_df}
\alias{rep_df}
\title{Repeat data frame}
\usage{
rep_df(x, n, fun_bind = rbind)
}
\description{
\code{rep_df()} repeats the input (\code{x}) a total of \code{n}
times and bind the result together using the function specified in
\code{bind}. Primarily intended for data frames (hence the name) and
matrices, but can be used for objects of another type (e.g. lists) as
long as the (default or user-supplied) binding function \code{bind} is
applicable to that type.
}
\examples{
rep_df(tibble(x=1:2), 3)
rep_df(1:3, 3)             # row-binds as a row vector
rep_df(t(t(1:3)), 3)       # row-binds as a column vector
rep_df(t(t(1:3)), 3)       # row-binds as a column vector

}
\seealso{
Other matrix/data-frame functions provided by utilbox: 
\code{\link{add_col}()},
\code{\link{identical_cols}()},
\code{\link{join_recurse}()},
\code{\link{rev_rows}()},
\code{\link{split_rows}()}
}
\concept{matrix/data-frame functions provided by utilbox}
