% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{rep_diag}
\alias{rep_diag}
\alias{rep_matrix}
\alias{rep_matrix_rbind}
\alias{rep_matrix_cbind}
\title{Create multiple copies of a matrix and bind them together}
\usage{
rep_diag(x, times = 1, sparse = FALSE)

rep_matrix(x, times = 1, sparse = FALSE)

rep_matrix_rbind(x, times = 1, sparse = FALSE)

rep_matrix_cbind(x, times = 1, sparse = FALSE)
}
\description{
\code{rep_diag()} makes \code{times} copies of a matrix and produces a block
diagonal matrix with the copies.
}
\details{
\code{rep_matrix()} returns a list of \code{times} copies of a matrix.

\code{rep_matrix_rbind()} and \code{rep_matrix_cbind} each return a matrix
made up of \code{times} copies of a matrix bound together by rows and
columns respectively.
}
