% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{replace3}
\alias{replace3}
\title{Value replacement}
\usage{
replace3(x, new_values, format)
}
\description{
\code{replace3()} replaces the values in \code{x} with \code{new_values} in the
order of occurrence. If \code{new_values} is missing, it is set to the
integers between 1 and the number of unique elements in \code{x}.
Optionally, via the argument \code{format} a function can be supplied
which attempts to set the class of the output (e.g.
\code{format=as.numeric}).
}
\examples{
x = c('b','b','aa','a','b','c','a')
replace3(x)
replace3(x, c('bar','Aaron','and','Cuba'))
replace3(x, format=as.numeric)
replace3(x, c('bar','Aaron','and','Cuba'), format=as.numeric)  # produces a warning about NAs during conversion

}
\seealso{
Other sequence-related functions provided by utilbox: 
\code{\link{every_kth}()},
\code{\link{filter_by_position}},
\code{\link{filter_by}},
\code{\link{groups_of_unique}()},
\code{\link{h1}()},
\code{\link{insert}()},
\code{\link{is_between}()},
\code{\link{is_run}},
\code{\link{last_element}()},
\code{\link{rep2}()},
\code{\link{replace2}()},
\code{\link{seq2}()},
\code{\link{seq_around}()},
\code{\link{sizechecks}},
\code{\link{split_into_groups}()},
\code{\link{to_factor}()},
\code{\link{what_bag}()}
}
\concept{sequence-related functions provided by utilbox}
