% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_script.R
\name{replace_code_tag}
\alias{replace_code_tag}
\title{Replace code tag in R help text}
\usage{
replace_code_tag(x, rds, add_brackets = FALSE, as_link = FALSE)
}
\description{
\code{replace_code_tag()} replaces all occurences of "\code{word}" with "\code{word}"
or "\code{word()}" (when \code{add_brackets=TRUE}) or "\code{\link{word}}" (when \code{as_link=TRUE}).
}
