% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{rnormpval}
\alias{rnormpval}
\title{Draw a normal distribution p-value}
\usage{
rnormpval(
  M,
  mu,
  Sigma = def_Sigma(rho, N = length(mu)),
  rho = 0,
  two.sided = FALSE,
  trace = 1,
  what = c("p", "X", "pX")
)
}
\description{
Draws a random sample of p-values relative to the normal distribution with 
parameters \code{m} (mean) and \code{s} (standard deviation).
}
