% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding.R
\name{rounding}
\alias{rounding}
\alias{ndigits}
\alias{ndecimals}
\alias{nsignif}
\alias{is_round}
\alias{int_part}
\alias{frac_part}
\alias{round_nearest}
\alias{round_nearest_power}
\alias{rsignif}
\alias{round2}
\alias{signif2}
\title{Number of digits}
\usage{
ndigits(x)

ndecimals(x)

nsignif(x)

is_round(x, eps = .Machine$double.eps)

int_part(x)

frac_part(x)

round_nearest(x, b = 1)

round_nearest_power(x, base = 10)

rsignif(x, ndig = 0)

round2(x, nd = 0)

signif2(x, ...)
}
\description{
\code{ndigits()} determines the number of digits for an integer or a
real (for which the output reflects only the integer part).

\code{ndecimals()} determines how many non-zero decimal places are
needed to print a number.

\code{nsignif()} calculates how many significant digits the supplied
numbers are rounded to.

\code{is_round()} checks if values in \code{x} are round (i.e. equal to
integers).

\code{int_part()} returns the integer parts of the values in in the
supplied vector.

\code{frac_part()} returns the fractional parts of the values in the
supplied vector.

\code{round2()} rounds the elements of vector \code{x} to \code{nd} digits while
making sure \code{x} does not become non-unique (in case it was unique).

\code{round_nearest()} rounds a number (can be a vector) to the nearest
multiple of \code{b}.

\code{round_nearest_power()} rounds a number (can be a vector) to the
nearest power of \code{base}.

\code{rsignif()} rounds \code{x} to \code{ndigit} digits after the decimal point
without losing precision to the left of the decimal point.

\code{signif2()} round numbers by leveraging \code{base::signif} except
with a wrapper which it allows a non-numeric input without throwing
an error.
}
\seealso{
Other check-performing functions provided by utilbox: 
\code{\link{anyNonNA}()},
\code{\link{extract_by_positiveness_negativeness}},
\code{\link{is_number}()},
\code{\link{numchecks}}
}
\concept{check-performing functions provided by utilbox}
