% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{rposdefmat}
\alias{rposdefmat}
\title{Draw a positive-definite matrix}
\usage{
rposdefmat(n, ev = runif(n, 1, 10), all.positive = FALSE)
}
\description{
Generates a "random" (more like arbitrary) positive-definite matrix with 
user-specified positive eigenvalues. If eigenvalues are not specified, 
they are generated from a uniform distribution on (1,10)
}
