% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{rvariance}
\alias{rvariance}
\title{Draw a random variance matrix with sigma2 on the diagonal}
\usage{
rvariance(n, sigma2 = rep(1, n), ev = runif(n, 1, 200))
}
\description{
Draw a random variance matrix with sigma2 on the diagonal
}
