% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{samplepdf}
\alias{samplepdf}
\title{Samples n variables from an arbitrary density function specified in 'pdf'}
\usage{
samplepdf(n, pdf, ..., spdf.lower = -Inf, spdf.upper = Inf)
}
\description{
Samples n variables from an arbitrary density function specified in 'pdf'
}
