% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripting.R
\name{script_dir}
\alias{script_dir}
\title{Directory of the current script}
\usage{
script_dir(n = 0)
}
\description{
\code{script_dir} returns the directory of the script file from which
it is called (via \code{source}).
}
