% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{set_class}
\alias{set_class}
\alias{has_class}
\title{Set class}
\usage{
set_class(x, class)

has_class(x, class)
}
\description{
\code{set_class()} attempts to set the class of \code{x} to \code{class} by
calling \verb{as.*} where \code{*} is substituted for the character value in
\code{class}.

\code{has_class()} checks if the class given in \code{class} is among the
classes of \code{x}.
}
\examples{
set_class(1, 'character')

has_class(1, 'character')
has_class(set_class(1, 'character'), 'character')

}
