% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{set_pkglib}
\alias{set_pkglib}
\title{Add path to the package path}
\usage{
set_pkglib(libpath)
}
\arguments{
\item{libpath}{path to be added}
}
\value{
A character vector with each command of the input function
as an element.
}
\description{
Adds a supplied path to the list of paths for searching packages.
}
\details{
Adds a supplied path to the list of paths for searching
packages. This is useful for instance when the default path is in a
location which is write-protected and packages need to be installed
into an alternative location.
}
\seealso{
Other package-related functions provided by utilbox: 
\code{\link{apply_pckg}()},
\code{\link{lib_dependencies}()},
\code{\link{list_installed_packages}()},
\code{\link{llibrary}()}
}
\concept{package-related functions provided by utilbox}
