% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{setdiffsym}
\alias{setdiffsym}
\title{Difference of sets}
\usage{
setdiffsym(x, y, labels = c("x", "y"))
}
\description{
Determines the \"symmetric difference\" of the supplied sets, 
which is all of the elements that are outside the intersection
of the supplied sets.
}
\examples{
setdiffsym(1:5,3:7)
setdiffsym(1:5,3:7, 'vector')

}
\seealso{
Other sequence-related functions provided by utilbox: 
\code{\link{as_factor}()},
\code{\link{dim_check}},
\code{\link{every_nth}()},
\code{\link{filter_by_position}},
\code{\link{filter_by}},
\code{\link{groups_of_unique}()},
\code{\link{h1}()},
\code{\link{insert}()},
\code{\link{is_between}()},
\code{\link{is_empty}()},
\code{\link{is_run}},
\code{\link{last_element}()},
\code{\link{midpoints}()},
\code{\link{nunique}()},
\code{\link{or_between}()},
\code{\link{rep2}()},
\code{\link{replace2}()},
\code{\link{replace3}()},
\code{\link{seq2}()},
\code{\link{seq_around}()},
\code{\link{shift}()},
\code{\link{split_into_groups}()},
\code{\link{unname_factor}()},
\code{\link{what_bag}()}
}
\concept{sequence-related functions provided by utilbox}
