% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets.R
\name{sets}
\alias{sets}
\alias{mintersect}
\alias{setdiffsym}
\alias{is_subset}
\title{Sets}
\usage{
mintersect(...)

setdiffsym(x, y, labels = c("x", "y"), report_counts = TRUE)

is_subset(x, y)
}
\description{
\code{mintersect()} finds the intersection of multiple sets.

\code{setdiffsym()} determines the \"symmetric difference\" of the
supplied sets, which is all of the elements that are outside the
intersection of the supplied sets.

\code{is_subset()} checks whether \code{x} is a subset of \code{y}.
}
\examples{
setdiffsym(1:5,3:7)
setdiffsym(1:5,3:7, 'vector')

}
\seealso{
Other numeric functions provided by utilbox: 
\code{\link{Invert}()},
\code{\link{downweight}()},
\code{\link{extract_by_value}},
\code{\link{locate_position_of_value}},
\code{\link{lower_tri}()},
\code{\link{min_dif}()},
\code{\link{permutations}},
\code{\link{replace_value}},
\code{\link{sequences}},
\code{\link{to_base_k}()},
\code{\link{which_max}()}
}
\concept{numeric functions provided by utilbox}
