% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripting.R
\name{setwd2}
\alias{setwd2}
\title{Set working directory}
\usage{
setwd2(dir, create = TRUE, ask = TRUE, dir2 = "")
}
\description{
A version of \code{setwd} which attempts to create the given path when
it does not exist. Unlike \code{setwd}, which throws an error when no path
is supplied, \code{setwd2} with missing path argument sets the working
directory to the output of \code{script_dir}. Additionally, it also takes
the argument \code{dir2} whose value is appended to the path being set.
}
