% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtp.R
\name{sfg.test}
\alias{sfg.test}
\title{Storey test (aka Storey-Fincher-Goncharuk test)}
\usage{
sfg.test(p, alpha = 0.05, lamSFG = 0.5, lam = 2, scale_up = TRUE, what = "nr")
}
\description{
Storey test (aka Storey-Fincher-Goncharuk test)
}
