% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{shift}
\alias{shift}
\title{Shift elements in an object}
\usage{
shift(x, lag = 1, rotate = TRUE, value_rotated = NA)
}
\description{
Shift (rotate) the elements in a vector in \code{x} by \code{lag} spaces.
With `rotate=FALSE` the rotated elements are replaced with `value_rotated`
(`NA` by default).
}
\seealso{
Other sequence-related functions provided by utilbox: 
\code{\link{as_factor}()},
\code{\link{dim_check}},
\code{\link{every_nth}()},
\code{\link{filter_by_position}},
\code{\link{filter_by}},
\code{\link{groups_of_unique}()},
\code{\link{h1}()},
\code{\link{insert}()},
\code{\link{is_between}()},
\code{\link{is_empty}()},
\code{\link{is_run}},
\code{\link{last_element}()},
\code{\link{midpoints}()},
\code{\link{nunique}()},
\code{\link{or_between}()},
\code{\link{rep2}()},
\code{\link{replace2}()},
\code{\link{replace3}()},
\code{\link{seq2}()},
\code{\link{seq_around}()},
\code{\link{setdiffsym}()},
\code{\link{split_into_groups}()},
\code{\link{unname_factor}()},
\code{\link{what_bag}()}
}
\concept{sequence-related functions provided by utilbox}
