% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{show_Colors}
\alias{show_Colors}
\title{Illustrate the utilbox color palette}
\usage{
show_Colors(
  n,
  col,
  cex = 3,
  pch = 16,
  pos1 = 2,
  pos2 = 4,
  adj1 = -0.25,
  adj2 = 0.2,
  sort = FALSE
)
}
\description{
Produces a plot which shows the colors returned by \code{palette_ub},
via a call to \code{Colors}.
}
\examples{
show_Colors(30)

}
\seealso{
Other plotting-related functions provided by utilbox: 
\code{\link{Colors}()},
\code{\link{add_box_above}()},
\code{\link{add_tight_axes}()},
\code{\link{all_devs}},
\code{\link{as_color}()},
\code{\link{get_legend}()},
\code{\link{hist2}()},
\code{\link{p16}()},
\code{\link{palette_ub}()}
}
\concept{plotting-related functions provided by utilbox}
