% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizechecks.R
\name{sizechecks}
\alias{sizechecks}
\alias{is_empty}
\alias{is_empty.default}
\alias{is_empty.data.frame}
\alias{has_empty}
\alias{is_not_empty}
\alias{is_scalar}
\alias{is_dimone}
\alias{is_dimtwo}
\alias{is_dimtwoplus}
\title{Size checks}
\usage{
is_empty(...)

\method{is_empty}{default}(x)

\method{is_empty}{data.frame}(x)

has_empty(x)

is_not_empty(...)

is_scalar(x)

is_dimone(x)

is_dimtwo(x)

is_dimtwoplus(x)
}
\description{
\code{isempty()} checks whether the supplied object has zero elements.
It is similar to \code{rlang::is_empty()}, except that it assesses
zero-row and/or zero-column data frames as empty by looking at
the product of the its dimensions instead of its length.

\code{has_empty()} is an element-wise check for emptiness. It applies
\code{is_empty()} to the elements of its argument via \code{base::sapply}. Useful
for checking emptiness within a \code{list}-class object (e.g., list, data frame).

\code{is_not_empty()} is a negation of \code{is_empty()}.

\code{is_scalar()} checks whether the supplied is a scalar, i.e.
unidimensional of size 1 (e.g., length-one vector, length-one list, etc.).

\code{is_dimone()} checks for unidimensionality (e.g. vector, list, etc.).

\code{is_dimtwo()} checks for bidimensionality (e.g. matrix, data frame).

\code{is_dimtwoplus()} checks whether the dimensionality is at least 2 (e.g.
matrix, data frame, 2D array, 3D array).
}
\seealso{
Other sequence-related functions provided by utilbox: 
\code{\link{every_kth}()},
\code{\link{filter_by_position}},
\code{\link{filter_by}},
\code{\link{groups_of_unique}()},
\code{\link{h1}()},
\code{\link{insert}()},
\code{\link{is_between}()},
\code{\link{is_run}},
\code{\link{last_element}()},
\code{\link{rep2}()},
\code{\link{replace2}()},
\code{\link{replace3}()},
\code{\link{seq2}()},
\code{\link{seq_around}()},
\code{\link{split_into_groups}()},
\code{\link{to_factor}()},
\code{\link{what_bag}()}
}
\concept{sequence-related functions provided by utilbox}
