% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting.R
\name{sort_df}
\alias{sort_df}
\title{Sort a data frame}
\usage{
sort_df(x, ..., decreasing = FALSE)
}
\arguments{
\item{x}{data frame/matrix/array to be sorted on a per-column basis.}

\item{...}{names of colums to sort by (symbolic).}

\item{decreasing}{Enable decreasing order sorting. Defaults to
\code{FALSE}.}
}
\value{
An object of the same type as \code{p} with sorted columns.
}
\description{
\code{sort_df()} sorts the columns of a given data frame, matrix, or any
two-dimensional array according to the supplied columns.
}
\examples{
x = data.frame(A = c(3,2,3,4,4,4,2,2), B = c('d','e','a','r','t','m','f','a2'), C=8:1)
# sort on only one column
sort_df(x, A)
sort_df(x, C)      # effectively, reverse rows
# sort on multiple columns
sort_df(x, A, B)

}
\seealso{
Other sorting utilities provided by utilbox: 
\code{\link{sort_by_pattern}()},
\code{\link{sort_pval_matrix_cols}()}
}
\concept{sorting utilities provided by utilbox}
