% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting.R
\name{sort_pval_matrix}
\alias{sort_pval_matrix}
\title{Sort a matrix of p-values (or any matrix, really)}
\usage{
sort_pval_matrix(p, n1 = 0, decreasing = FALSE)
}
\arguments{
\item{p}{Matrix of p-values to be sorted on a per-column basis.}

\item{n1}{The number of top rows that are to be sorted separately
from the rest of the rows.}

\item{decreasing}{Enable decreasing order sorting. Defaults to \code{FALSE}.}
}
\value{
An object of the same type as \code{p} with sorted columns.
}
\description{
Sorts the columns of a given matrix. If \code{n1} is supplied and positive,
the rows \code{1:n1} and \code{(n1+1):nrow(p)} are sorted separately.
}
\seealso{
Other sorting utilities provided by utilbox: 
\code{\link{pattern_sort}()}
}
\concept{sorting utilities provided by utilbox}
