% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting.R
\name{sort_pval_matrix_cols}
\alias{sort_pval_matrix_cols}
\title{Sort a matrix (of p-values)}
\usage{
sort_pval_matrix_cols(x, n1 = 0, decreasing = FALSE)
}
\arguments{
\item{n1}{The number of top rows that are to be sorted separately
from the rest of the rows.}

\item{decreasing}{Enable decreasing order sorting. Defaults to
\code{FALSE}.}

\item{p}{Matrix of p-values to be sorted on a per-column basis.}
}
\value{
An object of the same type as \code{p} with sorted columns.
}
\description{
\code{sort_pval_matrix_cols()} sorts the columns of a given matrix. This is
primarily intended for sorting matrices containing p-values with each
column corresponding to a single sample. If \code{n1} is
supplied and positive, the rows \code{1:n1} and \code{(n1+1):nrow(p)}
are sorted separately.
}
\seealso{
Other sorting utilities provided by utilbox: 
\code{\link{sort_by_pattern}()},
\code{\link{sort_df}()}
}
\concept{sorting utilities provided by utilbox}
