% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripting.R
\name{source_files}
\alias{source_files}
\title{Source multiple files}
\usage{
source_files(files, path = ".", announce = TRUE)
}
\description{
Sources all specified files (`files`) relative to 
the given path (`path`).
}
