% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripting.R
\name{source_pattern}
\alias{source_pattern}
\title{Source all files matching a pattern}
\usage{
source_pattern(pattern, path = ".")
}
\description{
Sources all files that match the supplied pattern (`pattern`)
relative to the supplied path (`path`) or the current working
directory.
}
