% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilbox.R
\name{source_utilbox}
\alias{source_utilbox}
\title{Source utilbox code files}
\usage{
source_utilbox(
  ...,
  path,
  normalize = FALSE,
  envir = .GlobalEnv,
  report_new = TRUE
)
}
\description{
Sources the specified files relative to the path in \code{path}. If
\code{files} is not specified, it sources all files in the path.

By default, this sources files from the utilbox location on my
system (the author).
}
\seealso{
Other utilbox-internal functions provided by utilbox: 
\code{\link{utilbox_internals}}
}
\concept{utilbox-internal functions provided by utilbox}
