% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic.R
\name{split_expression}
\alias{split_expression}
\alias{expr2terms}
\title{Split expressions into terms}
\usage{
split_expression(..., assume_language = FALSE, only_names = FALSE)

expr2terms(x, assume_language = FALSE, only_names = FALSE)
}
\description{
\code{split_expression()} takes one or more expressions (optionally as strings)
and splits them up into language elements (via \code{expr2terms()}).
}
\examples{
split_expression(c("I(A/B/1000/(D*G))", 'I(Wealth - Health)*Age'))
lapply(as.list(body(optimize)), FUN=split_expression, assume_language=TRUE) \%>\% unlist()

}
