% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping.R
\name{split_into_groups}
\alias{split_into_groups}
\title{Split an object into groups}
\usage{
split_into_groups(x, n)
}
\value{
List with the i-th element containing the elements that
were assigned into the i-th group (\"bag\")
}
\description{
Split the elements of an object into groups (referred to as
\"bags\") of given size.
}
\examples{
split_into_bags(1:22, 7)

}
\seealso{
Other sequence-related functions provided by utilbox: 
\code{\link{every_kth}()},
\code{\link{filter_by_position}},
\code{\link{filter_by}},
\code{\link{groups_of_unique}()},
\code{\link{h1}()},
\code{\link{insert}()},
\code{\link{is_between}()},
\code{\link{is_run}},
\code{\link{last_element}()},
\code{\link{rep2}()},
\code{\link{replace2}()},
\code{\link{replace3}()},
\code{\link{seq2}()},
\code{\link{seq_around}()},
\code{\link{sizechecks}},
\code{\link{to_factor}()},
\code{\link{what_bag}()}
}
\concept{sequence-related functions provided by utilbox}
