% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\name{split_rows}
\alias{split_rows}
\title{Split a matrix or data frame}
\usage{
split_rows(x, f = seq_len(nrow(x)), drop = FALSE, ...)
}
\arguments{
\item{x}{A two-dimensional object to be split.}

\item{f}{A vector giving the grouping of rows}

\item{drop}{A logical indicating if levels that do not occur should be dropped 
(if \code{f} is a factor or a list). See \code{base::split} for details.}

\item{...}{Further potential arguments passed on to \code{split::base}.}
}
\value{
A list of separated rows.
}
\description{
Takes a matrix, data frame or any two-dimensional object (array) and splits
its rows according to the grouping supplied in \code{f}. Uses \code{base::split} 
as the workhorse, see \code{?base::split} for more details.
}
\seealso{
Other matrix/data-frame functions provided by utilbox: 
\code{\link{add_col}()},
\code{\link{identical_cols}()},
\code{\link{join_recurse}()},
\code{\link{rep_df}()},
\code{\link{rev_rows}()}
}
\concept{matrix/data-frame functions provided by utilbox}
