% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_debug.R
\name{stopping_control}
\alias{stopping_control}
\alias{.sow}
\alias{.roe}
\alias{.q}
\title{Stopping and quitting}
\usage{
.sow(turn_on, announce = TRUE)

.roe(turn_on, announce = TRUE)

.q(save = "no")
}
\description{
\code{.sow()} sets \code{options()$warn} to either 0 or 2, which makes R
either stop when a warning is issued (>=2) or not (0). If the argument
\code{turn_on} is missing, the function acts as a toggle turning stopping
on warnings on and off on alternative calls.

\code{.roe()} analogously toggle the recovering on error option (same
logic as \code{.sow()}).

\code{.q()} is a shortcut for quitting R session without being asked to
save the current workspace.
}
\seealso{
Other system-related functions provided by utilbox: 
\code{\link{R_defaults}},
\code{\link{R_gui_adjust}()},
\code{\link{R_session_info}()},
\code{\link{environment_variables}},
\code{\link{is_win}()},
\code{\link{pause}},
\code{\link{system_path}},
\code{\link{timest}()}
}
\concept{system-related functions provided by utilbox}
