% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_abbreviate}
\alias{str_abbreviate}
\title{Shorten a string}
\usage{
str_abbreviate(x, n1 = 12, n2)
}
\description{
\code{str_abbreviate} takes a string and whenever it is long, it cuts
out the middle section and replaces it with information about how
many characters were cut out.
}
\examples{
let = collapse0(letters)
LET = collapse0(LETTERS)
str_abbreviate(collapse0(c(let, LET, let, LET)))

}
