% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_add_punct}
\alias{str_add_punct}
\title{Add punctuation}
\usage{
str_add_punct(x, p = ".", punct = ".!?", trim = TRUE, split_punct = TRUE)
}
\description{
Adds a punctuation (\code{p}) to the ends of all elements in a 
character vector that do not end in one of the punctuation marks
(\code{punct}).
}
\examples{
str_add_punct(c('hello','world!'), '!') # adds to the 1st, but not the 2nd

}
\seealso{
Other string-manipulation functions provided by utilbox: 
\code{\link{ascii}()},
\code{\link{collapse}()},
\code{\link{combine_names}()},
\code{\link{cumpaste0}()},
\code{\link{padding}},
\code{\link{patternize}()},
\code{\link{str_diff}()},
\code{\link{str_insert}()},
\code{\link{str_is_empty}()},
\code{\link{str_last}()},
\code{\link{str_replace}()},
\code{\link{str_scrub_space}()},
\code{\link{str_split}()},
\code{\link{str_trim_space}()},
\code{\link{strpos}()},
\code{\link{strrev}()},
\code{\link{sub_at}()},
\code{\link{subm}()},
\code{\link{toupperfirst}()}
}
\concept{string-manipulation functions provided by utilbox}
