% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_break_anywhere}
\alias{str_break_anywhere}
\title{Break string into lines (anywhere)}
\usage{
str_break_anywhere(
  x,
  max_width = Inf,
  drop_space_after_newline = FALSE,
  eol = "\\n"
)
}
\description{
Break strings into lines so that the maximum length of a line does
not exceed \code{max_width}. Line breaks are place anywhere (i.e., even
in the middle of words).
}
