% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_deempty}
\alias{str_deempty}
\title{Remove empty substrings}
\usage{
str_deempty(x)
}
\description{
\code{str_deempty()} drops all elements in \code{x} that have only spaces in them.
}
\examples{
str_deempty(c('a','',' ','   '))

}
