% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_drop_empty}
\alias{str_drop_empty}
\title{Remove empty substrings}
\usage{
str_drop_empty(x)
}
\description{
\code{str_drop_empty()} drops all elements in \code{x} that have only spaces in them.
}
\examples{
str_drop_empty(c('a','',' ','   '))

}
