% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{str_escape}
\alias{str_escape}
\alias{str_unescape}
\title{Escape special characters}
\usage{
str_escape(x, specials = NULL, specials0 = c(`\\\\n` = "\\\\\\\\n"))

str_unescape(x, specials = NULL, specials0 = c(`\\\\n` = "\\\\\\\\n"))
}
\description{
Escapes special characters in a string. It replaces `\\n` with a `\\\\n`, etc.
}
\examples{
str_escape('\n')

}
