% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{str_find_in}
\alias{str_find_in}
\alias{str_find_in_function}
\alias{str_find_in_package}
\title{Finding functions by substrings}
\usage{
str_find_in_function(string, fun, envir = parent.frame())

str_find_in_package(string, pckg, what = "all")
}
\description{
\code{str_find_in_funtion()} searches the function body of the specified
function (via \code{fun}) for the given \code{string}. If the string is found,
it returns the corresponding line(s) (as obtained by printing the
result of a call to \code{base::body()}), otherwise \code{NULL} is returned.
}
\details{
\code{str_find_in_package()} searches for the string in functions within
a package. Either all functions when (\code{what='all'}) or only exported
functions (when \code{what='exported'}) are searched.
}
\examples{
str_find_in_function('contrasts', 'lm', envir=asNamespace('stats')) # returns lines 3, 49, 59
str_find_in_package('contrasts', 'stats') # returns 28 functions from the package `stats`

}
