% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_grab}
\alias{str_grab}
\alias{str_first}
\alias{str_last}
\title{Extract the last characters in a string}
\usage{
str_grab(pattern, x, ..., workhorse = gregexpr)

str_first(x, n = 1, drop = FALSE)

str_last(x, n = 1, drop = FALSE)
}
\description{
\code{str_grab()} extracts the substring that matches the pattern
in \code{pattern} from a string.

\code{str_first()} extracts (\code{drop=FALSE}) or removes (\code{drop=TRUE}) the first
\code{n} characters from a string.

\code{str_last()} extracts (\code{drop=FALSE}) or removes (\code{drop=TRUE}) the last
\code{n} characters from a string.
}
\seealso{
Other string-manipulation functions provided by utilbox: 
\code{\link{ascii}()},
\code{\link{collapse}()},
\code{\link{cumpaste0}()},
\code{\link{multi_sub}()},
\code{\link{padding}},
\code{\link{patternize}()},
\code{\link{str_cut}()},
\code{\link{str_diff}()},
\code{\link{str_insert}()},
\code{\link{str_is_empty}()},
\code{\link{str_paste_grid}()},
\code{\link{str_pos}()},
\code{\link{str_rev}()},
\code{\link{str_trim_space}()},
\code{\link{sub_at}()},
\code{\link{toupperfirst}()}
}
\concept{string-manipulation functions provided by utilbox}
