% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_n_lines}
\alias{str_n_lines}
\title{Count the number of lines in a string}
\usage{
str_n_lines(x, eol = "\\\\n")
}
\description{
\code{str_n_lines()} counts the number of lines a string will print on
(without further wrapping). Effectively, it counts the new line symbols
(in \code{eol}) and adds 1.
}
\examples{
str_n_lines('this string\nhas only\nthree lines.')

}
