% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_replace}
\alias{str_replace}
\title{Insert substring into a string (replacive)}
\usage{
str_replace(x, substring, pos, insert_white = FALSE, str2_shift = 0)
}
\description{
Inserts a substring into a string at a given position instead 
of the characters contained in the string.
}
\details{
If the position in \code{pos} is larger than the character
count of the string, the substring either simply attached
to the end of the string (when \code{insert_white=FALSE},
default behavior) or the appropriate number of trailing 
white spaces are inserted before the substring is
attached.
}
\examples{
str_replace('My name is John Doe', ' Jack', 11)
str_replace(c('First name: ','Last name: ') , c('John','Doe'), 20)
str_replace(c('First name: ','Last name: ') , c('John','Doe'), 20, insert=TRUE)

}
\seealso{
Other string-manipulation functions provided by utilbox: 
\code{\link{ascii}()},
\code{\link{collapse}()},
\code{\link{combine_names}()},
\code{\link{cumpaste0}()},
\code{\link{padding}},
\code{\link{patternize}()},
\code{\link{str_add_punct}()},
\code{\link{str_diff}()},
\code{\link{str_insert}()},
\code{\link{str_is_empty}()},
\code{\link{str_last}()},
\code{\link{str_scrub_space}()},
\code{\link{str_split}()},
\code{\link{str_trim_space}()},
\code{\link{strpos}()},
\code{\link{strrev}()},
\code{\link{sub_at}()},
\code{\link{subm}()},
\code{\link{toupperfirst}()}
}
\concept{string-manipulation functions provided by utilbox}
