% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_scrub_space}
\alias{str_scrub_space}
\title{Remove extra white space}
\usage{
str_scrub_space(x, pattern = "[ ]+", s = " ", fixed = FALSE)
}
\description{
Replaces multiple consecutive white spaces with a single one.
}
\examples{
str_scrub_space(' fas  fdas    fdsfas fsdafdfs         dfsa ')

}
\seealso{
Other string-manipulation functions provided by utilbox: 
\code{\link{ascii}()},
\code{\link{collapse}()},
\code{\link{cumpaste0}()},
\code{\link{multi_sub}()},
\code{\link{padding}},
\code{\link{patternize}()},
\code{\link{str_add_punct}()},
\code{\link{str_diff}()},
\code{\link{str_insert}()},
\code{\link{str_is_empty}()},
\code{\link{str_last}()},
\code{\link{str_paste_grid}()},
\code{\link{str_pos}()},
\code{\link{str_rev}()},
\code{\link{str_split}()},
\code{\link{str_trim_space}()},
\code{\link{sub_at}()},
\code{\link{toupperfirst}()}
}
\concept{string-manipulation functions provided by utilbox}
