% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_split}
\alias{str_split}
\alias{str2vector}
\title{String splitting}
\usage{
str_split(x, split, ..., unlist_for_scalar = FALSE)

str2vector(x, split = "", ..., unlist_for_scalar = TRUE)
}
\description{
Splits a string vector `x` by character string `split`. 
`str_split` is basically the same as [base::strsplit]
except that for zero-length strings it does not return 
a zero-length element in the returned list. `str2vector`
is an alias with the default value for split set to '' 
(an empty string) so that it ends up producing a vector
of individual characters. It also unlists the returned
value for scalar `x`.

Splits string into a vector of individual characters.
}
\examples{
str_split(c('a-b','a-c','a-d','','c-d','c-e'),'-')

}
\seealso{
Other string-manipulation functions provided by utilbox: 
\code{\link{ascii}()},
\code{\link{collapse}()},
\code{\link{cumpaste0}()},
\code{\link{padding}},
\code{\link{patternize}()},
\code{\link{str_add_punct}()},
\code{\link{str_diff}()},
\code{\link{str_insert}()},
\code{\link{str_is_empty}()},
\code{\link{str_last}()},
\code{\link{str_paste_grid}()},
\code{\link{str_pos}()},
\code{\link{str_replace}()},
\code{\link{str_scrub_space}()},
\code{\link{str_trim_space}()},
\code{\link{strrev}()},
\code{\link{sub_at}()},
\code{\link{subm}()},
\code{\link{toupperfirst}()}

Other string-manipulation functions provided by utilbox: 
\code{\link{ascii}()},
\code{\link{collapse}()},
\code{\link{cumpaste0}()},
\code{\link{padding}},
\code{\link{patternize}()},
\code{\link{str_add_punct}()},
\code{\link{str_diff}()},
\code{\link{str_insert}()},
\code{\link{str_is_empty}()},
\code{\link{str_last}()},
\code{\link{str_paste_grid}()},
\code{\link{str_pos}()},
\code{\link{str_replace}()},
\code{\link{str_scrub_space}()},
\code{\link{str_trim_space}()},
\code{\link{strrev}()},
\code{\link{sub_at}()},
\code{\link{subm}()},
\code{\link{toupperfirst}()}
}
\concept{string-manipulation functions provided by utilbox}
