% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_strip_eol_space}
\alias{str_strip_eol_space}
\title{Strip space at the end of a line}
\usage{
str_strip_eol_space(y, width)
}
\description{
If a string in \code{y} has a space as each \code{width}+1 position, it is
removed.
}
