% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_wrap}
\alias{str_wrap}
\alias{str_wrap_one}
\title{Wrap text}
\usage{
str_wrap(x, max_width = Inf, eol = "\\n", break_only_at_space = FALSE)

str_wrap_one(
  x,
  max_width = Inf,
  eol = "\\n",
  break_only_at_space = FALSE,
  max_nlines = Inf
)
}
\description{
Wrap text. In other words, place newline symbols (\verb{\\\\n}) along a
string to make its printout (via e.g. \code{base::cat}) \"wrap nicely\",
that is not exceed a given width (set by \code{max_width}).

\code{str_wrap_one} does the actual work for input of length 1,
\code{str_wrap} is a wrapper around it useful for multivariate input (\code{x}).
}
\examples{
string = collapse0(rep(collapse0(letters[1:8]),10), sep=" ")

# insert line breaks at given width
catn(str_wrap(string, max_width=20))

# insert only at spaces
catn(str_wrap(string, max_width=20, break_only_at_space=TRUE))

}
