% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{strpos}
\alias{strpos}
\title{Find the first or last occurrence of a substring}
\usage{
strpos(
  string,
  substring,
  first = TRUE,
  last = FALSE,
  patternize = FALSE,
  fixed = TRUE
)
}
\description{
Find the first/last occurrence of a substring (given as regular pattern) 
in a string
}
\examples{
strpos('hello world.', 'o', first=TRUE)
strpos('hello world.', 'o', last=TRUE)
strpos('hello world.', '[el]', first=TRUE)
strpos('hello world.', '[el]', last=TRUE)
strpos('hello world.', '.')
strpos('hello world.', '.', patternize=TRUE)


}
\seealso{
Other string-manipulation functions provided by utilbox: 
\code{\link{ascii}()},
\code{\link{collapse}()},
\code{\link{combine_names}()},
\code{\link{cumpaste0}()},
\code{\link{padding}},
\code{\link{patternize}()},
\code{\link{str_add_punct}()},
\code{\link{str_diff}()},
\code{\link{str_insert}()},
\code{\link{str_is_empty}()},
\code{\link{str_last}()},
\code{\link{str_replace}()},
\code{\link{str_scrub_space}()},
\code{\link{str_split}()},
\code{\link{str_trim_space}()},
\code{\link{strrev}()},
\code{\link{sub_at}()},
\code{\link{subm}()},
\code{\link{toupperfirst}()}
}
\concept{string-manipulation functions provided by utilbox}
