% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table2}
\alias{table2}
\title{Contingency table}
\usage{
table2(..., obligate = NULL)
}
\description{
Does the same thing as \code{base::table} but makes sure that the 
counts of values in \code{obligate} are included in the final table,
although only if a tabulation of a single vector is performed.
}
\examples{
table2(
}
