% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tablenat}
\alias{tablenat}
\title{Quick tabulation}
\usage{
tablenat(x)
}
\description{
A quicker version of \code{base::table}. Assumes natural number values and every value outside 
1,2,... is counted as 0.
}
