% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtp.R
\name{test_pval}
\alias{test_pval}
\title{Perform multiple testing procedure}
\usage{
test_pval(
  p,
  method = "cbonf",
  alpha = 0.05,
  lamSFG = 0.5,
  lam,
  scale_up = TRUE
)
}
\value{
A vector of rejections.
}
\description{
Given a vector of p-values on input, performs a selected multiple testing
procedure.
}
