% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamp.R
\name{timest}
\alias{timest}
\title{Time stamp}
\usage{
timest(add_time = TRUE, add_pid = TRUE, format = "\%Y\%b\%d\%H\%M\%S")
}
\description{
Generates a "unique" stamp based on the current system time and 
the R process system ID.
}
\examples{
timest()
timest(add_pid=FALSE)
timest(format="\%Y\%m\%d\%H\%M\%S")

}
\seealso{
Other system-related functions provided by utilbox: 
\code{\link{.q}()},
\code{\link{R_gui_adjust}()},
\code{\link{R_session_info}()},
\code{\link{R_system_call_args}()},
\code{\link{append_system_PATH}()},
\code{\link{code_startup}},
\code{\link{environment_variables}},
\code{\link{get_system_PATH}()},
\code{\link{is_win}()}
}
\concept{system-related functions provided by utilbox}
