% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbase.R
\name{to_base_k}
\alias{to_base_k}
\alias{int2k}
\title{Convert base}
\usage{
to_base_k(x, k = 3)

int2k(x, k = 3)
}
\description{
`to_base_k` converts its length-1 argument (assumed to be 
`numeric`) from the decimal base to base `k`.
}
\details{
`int2k` is a vectorized version of `to_base_k`.
}
\seealso{
Other numeric functions provided by utilbox: 
\code{\link{Invert1}()},
\code{\link{Invert}()},
\code{\link{all_permutations}()},
\code{\link{as_numeric}()},
\code{\link{downweight}()},
\code{\link{force_as_}},
\code{\link{frac_part}()},
\code{\link{frac}()},
\code{\link{int_part}()},
\code{\link{lower_tri}()},
\code{\link{make_numeric}()},
\code{\link{max0}()},
\code{\link{min_dif}()},
\code{\link{mintersect}()},
\code{\link{round2}()},
\code{\link{round_nearest_power}()},
\code{\link{round_nearest}()},
\code{\link{rsignif}()},
\code{\link{signif2}()},
\code{\link{unspace}()},
\code{\link{unzero}()},
\code{\link{upper_tri}()},
\code{\link{value_locale}},
\code{\link{which_max}()}
}
\concept{numeric functions provided by utilbox}
