% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{toggle_dump_on_error}
\alias{toggle_dump_on_error}
\alias{.doe}
\title{Toggle dumping on frames on error}
\usage{
toggle_dump_on_error(
  dumpto = "last.dump",
  to_file = FALSE,
  include_GlobalEnv = FALSE
)

.doe(dumpto = "last.dump", to_file = FALSE, include_GlobalEnv = FALSE)
}
\description{
\code{.roe()} (or it alias \code{toggle_dump_on_error()}) toggle between dumping of
frames when an error is encoutered. This is done by setting \code{options()$error}
to a version of \code{utils::dump.frames()}. Optionally, the dumping is done to
to file (when \code{to_file=TRUE}) and/or includes the global environment
(when \code{include_GlobalEnv=TRUE}).
}
