% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{toupperfirst}
\alias{toupperfirst}
\title{Capitalization}
\usage{
toupperfirst(...)
}
\description{
Capitalize first letters of each element in the vector \code{string}.
}
\examples{
toupperfirst('hello')

}
\seealso{
Other string-manipulation functions provided by utilbox: 
\code{\link{ascii}()},
\code{\link{collapse}()},
\code{\link{cumpaste0}()},
\code{\link{padding}},
\code{\link{patternize}()},
\code{\link{str_add_punct}()},
\code{\link{str_diff}()},
\code{\link{str_insert}()},
\code{\link{str_is_empty}()},
\code{\link{str_last}()},
\code{\link{str_paste_grid}()},
\code{\link{str_pos}()},
\code{\link{str_replace}()},
\code{\link{str_scrub_space}()},
\code{\link{str_split}()},
\code{\link{str_trim_space}()},
\code{\link{strrev}()},
\code{\link{sub_at}()},
\code{\link{subm}()}
}
\concept{string-manipulation functions provided by utilbox}
