% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{transfer_objects}
\alias{transfer_objects}
\alias{transfer_object}
\title{Transfer objects between environments}
\usage{
transfer_objects(from, to, what, check_existence = FALSE, delete = TRUE)

transfer_object(from, to, what, check_existence = FALSE, delete = TRUE)
}
\description{
Transfers the specified objects (in \code{what}) or all objects (when
\code{what} not supplied) from one environment (\code{env1}) to another
(\verb{env2'). Optionally, an existence check in the destination  environment is performed (when }check_existence=TRUE`).
}
\examples{


}
