% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilbox.R
\name{utilbox_environment}
\alias{utilbox_environment}
\title{Get the utilbox environment}
\usage{
utilbox_environment(envir = .GlobalEnv)
}
\description{
Creates the working namespace called `.utilbox`, which is used by some of the
functions of the `utilbox` package (e.g. \code{cat0}) to store private variables.
}
\details{
This was used before `utilbox` was an installable package (now basically unnecessary).
}
\seealso{
Other utilbox-internal functions provided by utilbox: 
\code{\link{source_utilbox}()},
\code{\link{utilbox_namespace}()}
}
\concept{utilbox-internal functions provided by utilbox}
