% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilbox.R
\name{utilbox_internals}
\alias{utilbox_internals}
\alias{utilbox_namespace}
\alias{utilbox_environment}
\alias{get_utilbox}
\alias{assign_utilbox}
\alias{update_utilbox}
\alias{init_utilbox}
\alias{init1_utilbox}
\alias{utilbox_inits}
\title{Utilbox internals}
\usage{
utilbox_namespace()

utilbox_environment(envir = .GlobalEnv)

get_utilbox(x, envir = utilbox_namespace())

assign_utilbox(x, val, envir = utilbox_namespace())

update_utilbox(x, val, operation = `+`, envir = utilbox_namespace())

init_utilbox(x, envir = utilbox_namespace())

init1_utilbox(x, envir = utilbox_namespace())

utilbox_inits(x)
}
\description{
\code{get_utilbox} retrieves a value of a variabled named in \code{x} from
the utilbox's package namespace or the hidden environment created in
\code{.GlobalEnv} by a call to \code{utilbox_environment()}.

\code{assign_utilbox} is an analogue which assigns a supplied value
(\code{val}) to a variable (\code{x}).

\code{update_utilbox} updates the existing variable by applying a
function (\code{operation}) with the variable \code{x} and \code{val} supplied to it
as the first two arguments.

\code{init_utilbox} initializes a given variable to its default value
(defined in and obtained via \code{utilbox_inits}). If not variable name
is supplied, all variables defined in \code{utilbox_inits} are initialized.

\code{utilbox_namespace()} returns the environment which is the
namespace of the utilbox package. If it failed for some reason, it
returns creates and/or returns the hidden environment located in
\code{.GlobalEnv}.

\code{utilbox_environment()} creates the working namespace called
\code{.utilbox}, which is used by some of the functions of the \code{utilbox}
package (e.g. \code{cat0}) to store private variables.
}
\seealso{
Other utilbox-internal functions provided by utilbox: 
\code{\link{source_utilbox}()}
}
\concept{utilbox-internal functions provided by utilbox}
