% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilbox.R
\name{utilbox_namespace}
\alias{utilbox_namespace}
\title{Get the namespace of the utilbox package}
\usage{
utilbox_namespace()
}
\description{
Returns the environment which is the namespace of the utilbox package.
If it failed for some reason, it returns creates and/or returns 
the hidden environment located in `.GlobalEnv`.
}
\seealso{
Other utilbox-internal functions provided by utilbox: 
\code{\link{source_utilbox}()},
\code{\link{utilbox_environment}()}
}
\concept{utilbox-internal functions provided by utilbox}
