% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilbox.R
\name{utilbox_namespace_modify}
\alias{utilbox_namespace_modify}
\alias{get_utilbox}
\alias{assign_utilbox}
\alias{update_utilbox}
\alias{init_utilbox}
\alias{init1_utilbox}
\alias{utilbox_inits}
\title{Get/assign/update/initialize a variable from utilbox environment}
\usage{
get_utilbox(x, envir = utilbox_namespace())

assign_utilbox(x, val, envir = utilbox_namespace())

update_utilbox(x, val, operation = `+`, envir = utilbox_namespace())

init_utilbox(x, envir = utilbox_namespace())

init1_utilbox(x, envir = utilbox_namespace())

utilbox_inits(x)
}
\description{
`get_utilbox` retrieves a value of a variabled named in `x` 
from the utilbox's package namespace or the hidden 
environment created in `.GlobalEnv` by a call to
`utilbox_environment()`.
}
\details{
`assign_utilbox` is an analogue which assigns a supplied 
value (`val`) to a variable (`x`).

`update_utilbox` updates the existing variable by 
applying a function (`operation`) with the variable
`x` and `val` supplied to it as the first two arguments.

`init_utilbox` initializes a given variable to its default
value (defined in and obtained via `utilbox_inits`).
If not variable name is supplied, all variables defined
in `utilbox_inits` are initialized.
}
