% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{value_locale}
\alias{value_locale}
\alias{first_nonzero}
\alias{nth_nonzero}
\alias{last_nonzero}
\alias{first_positive}
\alias{last_positive}
\alias{first_above}
\alias{first_above_soft}
\alias{last_above}
\alias{last_above_soft}
\alias{first_below}
\alias{first_below_soft}
\alias{last_below}
\alias{last_below_soft}
\alias{w_nth_nonzero}
\alias{w_last_nonzero}
\alias{w_first_positive}
\alias{w_last_positive}
\alias{w_first_above}
\alias{w_first_above_soft}
\alias{w_last_above}
\alias{w_last_above_soft}
\alias{w_first_below}
\alias{w_first_below_soft}
\alias{w_last_below}
\alias{w_last_below_soft}
\title{Locate value}
\usage{
first_nonzero(x)

nth_nonzero(x, n)

last_nonzero(x)

first_positive(x)

last_positive(x)

first_above(x, y)

first_above_soft(x, y)

last_above(x, y, ...)

last_above_soft(x, y)

first_below(x, y)

first_below_soft(x, y)

last_below(x, y)

last_below_soft(x, y)

w_nth_nonzero(x, n)

w_last_nonzero(x)

w_first_positive(x)

w_last_positive(x)

w_first_above(x, y)

w_first_above_soft(x, y)

w_last_above(x, y)

w_last_above_soft(x, y)

w_first_below(x, y)

w_first_below_soft(x, y)

w_last_below(x, y)

w_last_below_soft(x, y)
}
\description{
`first_nonzero` finds the value of the last non-zero element in `x`.
}
\details{
`nth_nonzero` finds the value of the n-th non-zero element in `x`.

`last_nonzero` finds the value of the last non-zero element in `x`.

`first_positive` finds the value of the first positive element in `x`.

`last_positive` finds the value of the last positive element in `x`.

`first_above` (`first_above_soft`) finds the value of the first element 
in `x` with value larger than (larger than or equal to) `y`.

`last_above` (`last_above_soft`) finds the value of the last element 
in `x` with value larger than (larger than or equal to) `y`.

`first_below` (`first_below_soft`) finds the value of the first element 
in `x` with value smaller than (smaller than or equal to) `y`.

`last_below` (`last_below_soft`) finds the value of the last element 
in `x` with value smaller than (smaller than or equal to) `y`.
}
\seealso{
Other numeric functions provided by utilbox: 
\code{\link{Invert1}()},
\code{\link{Invert}()},
\code{\link{all_permutations}()},
\code{\link{as_numeric}()},
\code{\link{downweight}()},
\code{\link{force_as_}},
\code{\link{frac_part}()},
\code{\link{frac}()},
\code{\link{int_part}()},
\code{\link{lower_tri}()},
\code{\link{make_numeric}()},
\code{\link{max0}()},
\code{\link{min_dif}()},
\code{\link{mintersect}()},
\code{\link{round2}()},
\code{\link{round_nearest_power}()},
\code{\link{round_nearest}()},
\code{\link{rsignif}()},
\code{\link{signif2}()},
\code{\link{to_base_k}()},
\code{\link{unspace}()},
\code{\link{unzero}()},
\code{\link{upper_tri}()},
\code{\link{value_locale_position}},
\code{\link{which_max}()}
}
\concept{numeric functions provided by utilbox}
