% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{value_locale_position}
\alias{value_locale_position}
\alias{w_first_nonzero}
\title{Locate position of a value}
\usage{
w_first_nonzero(x)
}
\description{
`w_first_nonzero` finds the position of the last non-zero element in `x`.
}
\details{
`w_nth_nonzero` finds the position of the n-th non-zero element in `x`.

`w_last_nonzero` finds the position of the last non-zero element in `x`.

`w_first_positive` finds the position of the first positive element in `x`.

`w_last_positive` finds the position of the last positive element in `x`.

`w_last_above` finds the position of the last element in `x` with value larger than `y`.

`w_first_above` finds the position of the first element in `x` with value larger than `y`.

`w_last_below` finds the position of the last element in `x` with value below `y`.

`w_first_below` finds the position of the first element in `x` with value below `y`.

`w_first_nonempty` finds the position of the first non-empty element in `x`.
It is designed as a method, which currenlty only operates on lists.
}
\seealso{
Other numeric functions provided by utilbox: 
\code{\link{Invert1}()},
\code{\link{Invert}()},
\code{\link{all_permutations}()},
\code{\link{as_numeric}()},
\code{\link{downweight}()},
\code{\link{force_as_}},
\code{\link{frac_part}()},
\code{\link{frac}()},
\code{\link{int_part}()},
\code{\link{lower_tri}()},
\code{\link{make_numeric}()},
\code{\link{max0}()},
\code{\link{min_dif}()},
\code{\link{mintersect}()},
\code{\link{round2}()},
\code{\link{round_nearest_power}()},
\code{\link{round_nearest}()},
\code{\link{rsignif}()},
\code{\link{signif2}()},
\code{\link{to_base_k}()},
\code{\link{unspace}()},
\code{\link{unzero}()},
\code{\link{upper_tri}()},
\code{\link{value_locale}},
\code{\link{which_max}()}
}
\concept{numeric functions provided by utilbox}
