% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic.R
\name{void}
\alias{void}
\title{These functions are useful when processing \code{\link{tidyverse}}-style
functions.}
\usage{
void()
}
\description{
These functions are useful when processing \code{\link{tidyverse}}-style
functions.
}
