% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic.R
\name{void_general}
\alias{void_general}
\title{These functions are useful when processing \code{\link{tidyverse}}-style
functions.}
\usage{
void_general(...)
}
\description{
These functions are useful when processing \code{\link{tidyverse}}-style
functions.
}
