% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{vregexpr}
\alias{vregexpr}
\title{Vectorized regular pattern matching}
\usage{
vregexpr(
  pattern,
  text,
  ignore.case = FALSE,
  perl = FALSE,
  fixed = FALSE,
  useBytes = FALSE
)
}
\description{
Vectorized regular pattern matching, which takes multiple patterns.
}
