% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing.R
\name{warn}
\alias{warn}
\title{Print a warning and either stop execution or wait depending on arguments}
\usage{
warn(..., skip1 = 1, skip2 = 1)
}
\description{
Print a warning and either stop execution or wait depending on arguments
}
