% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping.R
\name{what_bag}
\alias{what_bag}
\title{Virtually bag the elements of \code{x}}
\usage{
what_bag(x, bagsize = 20)
}
\description{
"Puts" the elements of \code{x} into virtual bags each of size
\code{bagsize}, starting from bag labelled \"1\", then bag labeled \"2\",
and so on, and returns which bag each element ended up in.
}
\examples{
what_bag(1:40)
what_bag(1:90, 25)

}
\seealso{
Other sequence-related functions provided by utilbox: 
\code{\link{every_kth}()},
\code{\link{filter_by_position}},
\code{\link{filter_by}},
\code{\link{groups_of_unique}()},
\code{\link{h1}()},
\code{\link{insert}()},
\code{\link{is_between}()},
\code{\link{is_run}},
\code{\link{last_element}()},
\code{\link{rep2}()},
\code{\link{replace2}()},
\code{\link{replace3}()},
\code{\link{seq2}()},
\code{\link{seq_around}()},
\code{\link{sizechecks}},
\code{\link{split_into_groups}()},
\code{\link{to_factor}()}
}
\concept{sequence-related functions provided by utilbox}
