% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{y_day}
\alias{y_day}
\title{Get yesterday's date}
\usage{
y_day(
  format = "\%Y-\%m-\%d",
  lag = -1,
  relative_to_day,
  set_return_class = as.character
)
}
\description{
Calculated the date of the previous day or any other day relative
to today in the current time zone in the specified format. How many
days are shifted to the past is controlled via \code{lag}, which accepts
vector input. Negative values in \code{lag} shift the date to the past,
positive values shift the date into the future.
}
\examples{
y_day()          # yesterday
y_day(lag=7)     # a week ago
y_day(lag=-1)    # tomorrow
y_day(lag=-7)    # a week from today
y_day('\%Y')      # the year of yesterday
y_day('\%m')      # the month of yesterday

}
\seealso{
Other calendar functions provided by utilbox: 
\code{\link{closest_day}()},
\code{\link{t_day}()}
}
\concept{calendar functions provided by utilbox}
