% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipping.R
\name{zip_all_in_path}
\alias{zip_all_in_path}
\title{Compress the entire path}
\usage{
zip_all_in_path(
  path = ".",
  check_status = FALSE,
  extras = "-m",
  disable_warning = FALSE
)
}
\arguments{
\item{path}{Path to the files to zip up.}

\item{check_status}{Determines whether a check for zero status code is performed.}

\item{extras}{Modifiers sent to the zip archiver. Defaults to '-m', 
which moves the files into the archive.}

\item{disable_warning}{Disables warning about archiving all files in the given
path. Caution advised.}
}
\value{
List of names of archived files, list of the created archives and the 
       corresponding status codes returned by the archiver.
}
\description{
\code{zip_all_in_path} compresses all files in the supplied path (\code{path}) 
each into its own archive
}
\seealso{
Other archiving utilities provided by utilbox: 
\code{\link{list_zip}()},
\code{\link{read_table_zip}()},
\code{\link{unzip_files}()},
\code{\link{zip_files_pattern}()},
\code{\link{zip_files}()},
\code{\link{zip_file}()}
}
\concept{archiving utilities provided by utilbox}
