% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipping.R
\name{zip_file}
\alias{zip_file}
\alias{unzip_files_single}
\title{Compress files individually}
\usage{
zip_file(
  files,
  extras = "-m",
  appendix = ".zip",
  add_timestamp = FALSE,
  check_status = TRUE
)

unzip_files_single(
  zipfile,
  mask = ".*",
  mask_exclude = FALSE,
  patternize = TRUE
)
}
\arguments{
\item{files}{List of files to compress. Exact file names, not
regular patterns.}

\item{extras}{Modifiers sent to the zip archiver. Defaults to
'-m', which moves the files into the archive.}

\item{appendix}{File name extension added to the name of each file
by the archiver.}

\item{add_timestamp}{Sets whether a unique time stamp is also
added.}

\item{check_status}{Determines whether a check for zero status
code is performed.}
}
\value{
List of file names of the created archive and the
corresponding status codes returned by the archiver.
}
\description{
A collection of functions useful for creating zip archives and
unzipping them.
}
\details{
\code{zip_file} compresses files each into its own archive. Optionally
adding time stamps to the output files.
}
\seealso{
Other archiving utilities provided by utilbox: 
\code{\link{list_zip}()},
\code{\link{read_table_zip}()},
\code{\link{unzip_files}()},
\code{\link{zip_all_in_path}()},
\code{\link{zip_files_pattern}()},
\code{\link{zip_files}()}
}
\concept{archiving utilities provided by utilbox}
