% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.R
\name{append_body}
\alias{append_body}
\title{Append the body of a function}
\usage{
append_body(fun, calls, at_top = FALSE)
}
\description{
Takes an existing function (`fun`) and changes its body
by appending to it the calls supplied in `calls`.
}
\examples{
f = function() { x = pi }
print(f())                           # returns pi (i.e. 3.141593)
f = append_body(f, quote(x^2))
print(f())                           # returns pi^2 (i.e. 9.869604)
f = append_body(f, c(quote(x <- x^2), quote(x-1)))
print(f())                           # returns pi^2 - 1 (i.e. 8.869604)

# modify an existing function
call = str2lang("cat(\"And the symmetric set difference is...\n\")")
f = append_body(setdiffsym, call, at_top=TRUE)
f(1:5, 3:10)

}
\seealso{
Other coding-related functions provided by utilbox: 
\code{\link{call_to_list}},
\code{\link{compare_functions}()},
\code{\link{fun_code_to_text}()},
\code{\link{hijack}()},
\code{\link{this_fun_name}()}
}
\concept{coding-related functions provided by utilbox}
