% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{apply_pckg}
\alias{apply_pckg}
\title{Apply function to an object in a package or environment}
\usage{
apply_pckg(objs, pckg, f, ..., workhorse = sapply, value_error = NA)
}
\description{
Applies function \code{f} to the object(s) whose names are in \code{objs}
and that are found in \code{pckg}, which is primarily intended to be a
package name (assume to be one when of class \code{character}), but can
also be an environment. The elipsis (\code{...}) is supplied to the
function \code{f}. \code{f} can have any number of arguments and only as many
as it accepts are supplied to it (as determined via \code{\link{nformals}}).

This function is useful for applying a function \code{f} (e.g.
\code{\link[base:class]{base::class}}) to exported and/or non-exported objects inside a
package as done for example by \code{\link{object_table}}.
}
\examples{
apply_pckg('rev_cols', 'utilbox', class)    # returns 'function'

}
\seealso{
Other package-related functions provided by utilbox: 
\code{\link{list_package}},
\code{\link{llibrary}()},
\code{\link{set_pkglib}()}
}
\concept{package-related functions provided by utilbox}
