% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_number.R
\name{as_numeric}
\alias{as_numeric}
\title{Type conversion}
\usage{
as_numeric(x)
}
\description{
Convert to type \code{numeric} without any warnings about non-numeric
elements. Non-numeric elements are turned into \code{NA}. Basically, a
less verbose version of \code{base::as.numeric}.
}
\examples{
as_numeric(c("1","2"))
as_numeric(c("1","a"))

}
\seealso{
Other numeric functions provided by utilbox: 
\code{\link{Invert}()},
\code{\link{downweight}()},
\code{\link{extract_by_value}},
\code{\link{force_as_}},
\code{\link{locate_position_of_value}},
\code{\link{lower_tri}()},
\code{\link{make_numeric}()},
\code{\link{min_dif}()},
\code{\link{permutations}},
\code{\link{replace_value}},
\code{\link{sequences}},
\code{\link{sets}},
\code{\link{to_base_k}()},
\code{\link{which_max}()}
}
\concept{numeric functions provided by utilbox}
