% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{assign_locked}
\alias{assign_locked}
\title{Assign for locked environments}
\usage{
assign_locked(x, value, envir, namespace, keep_unlocked = FALSE)
}
\description{
\code{assign_locked()} assigns the supplied value in \code{value} to an object named in \code{x}
inside the namespace named in \code{namespace} (as character). If the object named in \code{x}
is locked, the binding is first unlocked and, depending on the value in \code{keep_unlocked}
is relocked afterwards. In locked environments only existing objects can be
modified by \code{assign_locked()}.
}
