% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{Bootstrap}
\alias{Bootstrap}
\title{Simple bootstrap}
\usage{
Bootstrap(x, statistic, B = 100, ..., portion = 100000000)
}
\description{
Returns a vector of B bootstrap values of real-valued statistic
\code{T}, where \code{T} should be an R-function which returns a scalar.
Arguments of \code{T} can be supplied via the ellipsis (\code{...}). If length
of \code{x} and \code{B} are small enough, do it all at once Otherwise, do it
for each iteration separately.
}
\examples{
sd(bootstrap(rnorm(100), mean, B=1000))    # compare with 1/sqrt(100)

}
