% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{cat0}
\alias{cat0}
\alias{catn}
\alias{catnn}
\alias{msgf}
\alias{msg}
\alias{warn}
\alias{note}
\alias{.fc}
\alias{flush_console}
\title{Print text}
\usage{
cat0(
  ...,
  file = "",
  sep = "",
  fill = FALSE,
  pad = FALSE,
  padding = " ",
  pad_side = "right",
  labels = NULL,
  append = FALSE,
  flush = TRUE,
  flush_cycle = 1,
  envir = utilbox_environment(),
  fill_using_options = FALSE
)

catn(
  ...,
  file = "",
  sep = "",
  fill = FALSE,
  labels = NULL,
  append = FALSE,
  eol = "\\n"
)

catnn(..., sep = "\\n", fill = TRUE)

msgf(..., flush = interactive())

msg(
  t,
  ...,
  lead = "",
  sep = "",
  quit = FALSE,
  wait = FALSE,
  skip1 = 0,
  skip2 = 0,
  flush = TRUE
)

warn(..., skip1 = 1, skip2 = 1)

note(...)

.fc()

flush_console(flush = TRUE, flush_cycle = 1, envir = utilbox_environment())
}
\description{
\code{cat0()} prints text to console or file via \code{\link[base:cat]{base::cat()}}. It is
basically an alias for \code{\link[base:cat]{base::cat()}} that automatically flushes the
console buffer (when \code{flush=TRUE}). This only affects buffered output
enabled R sessions (e.g. Rgui on Windows)

\code{catn()} is an alias for \code{cat0()} which in the default setting
adds a single end-of-line symbol (specified in \code{eol}, which \verb{\\\\n} by
default) at the end of the strings.

\code{catnn()} extends \code{catn} to also use \code{eol} as the default value
for \code{sep}.

All of these allow a smart way of flushing the console.

\code{msg()} prints a message with more control by user. Optionally, it can
stop the execution (when \code{quit=TRUE}) or wait for user interaction
(when \code{wait=TRUE}).

\code{warn()} is an alias for \code{msg()} which appends the string
\"WARNING\" to the printed message.

\code{note()} prints a note and a request to wait (optional).

\code{msgf()} is a wrapper around \code{base::message()} which flushes the output
by default (for interactive sessions).

\code{.fc()} is an alias for \code{\link[utils:flush.console]{utils::flush.console()}}.

\code{flush_console()} is a more adjustable way of flushing the console.
It allows to flush only on every n-th print, where n is set by the
value in \code{flush_cycle}.
}
\concept{printing function provided by utilbox}
