% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{collapse}
\alias{collapse}
\alias{collapse0}
\alias{str_collapse}
\alias{collapse0n}
\alias{collapsen}
\alias{collapse0nq}
\title{Collapse a string vector}
\usage{
collapse0(x, ..., sep = "")

str_collapse(x, ..., sep = "")

collapse0n(x, ..., sep = "\\n")

collapsen(x, ..., sep = "\\n")

collapse0nq(x, ..., sep = "'\\n'")
}
\description{
Takes a vector and collapses it into a single string. Equivalent to
[base::paste] with `collapse=''`.
}
\examples{
collapse0(c(0,1,1,1,0,1,0,0))

}
\seealso{
Other string-manipulation functions provided by utilbox: 
\code{\link{ascii}()},
\code{\link{cumpaste0}()},
\code{\link{padding}},
\code{\link{patternize}()},
\code{\link{str_add_punct}()},
\code{\link{str_diff}()},
\code{\link{str_insert}()},
\code{\link{str_is_empty}()},
\code{\link{str_last}()},
\code{\link{str_paste_grid}()},
\code{\link{str_pos}()},
\code{\link{str_replace}()},
\code{\link{str_scrub_space}()},
\code{\link{str_split}()},
\code{\link{str_trim_space}()},
\code{\link{strrev}()},
\code{\link{sub_at}()},
\code{\link{subm}()},
\code{\link{toupperfirst}()}
}
\concept{string-manipulation functions provided by utilbox}
