% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{concatenate}
\alias{concatenate}
\alias{\%_\%}
\alias{\%pastefront\%}
\alias{\%pasteend\%}
\title{Concatenation operators}
\usage{
... \%_\% sep

LHS \%pastefront\% RHS

LHS \%pasteend\% RHS
}
\description{
Concatenation operators, which are aliases for \code{base::paste}.

\code{\%_\%} is an operator versions of \code{base::paste} with
\code{sep='_'} and which takes symbols as arguments (see the examples
below).

\code{\%pastefront\%} is a conditional concatenatation operator,
which appends the string in \code{LHS} onto \code{RHS} from the left but
only if \code{RHS} does not already start with the string in \code{LHS}.

\code(\\%pasteend\\%} appends the string in \code{LHS} onto \code{RHS} from
the right (i.e., onto the end) but only if \code{RHS} does not already
end with the string inside the \code{LHS}.
}
\examples{
a = 'multi'; b = 'tasking'
a \%_\% b                    # BEWARE: yields 'a_b' (the inputs are treated as symbolic!)
'name:' \%pastefront\% 'james'        # yields 'name:james'
'name:' \%pastefront\% 'name:james'   # yields 'name:james'
'james' \%pasteend\% 'name:'          # yields 'name:james'
'james' \%pasteend\% 'name:james'     # yields 'name:james'

}
