% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{concatenate}
\alias{concatenate}
\alias{\%p\%}
\alias{\%..\%}
\alias{\%_\%}
\alias{\%__\%}
\alias{\%.^\%}
\alias{\%^.\%}
\title{Concatenation operators}
\usage{
x \%p\% y

x \%..\% y

x \%_\% y

... \%__\% sep

x \%.^\% y

x \%^.\% y
}
\description{

}
\examples{
a = 'multi'; b = 'tasking'
a \%p\% b                    # yields 'multitasking'
a \%..\% b                   # yields 'multi tasking'
a \%_\% b                    # yields 'multi_tasking'
a \%__\% b                   # BEWARE: yields 'a_b' (the inputs are treated as symbolic!)
'name:' \%.^\% 'james'        # yields 'name:james'
'name:' \%.^\% 'name:james'   # also yields 'name:james'
'name:' \%^.\% 'james'        # yields 'name:james'
'name:james' \%^.\% 'james'   # also yields 'name:james'

}
\seealso{
Other operators provided by utilbox: 
\code{\link{\%anyin\%}()},
\code{\link{\%modify\%}()},
\code{\link{not_in}},
\code{\link{operator_NULL}},
\code{\link{operator_append}},
\code{\link{operator_c}},
\code{\link{operator_match}},
\code{\link{operator_try}},
\code{\link{renaming_operators}}
}
\concept{operators provided by utilbox}
