% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{cumpaste0}
\alias{cumpaste0}
\title{Cumulative paste0}
\usage{
cumpaste0(x, .sep = "")
}
\description{
Takes a vector and cumulatively pastes it together.
}
\examples{
cumpaste0(c(0,1,1,1,0,1,0,0))

}
\seealso{
Other string-manipulation functions provided by utilbox: 
\code{\link{ascii}()},
\code{\link{collapse}},
\code{\link{multi_sub}()},
\code{\link{padding}},
\code{\link{patternize}()},
\code{\link{str_cut}()},
\code{\link{str_diff}()},
\code{\link{str_grab}()},
\code{\link{str_insert}()},
\code{\link{str_is_empty}()},
\code{\link{str_paste_grid}()},
\code{\link{str_pos}()},
\code{\link{str_rev}()},
\code{\link{str_trim_space}()},
\code{\link{sub_at}()},
\code{\link{toupperfirst}()}
}
\concept{string-manipulation functions provided by utilbox}
