% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighing.R
\name{downweight}
\alias{downweight}
\title{Down-weighing}
\usage{
downweight(
  x,
  pivot,
  speed = 0.15,
  side = c("left", "right", "both"),
  nonnegative = FALSE
)
}
\description{
Down-weighs the values in \code{x} if they are below the pivot (\code{from}) 
at the rate of determined by \code{speed}. More specifically, the value of 
\code{speed} is used as the standard deviation of the normal density which 
is used to bring the values in \code{x} to zero as they deviate from the pivot.
}
\details{
The argument \code{side} determines whether the down-weighing is applied only 
to the values below the pivot (\code{side='left'}), above the pivot 
(\code{side='right'}), or to both sides of the pivot (\code{side='both'}).

The argument \code{speed} expects values between '0' (no down-weighing) and
'1' (maximum down-weighing). Values outside this range are set to one of the
two boundaries.
}
\examples{
downweight(1:10, 5)

}
\seealso{
Other numeric functions provided by utilbox: 
\code{\link{Invert1}()},
\code{\link{Invert}()},
\code{\link{all_permutations}()},
\code{\link{as_numeric}()},
\code{\link{force_as_}},
\code{\link{frac_part}()},
\code{\link{frac}()},
\code{\link{int_part}()},
\code{\link{lower_tri}()},
\code{\link{make_numeric}()},
\code{\link{max0}()},
\code{\link{min_dif}()},
\code{\link{mintersect}()},
\code{\link{round2}()},
\code{\link{round_nearest_power}()},
\code{\link{round_nearest}()},
\code{\link{rsignif}()},
\code{\link{signif2}()},
\code{\link{to_base_k}()},
\code{\link{unspace}()},
\code{\link{unzero}()},
\code{\link{upper_tri}()},
\code{\link{value_locale}},
\code{\link{which_max}()}
}
\concept{numeric functions provided by utilbox}
