% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{environments}
\alias{environments}
\alias{orig_env}
\alias{all_parent_envs}
\title{Original and parent environments}
\usage{
orig_env(obj, na = "")

all_parent_envs(x = parent.frame())
}
\description{
\code{orig_env()} get the original environment of a function or another object.
When an object defined in a namespace is printed, its namespace is
shown as the last line. If a function in a package has been defined
as an alias of a function from another package, this original
information is still present when the function is printed For
instance, \code{utilbox::head} is just an alias for \code{utils::head}, as
shown when \code{print(utilbox::head)} is called. The function
\code{orig_env()} captures that information and returns it as a character
string. For objects with no such information printed it simply
returns the last element of a print call of that object.

\code{all_parent_envs()} returns the cascade of all parent environments
of \code{x}. If no input is specified during a call to \code{all_parent_envs()},
the default is to take the calling environment.
}
\examples{
orig_env(mean)

all_parent_envs()
all_parent_envs(mean)

}
