% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halting.R
\name{error}
\alias{error}
\title{Throw an error and stop}
\usage{
error(t, ..., sep = "", quit = TRUE, Q, browser = interactive(), nskip1 = 0)
}
\description{
A more elaborate version of stop(). It can do beeping and if inside an interactive
session it doesn't stop execution but enters a browsing mode
}
