% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{filter_by_position}
\alias{filter_by_position}
\alias{nth}
\alias{nthm}
\alias{nthr}
\alias{nthmr}
\alias{nth<-}
\alias{nthm<-}
\alias{nthr<-}
\alias{nthmr<-}
\title{Get every n-th element of an object}
\usage{
nth(x, n)

nthm(x, n)

nthr(x, n)

nthmr(x, n)

nth(x, n) <- value

nthm(x, n) <- value

nthr(x, n) <- value

nthmr(x, n) <- value
}
\description{
\code{nth} extracts the n-th element (single value) of an object \code{x}
(vector, list, etc.). It relies on \verb{\\}[[\`` for the extraction.

\code{nthm} allows for multivariate input in \code{n}. It relies on \verb{\\}[\``
for the extraction.

\code{nthr} and \verb{nthmr are the analogue of }nth\code{and}nthm` which
determine index the position in reversed order.
}
\examples{
# extaction of value
nth(1:10, 2)             # second
nthm(1:10, 3:4)          # third and fourth
nthr(1:10, 2)            # second from the end
nthmr(1:10, 3:4)         # third and fourth from the end

# extaction and assignment of value
y = as.list(1:10)
nth(y, 2) = Inf       # second
nthm(y, 3:5) = Inf    # third thru fifth
nthr(y, 2) = Inf      # second from the end
nthmr(y, 3:5) = Inf   # third thru fifth from the end

# extract via logicals
nthm(y, sapply(y, is.finite))

}
\seealso{
Other sequence-related functions provided by utilbox: 
\code{\link{every_nth}()},
\code{\link{filter_by}},
\code{\link{groups_of_unique}()},
\code{\link{h1}()},
\code{\link{insert}()},
\code{\link{is_between}()},
\code{\link{is_run}},
\code{\link{last_element}()},
\code{\link{rep2}()},
\code{\link{replace2}()},
\code{\link{replace3}()},
\code{\link{seq2}()},
\code{\link{seq_around}()},
\code{\link{sizechecks}},
\code{\link{split_into_groups}()},
\code{\link{to_factor}()},
\code{\link{what_bag}()}
}
\concept{sequence-related functions provided by utilbox}
