% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{function_separate_code}
\alias{function_separate_code}
\title{Separate lines of the code of a function}
\usage{
function_separate_code(fun)
}
\description{
\code{function_separate_code()} converts the body of a function (\code{fun}) to a list
where each element in the list corresponds to a block of code (often a
line). This is used when altering the code for instance via \code{append_body()}
below or via \code{base::trace()}.
}
