% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorchecks.R
\name{hide}
\alias{hide}
\title{Hide input}
\usage{
hide(x, ...)
}
\description{
\code{hide()} is a simple funciton which takes input and returns it back invisibly.
It is useful for instance for hiding errors and warnings within \code{base::tryCatch}
(see the examples).
}
\examples{
tryCatch(as.numeric('a'), warning=hide)
tryCatch(sin('a'), error=hide)

}
\seealso{
Other coding-related functions provided by utilbox: 
\code{\link{append_body}()},
\code{\link{call_to_list}},
\code{\link{compare_functions}()},
\code{\link{fun_code_to_text}()},
\code{\link{fun_dump_code}()},
\code{\link{hijack}()},
\code{\link{is_error}()},
\code{\link{namespaces}},
\code{\link{this_fun_name}()}
}
\concept{coding-related functions provided by utilbox}
