% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.R
\name{hijack}
\alias{hijack}
\title{Argument hijack function}
\usage{
hijack(FUN, ...)
}
\description{
A function that allows for changing of default values for arguments of other 
functions
}
\examples{
cat = hijack(cat, sep='')        # equivalent to cat0
mean = hijack(mean, na.rm=TRUE)  # equivalent to mean2

}
\seealso{
Other coding-related functions provided by utilbox: 
\code{\link{append_body}()},
\code{\link{call_to_list}},
\code{\link{compare_functions}()},
\code{\link{fun_code_to_text}()},
\code{\link{get_package_namespace}()},
\code{\link{this_fun_name}()}
}
\concept{coding-related functions provided by utilbox}
