% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqchecks.R
\name{is_between}
\alias{is_between}
\alias{or_between}
\title{Between check}
\usage{
is_between(x, a, b, sharp_left = FALSE, sharp_right = FALSE)

or_between(x, intervals, sharp_left = FALSE, sharp_right = FALSE)
}
\description{
\code{is_between()} is a shortcut for \code{x>=a & x<=b} (for
\code{sharp_left=TRUE} and \code{sharp_right=TRUE}). Other combinations of
\verb{TRUE/FALSE for }sharp_left,sharp_right` lead to non-sharp
comparisons.

\code{or_between()} performs is a between check with many intervals.
Checks whether \code{x} is in at least one of the supplied intervals.
}
\examples{
between(1:2, 1, 2))                # both
between(1:2, 1, 2, TRUE))          # second only
between(1:2, 1, 2, FALSE, TRUE)    # first only
between(1:2, 1, 2, TRUE, TRUE)     # none

or_between(runif(3), list(c(0,0.1), c(0.9,1)))

}
\seealso{
Other sequence-related functions provided by utilbox: 
\code{\link{every_kth}()},
\code{\link{filter_by_position}},
\code{\link{filter_by}},
\code{\link{groups_of_unique}()},
\code{\link{h1}()},
\code{\link{insert}()},
\code{\link{is_run}},
\code{\link{last_element}()},
\code{\link{rep2}()},
\code{\link{replace2}()},
\code{\link{replace3}()},
\code{\link{seq2}()},
\code{\link{seq_around}()},
\code{\link{sizechecks}},
\code{\link{split_into_groups}()},
\code{\link{to_factor}()},
\code{\link{what_bag}()}
}
\concept{sequence-related functions provided by utilbox}
