% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.R
\name{lang_sub}
\alias{lang_sub}
\alias{gsub_lang}
\alias{sub_lang}
\title{Replacement in expressions}
\usage{
lang_sub(x, ...)

gsub_lang(...)

sub_lang(...)
}
\description{
Similarly to \code{base::gsub()} and \code{base::sub()}, the functions \code{gsub_lang()}
and \code{sub_lang()} perform replacement in expressions. The expressions in
\code{code} are first converted to strings, the replacement is executed
(via \code{base::gsub()} or \code{base::sub()}) and the result is converted back
to language expressions. The input via \code{code} can be a list of
expressions or a function. \code{pattern} and \code{repl} are the analogues
\code{pattern} and \code{replacement} in \code{base::sub()} and \code{base::gsub()}.
}
\examples{
# Modify the `get0` function to return an NA when the requested
# object is not found, and to change the name of the argument
# 'ifnotfound' to 'default_value':
get3 = gsub_lang(base::get0, 'ifnotfound', 'default_value')
get3 = hijack(get3, default_value = NA)

# Modify the mean function to make a median
median2 = lang_sub(mean.default, 'trim = 0','trim = 0.5', fixed=TRUE)
x = rnorm(1000)
identical(median2(x), median(x))     # TRUE

}
