% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_clean}
\alias{list_clean}
\title{Removes zero-length elements from a list}
\usage{
list_clean(x, clean_by = length)
}
\description{
Takes a list and removes all of its zero-length elements.
Optionally, the filtering function can be changed via \code{clean_by}.
}
\examples{
list_clean(list(1:10, NULL, 'a'))

}
\seealso{
Other list utilities provided by utilbox: 
\code{\link{is_all_character}()},
\code{\link{list_empty}()},
\code{\link{list_flatten}()},
\code{\link{list_mirror}()},
\code{\link{list_set_attr}()},
\code{\link{list_update}()},
\code{\link{merge_recurse}()},
\code{\link{nlist}()}
}
\concept{list utilities provided by utilbox}
