% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripting.R
\name{list_created_by_sourcing}
\alias{list_created_by_sourcing}
\title{Get the list of objects created by sourcing a file}
\usage{
list_created_by_sourcing(file, ..., all.names = TRUE)
}
\description{
Find out what objects would be created if a file was sourced (via
\code{\link[base:sys.source]{base::sys.source}}) and return the list.
}
\examples{
tmp = '.~temp.R'
cat('test_function = function() NULL\n', file=tmp_file)
list_created_by_sourcing(tmp)
unlink(tmp)

}
