% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{list_files}
\alias{list_files}
\alias{list_files_by_date}
\alias{list_files_by_size}
\title{List files matching regular pattern}
\usage{
list_files(
  pattern = NULL,
  path = ".",
  full.names = FALSE,
  ...,
  sort = FALSE,
  attrib = c("mtime", "size", "isdir", "mode", "ctime", "atime", "exe"),
  decreasing = FALSE,
  append_path = FALSE
)

list_files_by_date(...)

list_files_by_size(...)
}
\description{
A wrapper for \code{\link[base:list.files]{base::list.files()}} that allows multiple patterns
at once and files that match at least on of the patterns are
returned. Optionally, files can be sorted according to attribute in
\code{attrib}.

\code{list_files_by_date()} and \code{list_files_by_size()} are shortcuts to
sorting according to attributes "modification time" and "size",
respectively.
}
\seealso{
Other file system functions provided by utilbox: 
\code{\link{dir_create}()},
\code{\link{dir_exist_check}()},
\code{\link{file_path}()},
\code{\link{file_rename}()}
}
\concept{file system functions provided by utilbox}
