% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{llibrary}
\alias{llibrary}
\alias{llib}
\alias{unload_library}
\alias{list_installed_packages}
\alias{list_loaded_packages}
\alias{package_is_installed}
\title{Library operations}
\usage{
llibrary(
  pckgs = NULL,
  quietly = TRUE,
  character.only = FALSE,
  fail = warn,
  detach_first = FALSE,
  default_src = "CRAN",
  url_CRAN = "https://cloud.r-project.org/"
)

llib(..., detach_first = FALSE)

unload_library(pckgs = NULL, character.only = FALSE)

list_installed_packages()

list_loaded_packages()

package_is_installed(pckgs, character.only = FALSE)
}
\description{
\code{llibrary()} loads specified libraries. On input, \code{pckgs} must
either be a character vector of package names or a list with each of
its elements having the structure of \code{list(name=..., src=...)}
where name is the package name and src is the package source
(repository or local) from which the package is to be loaded.

\code{llib()} is an alias for \code{llibrary()} which allows unquoted input.

\code{unload_library()} unloads/detaches a loaded library.

\code{list_installed_packages()} lists all installed packages.

\code{list_loaded_packages()}) list all loaded packages.

\code{package_is_installed()} checks if a package is installed.
}
\examples{
list_installed_packages()
list_loaded_packages()
package_is_installed('utilbox')

}
\seealso{
Other package-related functions provided by utilbox: 
\code{\link{apply_pckg}()},
\code{\link{list_package}},
\code{\link{set_pkglib}()}
}
\concept{package-related functions provided by utilbox}
