% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{lsos}
\alias{lsos}
\alias{object_size}
\alias{object_sizes}
\alias{list_object_sizes}
\title{List all objects and their sizes}
\usage{
lsos(..., envir = parent.frame(), n = 10, ndigits = 2, horse = NULL)

object_size(name, envir = parent.frame(), horse = object.size)

object_sizes(
  ...,
  list = character(),
  unit = "B",
  with_unit = TRUE,
  envir = parent.frame(),
  ndigits = 2,
  horse = NULL
)

list_object_sizes(
  pos = 1,
  pattern,
  envir = parent.frame(),
  order_by,
  decreasing = FALSE,
  head = FALSE,
  n = 5,
  all.names = TRUE,
  ndigits = 2,
  horse = NULL
)
}
\description{
\code{lsos()} lists objects in the given environment and returns their
names and sizes. It does so via \code{list_object_sizes()}, which has
a few more optional arguments and can also be called directly.

\code{object_sizes()} returns the sizes of supplied objects (in memory unit
given in 'unit').
}
