% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/means.R
\name{means}
\alias{means}
\alias{amean}
\alias{amean2}
\alias{hmean}
\alias{gmean}
\alias{gmeanw}
\alias{lmean}
\title{Various means}
\usage{
amean(x, na.rm = FALSE)

amean2(..., na.rm = TRUE)

hmean(...)

gmean(x, na.rm = FALSE)

gmeanw(x, w, ...)

lmean(x, w, ...)
}
\description{
`amean` gives the arithmetic mean of `x`. It has the same 
functionality as \code{base::mean} except with 
different output when input has zero length.
}
\details{
`amean2` is an alias for [base::mean] with a different
default for `na.rm` (unlike `base::mean`, it removes them by default).

`gmean` and `gmeanw` calculate the geometric mean of `x`, while the latter
permits weighing.

`hmean` calculates the harmonic mean (single value) of a single vector created as a
combination of all its input arguments.

`lmean` calculates the (weighted) logistic mean.
}
\examples{
x = c(1,NA); mean(x); mean2(x)

}
