% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{namespace_exists}
\alias{namespace_exists}
\title{Check for namespace}
\usage{
namespace_exists(ns)
}
\description{
Checks whether a namespace exists. For packages, this basically
checks whether a package is installed. The check is done by
calling \code{base::asNamespace(ns)} and returning a logical which
indicates whether an error occurred on that call.
}
