% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{namespaces}
\alias{namespaces}
\alias{get_package_namespace}
\alias{check_namespace}
\title{Get the namespace of a package}
\usage{
get_package_namespace(pckg, character.only = FALSE, stop_on_error = FALSE)

check_namespace(pckg, envir = parent.frame())
}
\description{
\code{get_package_namespace()} returns the namespace of the package
with name in \code{pckg}.

\code{check_namespace()} checks whether the package \code{pckg} is installed
and available and throws an error if not.
}
\examples{
get_package_namespace('base')

check_namespace('base')          # no error
check_namespace('XYZABC12323')   # error

}
\seealso{
Other coding-related functions provided by utilbox: 
\code{\link{append_body}()},
\code{\link{call_to_list}},
\code{\link{compare_functions}()},
\code{\link{fun_code_to_text}()},
\code{\link{fun_dump_code}()},
\code{\link{hide}()},
\code{\link{hijack}()},
\code{\link{is_error}()},
\code{\link{this_fun_name}()}
}
\concept{coding-related functions provided by utilbox}
