% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{operator_match}
\alias{operator_match}
\alias{\%match\%}
\alias{\%m\%}
\alias{\%notmatch\%}
\alias{\%nm\%}
\alias{\%matches\%}
\alias{\%notmatches\%}
\title{Regular expression match operator}
\usage{
pattern \%match\% x

pattern \%m\% x

... \%notmatch\% NA

... \%nm\% NA

x \%matches\% pattern

... \%notmatches\% NA
}
\description{
\code{\%match\%} is an operator version of \code{base::regexpr}. Uses 
the left-hand side argument as the \code{pattern} and the right-hand 
side argument as the \code{text} arguments of a call to \code{regexpr}. 
The call to \code{regexpr} is vectorized, which means that the operator 
also takes vector arguments (on either side).
}
\details{
\code{\%matches\%} does the same except with the roles of the two 
arguments reversed.

`\%notmatch%` and `\%notmatches%` are the negations of the two operators and
`\%m\%` and `\%nm%\` are aliases for `\%match\%` and `\%notmatch\%`.
}
\examples{
# in one direction:
"ay" \%match\% "daylight"
"ai" \%match\% "daylight"
# and in the opposite direction:
"daylight" \%matches\% "ay"
"daylight" \%matches\% "ai"

}
\seealso{
Other operators provided by utilbox: 
\code{\link{\%anyin\%}()},
\code{\link{\%append\%}()},
\code{\link{\%c\%}()},
\code{\link{\%modify\%}()},
\code{\link{concatenate}},
\code{\link{not_in}},
\code{\link{operator_NULL}},
\code{\link{renaming_operators}}
}
\concept{operators provided by utilbox}
