% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{operator_NULL}
\alias{operator_NULL}
\alias{\%|||\%}
\alias{\%||||\%}
\alias{\%|||||\%}
\alias{\%NA\%}
\alias{\%ERR\%}
\title{Default value for NULL and zero-length objects}
\usage{
x \%|||\% y

x \%||||\% y

x \%|||||\% y

x \%NA\% y

x \%ERR\% y
}
\description{
\verb{\\\%|||\\\%} is intented to be equivalent "\code{\%||\%}" operator in
the package \code{rlang}, but named differently as not to clash with it.

\verb{\\\%||||\\\%} extends this behavior to all zero-length objects.

\verb{\\\%|||||\\\%} extends this behavior further to include empty strings.

\verb{\\\%NA\\\%} is an operator version of `ifelse(is.na(x), y, x), which
allows input of various types (vectors, lists, closures, etc.).

\verb{\\\%ERR\\\%} works analogously except that it checks for class
\code{try-error}. If its first argument is of class \code{try-error} it returns
its second argument (\code{y}). Otherwise it returns the first (\code{x}).
}
\examples{
1 \%|||\% 2
NULL \%|||\% 2
NULL \%||||\% 2
numeric(0) \%||||\% 2

}
\seealso{
Other operators provided by utilbox: 
\code{\link{\%anyin\%}()},
\code{\link{\%modify\%}()},
\code{\link{concatenate}},
\code{\link{not_in}},
\code{\link{operator_append}},
\code{\link{operator_c}},
\code{\link{operator_match}},
\code{\link{renaming_operators}}
}
\concept{operators provided by utilbox}
