% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_gui.R
\name{R_gui_adjust}
\alias{R_gui_adjust}
\alias{R_gui_adjust2}
\alias{R_gui_reset}
\title{Adjust R GUI}
\usage{
R_gui_adjust(cmd = c("points=7", "rows=51", "columns=248"))

R_gui_adjust2(cmd = c("points=10", "rows=37", "columns=154"))

R_gui_reset()
}
\description{
\code{R_gui_adjust()} changes the way the GUI looks (e.g. font size,
font style). Without changing the default arguments, a call to
\code{Rgui_adjust} makes the font smaller so that more info fits the
screen. For changes in other ways set the value of \code{cmd}
appropriately.

\code{R_gui_adjust2()} is a wrapper for \code{R_gui_adjust()}, which by
default only resizes the active window.

\code{Rgui_reset()} resets the settings to their defaults.
}
\seealso{
Other system-related functions provided by utilbox: 
\code{\link{R_defaults}},
\code{\link{R_session_info}()},
\code{\link{environment_variables}},
\code{\link{is_win}()},
\code{\link{pause}},
\code{\link{stopping_control}},
\code{\link{system_path}},
\code{\link{timest}()}
}
\concept{system-related functions provided by utilbox}
