% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{R_session_info}
\alias{R_session_info}
\title{R session info}
\usage{
R_session_info()
}
\description{
Prints information about the current R session including the version of R
and the trailing arguments of the system call that started this session.
}
\seealso{
Other system-related functions provided by utilbox: 
\code{\link{.q}()},
\code{\link{R_gui_adjust}()},
\code{\link{R_system_call_args}()},
\code{\link{append_system_PATH}()},
\code{\link{code_startup}},
\code{\link{environment_variables}},
\code{\link{get_system_PATH}()},
\code{\link{is_win}()},
\code{\link{timest}()}
}
\concept{system-related functions provided by utilbox}
