% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{rep_list}
\alias{rep_list}
\alias{rep_list.list}
\alias{rep_list.default}
\alias{rep_list.matrix}
\title{Repeat elements in a list}
\usage{
rep_list(...)

\method{rep_list}{list}(
  x,
  n,
  each,
  length.out,
  bind = append_recurse,
  flatten = FALSE,
  rep_as_list = FALSE
)

\method{rep_list}{default}(
  x,
  n,
  each,
  length.out,
  bind = append_recurse,
  flatten = FALSE,
  rep_as_list = FALSE
)

\method{rep_list}{matrix}(
  x,
  n,
  each,
  length.out,
  bind = append_recurse,
  flatten = FALSE,
  rep_as_list = FALSE
)
}
\description{
An extension of \code{\link[base:rep]{base::rep()}} which works with lists which are
not unlisted. For non-list types it works the same as \code{\link[base:rep]{base::rep()}}.
}
\examples{
rep_list(list(a=1,b=2), 3)
rep_list(list(a=1, mean), 2)

}
