% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\name{rev_rows}
\alias{rev_rows}
\alias{rev_cols}
\title{Reverse the order of rows/colums}
\usage{
rev_rows(x)

rev_cols(x)
}
\value{
Object of the same class as the input.
}
\description{
Reverses the order of rows (\code{rev_rows}) or columns (\code{rev_cols})
of a matrix, a data frame or a 2-dimensional array.
}
\seealso{
Other matrix/data-frame functions provided by utilbox: 
\code{\link{add_col}()},
\code{\link{identical_cols}()},
\code{\link{join_recurse}()},
\code{\link{rep_df}()},
\code{\link{split_rows}()}

Other matrix/data-frame functions provided by utilbox: 
\code{\link{add_col}()},
\code{\link{identical_cols}()},
\code{\link{join_recurse}()},
\code{\link{rep_df}()},
\code{\link{split_rows}()}
}
\concept{matrix/data-frame functions provided by utilbox}
