% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{sequences}
\alias{sequences}
\alias{shift}
\alias{nunique}
\alias{frac}
\alias{midpoints}
\title{Points in a sequence}
\usage{
shift(x, lag = 1, rotate = TRUE, value_rotated = NA)

nunique(x)

frac(x)

midpoints(x)
}
\description{
\code{shift()} shifts (rotate) the elements in a vector in \code{x} by \code{lag}
spaces. With \code{rotate=FALSE} the rotated elements are replaced with
\code{value_rotated} (\code{NA} by default).

\code{frac} gives sequential fractions. It is similar to \code{base::diff}
except that it returns lagged ratios instead of lagged differences.

\code{midpoints} find the midpoints between individual elements of a
sequence given in \code{x}.

\code{nunique()} counts the number of unique values in the supplied
argument. Simply an alias for \code{length(unique(x))}.
}
\examples{
frac(1:10)

midpoints(1:10)
midpoints(c(1:5,3*1:5))

}
\seealso{
Other numeric functions provided by utilbox: 
\code{\link{Invert}()},
\code{\link{downweight}()},
\code{\link{extract_by_value}},
\code{\link{locate_position_of_value}},
\code{\link{lower_tri}()},
\code{\link{min_dif}()},
\code{\link{permutations}},
\code{\link{replace_value}},
\code{\link{sets}},
\code{\link{to_base_k}()},
\code{\link{which_max}()}
}
\concept{numeric functions provided by utilbox}
