% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notation.R
\name{short_notation}
\alias{short_notation}
\title{Convert a number to an abbreviated notation}
\usage{
short_notation(x, exact_cutoff = 999, ndig = 1, word_units = FALSE)
}
\description{
Convert a number to an abbreviated notation
}
