% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_break}
\alias{str_break}
\title{Wrap text}
\usage{
str_break(x, max_width = Inf, eol = "\\n", break_only_at_space = FALSE)
}
\description{
\code{str_break()} wraps (or more generally breaks) a string text. In other words,
it places the value in \code{eol} (the new line symbol \verb{\\\\n} by default) along a
string to make its printout (via e.g. \code{base::cat}) \"wrap nicely\",
that is not exceed a given width (set by \code{max_width}). By default,
it places the breaks (\code{eol}) anywhere, while with \code{break_only_at_space=TRUE}
it only places the breaks between words.
}
\examples{
string = collapse0(rep(collapse0(letters[1:8]),10), sep=" ")

# insert line breaks at given width
message(str_break(string, max_width=20))

# insert only at spaces
message(str_break(string, max_width=20, break_only_at_space=TRUE))

}
