% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_diff}
\alias{str_diff}
\title{Lagged differences for strings}
\usage{
str_diff(x)
}
\description{
Compares neighbouring elements in a string vector and indicates as
\code{TRUE/FALSE} the differences. Returns \code{TRUE} if the corresponding
elements are different, and \code{FALSE} whet they are the same. Similar
to \code{\link[base:diff]{base::diff}} but for strings.
}
\examples{
str_diff(c('a','a','b','b','b','c'))

}
\seealso{
Other string-manipulation functions provided by utilbox: 
\code{\link{ascii}()},
\code{\link{collapse}()},
\code{\link{cumpaste0}()},
\code{\link{multi_sub}()},
\code{\link{padding}},
\code{\link{patternize}()},
\code{\link{str_cut}()},
\code{\link{str_grab}()},
\code{\link{str_insert}()},
\code{\link{str_is_empty}()},
\code{\link{str_paste_grid}()},
\code{\link{str_pos}()},
\code{\link{str_rev}()},
\code{\link{str_trim_space}()},
\code{\link{sub_at}()},
\code{\link{toupperfirst}()}
}
\concept{string-manipulation functions provided by utilbox}
