% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_split}
\alias{str_split}
\alias{str2vector}
\title{String splitting}
\usage{
str_split(x, split, ..., unlist_for_scalar = FALSE)

str2vector(x, split = "", ..., unlist_for_scalar = TRUE)
}
\description{
\code{str_split()} splits a string vector \code{x} by character string \code{split}.
Basically, it is the same as \code{\link[base:strsplit]{base::strsplit}} except that
for zero-length strings it does not return a zero-length element in
the returned list.

\code{str2vector()} is an alias for \code{str_split()} with the default value
for split set to '' (an empty string) so that it ends up producing a
vector of individual characters. It also unlists the returned value
for scalar \code{x}.
}
\examples{
str_split(c('a-b','a-c','a-d','','c-d','c-e'),'-')

}
\seealso{
Other string-manipulation functions provided by utilbox: 
\code{\link{ascii}()},
\code{\link{collapse}()},
\code{\link{cumpaste0}()},
\code{\link{multi_sub}()},
\code{\link{padding}},
\code{\link{patternize}()},
\code{\link{str_diff}()},
\code{\link{str_grab}()},
\code{\link{str_insert}()},
\code{\link{str_is_empty}()},
\code{\link{str_paste_grid}()},
\code{\link{str_pos}()},
\code{\link{str_rev}()},
\code{\link{str_trim_space}()},
\code{\link{sub_at}()},
\code{\link{toupperfirst}()}
}
\concept{string-manipulation functions provided by utilbox}
