% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_wrap}
\alias{str_wrap}
\alias{str_wrap1}
\title{Wrap text}
\usage{
str_wrap(x, max_width = Inf, eol = "\\n", break_only_at_space = FALSE)

str_wrap1(
  x,
  max_width = Inf,
  eol = "\\n",
  break_only_at_space = FALSE,
  max_nlines = Inf
)
}
\description{
Places newline symbols (`\\n`) along a string to make its
print via `cat` not exceed a given width (`max_width`).
}
\examples{
catn(str_wrap(letters, max_width=10)

}
