% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic.R
\name{symbolic_call_names}
\alias{symbolic_call_names}
\title{\code{symbolic_call_names()} returns the names of arguments that are
needed by a symbolic call.}
\usage{
symbolic_call_names(n)
}
\description{
\code{symbolic_call_names()} returns the names of arguments that are
needed by a symbolic call.
}
