% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{system_path}
\alias{system_path}
\alias{get_system_path}
\alias{append_system_path}
\title{Modify system search path}
\usage{
get_system_path(path)

append_system_path(path)
}
\description{
\code{get_system_path()} reads and returns the system search path variables.

\code{append_system_path()} adds the supplied path into the system
search path variables.
}
\examples{
get_system_path()

}
\seealso{
Other system-related functions provided by utilbox: 
\code{\link{R_defaults}},
\code{\link{R_gui_adjust}()},
\code{\link{R_session_info}()},
\code{\link{environment_variables}},
\code{\link{is_win}()},
\code{\link{pause}},
\code{\link{stopping_control}},
\code{\link{timest}()}
}
\concept{system-related functions provided by utilbox}
