% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.R
\name{this_fun_name}
\alias{this_fun_name}
\title{Get function name}
\usage{
this_fun_name()
}
\description{
Returns the name of the parent function, i.e. that from whose body it is invoked.
}
\examples{
lll = function() print(this_fun_name()); lll() # should return 'lll'

}
\seealso{
Other coding-related functions provided by utilbox: 
\code{\link{append_body}()},
\code{\link{call_to_list}},
\code{\link{compare_functions}()},
\code{\link{fun_code_to_text}()},
\code{\link{hijack}()}
}
\concept{coding-related functions provided by utilbox}
