% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{toupperfirst}
\alias{toupperfirst}
\title{Capitalization}
\usage{
toupperfirst(x)
}
\description{
Capitalize first letters of each element in the vector \code{string}.
}
\examples{
toupperfirst('hello')

}
\seealso{
Other string-manipulation functions provided by utilbox: 
\code{\link{ascii}()},
\code{\link{collapse}},
\code{\link{cumpaste0}()},
\code{\link{multi_sub}()},
\code{\link{padding}},
\code{\link{patternize}()},
\code{\link{str_cut}()},
\code{\link{str_diff}()},
\code{\link{str_grab}()},
\code{\link{str_insert}()},
\code{\link{str_is_empty}()},
\code{\link{str_paste_grid}()},
\code{\link{str_pos}()},
\code{\link{str_rev.character}()},
\code{\link{str_trim_space}()},
\code{\link{sub_at}()}
}
\concept{string-manipulation functions provided by utilbox}
