% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilbox.R
\name{utilbox_internals}
\alias{utilbox_internals}
\alias{utilbox_namespace}
\alias{utilbox_environment}
\alias{get_in_utilbox_env}
\alias{assign_in_utilbox_env}
\alias{update_in_utilbox_env}
\alias{init_in_utilbox_env}
\alias{get_utilbox_var_inits}
\title{Utilbox internals}
\usage{
utilbox_namespace()

utilbox_environment(envir = .GlobalEnv)

get_in_utilbox_env(nam, envir = utilbox_namespace())

assign_in_utilbox_env(nam, val, envir = utilbox_namespace())

update_in_utilbox_env(nam, val, operation = `+`, envir = utilbox_namespace())

init_in_utilbox_env(nams, envir = utilbox_namespace())

get_utilbox_var_inits(nam)
}
\description{
\code{get_in_utilbox_env} retrieves a value of a variabled named in \code{x} from
the utilbox's package namespace or the hidden environment created in
\code{.GlobalEnv} by a call to \code{utilbox_environment()}.

\code{assign_in_utilbox_env} is an analogue which assigns a supplied value
(\code{val}) to a variable (\code{x}).

\code{update_in_utilbox_env} updates the existing variable by applying a
function (\code{operation}) with the variable \code{x} and \code{val} supplied to it
as the first two arguments.

\code{init_in_utilbox_env} initializes a given variable to its default value
(defined in and obtained via \code{get_utilbox_var_inits}). If not variable name
is supplied, all variables defined in \code{get_utilbox_var_inits} are initialized.

\code{utilbox_namespace()} returns the environment which is the
namespace of the utilbox package. If it failed for some reason, it
returns creates and/or returns the hidden environment located in
\code{.GlobalEnv}.

\code{utilbox_environment()} creates the working environment called
\code{.utilbox}, which is used by some of the functions of the \code{utilbox}
package (e.g. \code{cat0}) to store private variables.
}
\seealso{
Other utilbox-internal functions provided by utilbox: 
\code{\link{source_utilbox}()}
}
\concept{utilbox-internal functions provided by utilbox}
