% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{value_locale}
\alias{value_locale}
\alias{first_nonzero}
\alias{nth_nonzero}
\alias{last_nonzero}
\alias{first_positive}
\alias{last_positive}
\alias{last_above}
\alias{first_above}
\alias{last_below}
\alias{first_below}
\title{Locate value}
\usage{
first_nonzero(x)

nth_nonzero(x, n)

last_nonzero(x)

first_positive(x)

last_positive(x)

last_above(x, y, ...)

first_above(x, y, ...)

last_below(x, y, ...)

first_below(x, y, ...)
}
\description{
`first_nonzero` finds the position of the last non-zero element in `x`.
}
\details{
`nth_nonzero` finds the position of the n-th non-zero element in `x`.

`last_nonzero` finds the position of the last non-zero element in `x`.

`first_positive` finds the position of the first positive element in `x`.

`last_positive` finds the position of the last positive element in `x`.

`last_above` finds the position of the last element in `x` with value larger than `y`.

`first_above` finds the position of the first element in `x` with value larger than `y`.

`last_below` finds the position of the last element in `x` with value below `y`.

`first_below` finds the position of the first element in `x` with value below `y`.
}
\seealso{
Other numeric functions provided by utilbox: 
\code{\link{Invert1}()},
\code{\link{Invert}()},
\code{\link{all_permutations}()},
\code{\link{as_numeric}()},
\code{\link{downweight}()},
\code{\link{force_as_}},
\code{\link{frac_part}()},
\code{\link{frac}()},
\code{\link{int_part}()},
\code{\link{lower_tri}()},
\code{\link{make_numeric}()},
\code{\link{max0}()},
\code{\link{min_dif}()},
\code{\link{mintersect}()},
\code{\link{round2}()},
\code{\link{round_nearest_power}()},
\code{\link{round_nearest}()},
\code{\link{rsignif}()},
\code{\link{signif2}()},
\code{\link{to_base_k}()},
\code{\link{unspace}()},
\code{\link{unzero}()},
\code{\link{upper_tri}()},
\code{\link{which_max}()}
}
\concept{numeric functions provided by utilbox}
