% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmax.R
\name{which_max}
\alias{which_max}
\alias{which_min}
\title{Maximum/minimum}
\usage{
which_max(
  x,
  last = TRUE,
  first = FALSE,
  all = FALSE,
  na.rm = FALSE,
  na.ignore = TRUE,
  arr.ind = FALSE
)

which_min(
  x,
  last = TRUE,
  first = FALSE,
  all = FALSE,
  na.rm = FALSE,
  na.ignore = TRUE,
  arr.ind = FALSE,
  eps = 0
)
}
\description{
A more versatile version of \code{base:which.max} and \code{which.min}.
}
\details{
Essentially, the functions \code{which_max} and \code{which_min} 
behave just like their base equivalents except they can return 
the position of both first or last extreme in a vector (the last 
by default). They also differ in the treatment of \code{NA}s. The
removal of NA values can be done in two ways, they are either removed
and then the position of the extreme in a vector stripped of \code{NA}s 
is returned, or they are ignored (default) and the position of a extreme
within the entire original vector is returned.
}
\examples{
which_max(1:5)
which_max(1:5, last=FALSE)
which_max(c(NA,1:5))
which_max(c(NA,1:5), na.ignore=FALSE, na.rm=TRUE)

}
\seealso{
Other numeric functions provided by utilbox: 
\code{\link{Invert1}()},
\code{\link{Invert}()},
\code{\link{all_permutations}()},
\code{\link{as_numeric}()},
\code{\link{downweight}()},
\code{\link{force_as_}},
\code{\link{frac_part}()},
\code{\link{frac}()},
\code{\link{int_part}()},
\code{\link{lower_tri}()},
\code{\link{make_numeric}()},
\code{\link{max0}()},
\code{\link{min_dif}()},
\code{\link{mintersect}()},
\code{\link{round2}()},
\code{\link{round_nearest_power}()},
\code{\link{round_nearest}()},
\code{\link{rsignif}()},
\code{\link{signif2}()},
\code{\link{to_base_k}()},
\code{\link{unspace}()},
\code{\link{unzero}()},
\code{\link{upper_tri}()},
\code{\link{value_locale}}

Other numeric functions provided by utilbox: 
\code{\link{Invert1}()},
\code{\link{Invert}()},
\code{\link{all_permutations}()},
\code{\link{as_numeric}()},
\code{\link{downweight}()},
\code{\link{force_as_}},
\code{\link{frac_part}()},
\code{\link{frac}()},
\code{\link{int_part}()},
\code{\link{lower_tri}()},
\code{\link{make_numeric}()},
\code{\link{max0}()},
\code{\link{min_dif}()},
\code{\link{mintersect}()},
\code{\link{round2}()},
\code{\link{round_nearest_power}()},
\code{\link{round_nearest}()},
\code{\link{rsignif}()},
\code{\link{signif2}()},
\code{\link{to_base_k}()},
\code{\link{unspace}()},
\code{\link{unzero}()},
\code{\link{upper_tri}()},
\code{\link{value_locale}}
}
\concept{numeric functions provided by utilbox}
